% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action_class.R
\docType{class}
\name{Action}
\alias{Action}
\title{La classe Action}
\description{
Classe pour les actifs de type Action
}
\section{Slots}{

\describe{
\item{\code{ptf_action}}{est un dataframe, chaque ligne represente un actif action du portefeuille d'action.}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

