% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlmEngine_class.R
\docType{class}
\name{AlmEngine}
\alias{AlmEngine}
\title{La classe ALMEngine}
\description{
Classe ayant pour principal vocation de contenir des methodes de reallocation.
}
\section{Slots}{

\describe{
\item{\code{journal_achat_vente}}{outil permettant de memoriser l'ensemble des operations d'achat-vente.}
}}
\author{
Prim'Act
}
\seealso{
La fonction de reallocation du Portefeuille \code{\link{reallocate}}
}

