% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutresPassifs-load.R
\docType{methods}
\name{autres_passif_load}
\alias{autres_passif_load}
\title{Methode permettant de charger la valeur initiale des autres passifs.}
\usage{
autres_passif_load(file_autres_passif_address)
}
\arguments{
\item{file_autres_passif_address}{est un \code{character} contenant l'adresse exacte
du fichier d'input utilisateur permettant de renseigner un objet \code{\link{AutresPassifs}}.}
}
\value{
L'objet de la classe \code{\link{AutresPassifs}} construit a partir des inputs renseignes par l'utilisateur.
}
\description{
\code{autres_passif_load} est une methode permettant de charger les donnees associees a un
objet de classe \code{\link{AutresPassifs}}.
}
\seealso{
La classe \code{\link{Initialisation}} et sa methode \code{\link{set_architecture}}
 pour renseigner l’input.
}
\author{
Prim'Act
}
