% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutresReserves-load.R
\docType{methods}
\name{autres_reserves_load}
\alias{autres_reserves_load}
\title{Methode permettant de charger la valeur initiale de la PSAP et de la PGG.}
\usage{
autres_reserves_load(file_autres_reserves_address)
}
\arguments{
\item{file_autres_reserves_address}{est un \code{character} contenant l'adresse exacte
du fichier d'input utilisateur
permettant de renseigner un objet \code{\link{AutresReserves}}.}
}
\value{
L'objet de la classe \code{\link{AutresReserves}} construit a partir des inputs renseignes par l'utilisateur.
}
\description{
\code{autres_reserves_load} est une methode permettant de charger les donnees associees a un
objet de classe \code{\link{AutresReserves}}.
}
\author{
Prim'Act
}
\seealso{
La classe \code{\link{Initialisation}} et sa methode \code{\link{set_architecture}}
 pour renseigner l’input.
}

