% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutresPassifs-class.R
\docType{class}
\name{AutresPassifs}
\alias{AutresPassifs}
\title{La classe AutresPassifs}
\description{
Classe pour les Autres Passifs
}
\section{Slots}{

\describe{
\item{\code{mp}}{un objet data.frame au format fige contenant l'ensemble de model points Autres Passifs}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

