% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutresReserves-class.R
\docType{class}
\name{AutresReserves}
\alias{AutresReserves}
\title{La classe AutreReserves}
\description{
Classe permettant de gerer le stock de provision globale de gestion (PGG) et de
provision pour sinistres a payer (PSAP)
}
\section{Slots}{

\describe{
\item{\code{pgg_debut}}{la valeur de la PGG en debut de periode}

\item{\code{psap_debut}}{la valeur de la PSAP en debut de periode}

\item{\code{pgg_valeur}}{la valeur courant de la PGG.}

\item{\code{psap_valeur}}{la valeur courant de la PSAP.}

\item{\code{tx_pgg_ep}}{le taux de PGG applique sur l'epargne.}

\item{\code{tx_pgg_autres}}{le taux de PGG applique sur les autres passifs.}

\item{\code{tx_psap_ep}}{le taux de PGG applique sur l'epargne.}

\item{\code{tx_psap_autres}}{le taux de PGG applique sur les autres passifs.}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

