% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevaloEngine_base_prod_fin.R, R/RevaloEngine_calc_revalo.R, R/RevaloEngine_finance_cible_marge.R, R/RevaloEngine_finance_cible_pmvl.R, R/RevaloEngine_finance_cible_ppb.R, R/RevaloEngine_finance_tmg.R, R/RevaloEngine_pb_contr.R
\docType{methods}
\name{base_prod_fin}
\alias{RevaloEngine}
\alias{base_prod_fin}
\title{Calcule la base de produits financiers attribuables.}
\usage{
base_prod_fin(tra, pm_moy, ppb)

calc_revalo(x, passif_av_pb, tra, plac_moy_vnc, result_tech)

finance_cible_marge(marge_fin, bes_cible, rev_stock_nette, marge_min)

finance_cible_pmvl(bes_cible, rev_stock_nette, base_fin, seuil_pmvl, tx_pb)

finance_cible_ppb(bes_cible, rev_stock_nette, ppb)

finance_tmg(bes_tmg_prest, bes_tmg_stock, ppb)

pb_contr(base_fin, tx_pb, rev_stock_brut, ch_enc_th, tx_enc_moy)
}
\arguments{
\item{tra}{est un objet \code{numeric} donnant le taux de rendement de l'actif.}

\item{pm_moy}{est un vecteur \code{numeric} comprenant le montant de PM moyenne par produit.}

\item{ppb}{est un objet de la classe \code{Ppb} qui renvoie l'etat courant de la PPB.}

\item{x}{un objet de la classe \code{Canton}.}

\item{passif_av_pb}{est une liste produit par la methode \code{viellissement_av_pb}
appliquee a un portefeuille de passif.}

\item{plac_moy_vnc}{est la valeur \code{numeric} moyenne des actifs.}

\item{result_tech}{est la valeur \code{numeric} du resultats techniques prise en compte de la PB.}

\item{marge_fin}{est valeur \code{numeric} donnant le montant courant de la marge financiere de l'assureur.}

\item{bes_cible}{est un vecteur de type \code{numeric} comprenant par produit
le besoin de financement aux taux cible.}

\item{rev_stock_nette}{est un vecteur de type \code{numeric} comprenant par produit
le montant de revalorisation nette atteint.}

\item{marge_min}{est une valeur de type \code{numeric} correspondant
au montant minimum de marge financiere souhaite par l'assureur.}

\item{base_fin}{est un vecteur de type \code{numeric} comprenant par produit la base de produits financiers.}

\item{seuil_pmvl}{est une valeur \code{numeric} comprenant au montant de plus-values latentes qui peut etre liquidee. Ce montant
doit etre exprime en tenant compte de l'abattement realise pour rapport les plus-values latentes de l'actiff general au passif.}

\item{tx_pb}{est un vecteur de type \code{numeric} comprenant par produit les taux de particippation aux benefices contractuels.}

\item{bes_tmg_prest}{est un vecteur de type \code{numeric} comprenant
par produit le besoin de finance des TMG sur prestations.
@param bes_tmg_stock est un vecteur de type \code{numeric} comprenant
par produit le besoin de finance des TMG sur le stock.}

\item{rev_stock_brut}{est un vecteur de type \code{numeric} comprenant par produit la revalorisation
appliquee sur le stock au taux minimum.}

\item{ch_enc_th}{est un vecteur de type \code{numeric} comprenant par produit le montant total
des chargements sur encours appliques stock, revalorise au taux minimum.}

\item{tx_enc_moy}{est un vecteur de type \code{numeric} comprenant par produit
les taux de chargements sur encours moyens.}

\item{tra}{est la valeur \code{numeric} du taux de rendement de l'actif.}

\item{bes_cible}{est un vecteur de type \code{numeric} comprenant par produit le besoin de financement aux taux cible.}

\item{rev_stock_nette}{est un vecteur de type \code{numeric} comprenant par produit le montant de revalorisation nette atteint.}

\item{bes_cible}{est un vecteur de type \code{numeric} comprenant par produit le besoin de financement aux taux cible.}

\item{rev_stock_nette}{est un vecteur de type \code{numeric} comprenant par produit le montant de revalorisation nette atteint.}

\item{ppb}{est un objet de la classe \code{Ppb} qui renvoie l'etat courant de la PPB.}

\item{ppb}{est un objet de la classe \code{Ppb} qui renvoie l'etat courant de la PPB.}

\item{base_fin}{est un vecteur de type \code{numeric} comprenant par produit la base de produits financiers.}

\item{tx_pb}{est un vecteur de type \code{numeric} comprenant par produit les taux de participation aux benefices contractuels.}
}
\value{
La valeur de la base de produit financier par produit et au total pour le portefeuille.

Une liste avec la valeur de la revalorisation additionelle nette servie par
 produit, la ppb mise a jour, la PPB mise a jour et le resultat technique avant PB.

Une liste avec la valeur de la revalorisation nette servie par produit
 et le montant de marge de l'assureur.

Une liste avec la valeur de la revalorisation nette servie par produit et le montant de plus-values a liquider, ramene a
la valeur du passif pour financer la revalorisation.

Une liste avec la valeur de la revalorisation nette servie par produit, les montants de dotation
et reprise sur PPB et la PPB mise à jour.

Une liste comprenant la valeur de la contribution au financement des TMG sur prestations et sur stock.
Cette liste comprend egalement la PPB mise a jour.

Une liste composee de deux vecteurs comprenant par produit les chargements sur encours appliques et la
revalorisation contractuelle nette.
}
\description{
\code{base_prod_fin} est une methode permettant de calculer la base de produits financiers attribuables
 pour la revalorisation des contrats.

\code{calc_revalo} est une methode permettant de
 d'appliquer l'ensemble de la politique de revalorisation d'un assureur.

\code{finance_cible_ppb} est une methode permettant de
 determiner le financement de taux cibles par la provision pour participation aux benefices.

\code{finance_cible_pmvl} est une methode permettant de
 determiner le financement de taux cibles par des cessions de plus-values latentes.

\code{finance_cible_ppb} est une methode permettant de
 determiner le financement de taux cibles par la provision pour participation aux benefices.

\code{finance_tmg} est une methode permettant d'evaluer le contribution de la PPB
au financement des taux minimum garanti sur prestations et sur stock.

\code{pb_contr} est une methode permettant de calculer la participation aux benefices contractuelle par produit.
}
\details{
Lorsque la revalorisation nette est superieure au besoin de financement des taux cibles, on sert le taux cible
et on partage le surplus. A l'inverse, les taux cible sont finances par
les compensations entre produits lorsque certains prevoient une revalorisation superieure au taux cible, et
par une liquidation de plus-values latentes.

Lorsque la revalorisation nette est superieure au besoin de financement des taux cibles, on sert le taux cible
et on dote le reste a la PPB dans la limite du plafond de dotation annuel. A l'inverse, les taux cible sont finances par
les compensations entre produits lorsque certains prevoient une revalorisation superieure au taux cible, et
par une reprise sur PPB.

Cette fonction priorise le financement des TMG prestations. Elle met a jour la PPB des reprises effectuees.

Le montant des chargements \code{ch_enc_th} est theorique et peut
conduire a l'application d'une revalorisation nette negative.
}
\author{
Prim'Act

Prim'Act

Prim'Act

Prim'Act

Prim'Act

Prim'Act

Prim'Act
}

