% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Be_class.R
\docType{class}
\name{Be}
\alias{Be}
\title{La classe Be}
\description{
Classe pour le BE d'un assureur
}
\section{Slots}{

\describe{
\item{\code{param_be}}{est un objet de type \code{ParamBe} qui regroupe les parametres de base du best estimate.}

\item{\code{canton}}{est un objet de type \code{Canton} correspond au canton en date initiale.}

\item{\code{esg}}{est un objet de type \code{ESG}.}

\item{\code{tab_flux}}{est une liste qui contient les flux moyens de best estimate.}

\item{\code{tab_be}}{est une liste qui contient la valeur du BE et de ses composantes.}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

