% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Be_data.R
\docType{data}
\name{Be_DataSet}
\alias{Be_DataSet}
\alias{be}
\title{Dataset Be.}
\format{
Objet de la classe \code{\link{Be}}.
}
\usage{
data("be")
}
\description{
Dataset de la classe \code{\link{Be}} contenant des donnees correspondant aux classes : \code{\link{Canton}}, \code{\link{ParamBe}}, \code{\link{ESG}}.
}
\note{
Il s'agit de donnees simulees, elles ont ete creees afin de tester les differents fonctions du package.
}
\examples{
data(be)
param_be <- be@param_be
portefeuille_passif <- be@canton@ptf_passif
}
\author{
Prim'Act
}
\keyword{datasets}
