% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_buy.R
\docType{methods}
\name{buy_oblig}
\alias{buy_oblig}
\title{Mise a jour de chaque composante d'un portefeuille obligataire suite a un achat d'un autre portefeuille obligataire.}
\usage{
buy_oblig(x, ptf_bought)
}
\arguments{
\item{x}{objet de la classe \code{\link{Oblig}} (decrivant le portefeuille obligataire en detention).}

\item{ptf_bought}{objet de la classe \code{\link{Oblig}} (decrivant le portefeuille obligataire achete).}
}
\value{
L'objet \code{x} complete des elements de \code{ptf_bought}.
}
\description{
\code{buy_oblig} est une methode permettant de mettre a jour le portefeuille obligataire suite a l'achat d'un autre portefeuille obligataire.
de chaque composante d'un portefeuille obligataire.
}
\author{
Prim'Act
}

