% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamTableMort-calc_coef_rente.R
\docType{methods}
\name{calc_ax}
\alias{calc_ax}
\title{Calcule la valeur d'un coefficient actuariel d'une rente sur une tete.}
\usage{
calc_ax(tx_tech, freq_rente, echu, table_mort, age, gen)
}
\arguments{
\item{tx_tech}{un \code{numeric} correspondant au taux technique utilise pour l'actualisation.}

\item{freq_rente}{un \code{integer} correspondant au nombre de versements dans l'annee.}

\item{echu}{un \code{logical} valant \code{TRUE} si la rente est a terme echue et \code{FALSE}
si la rente est a terme echoir.}

\item{table_mort}{un objet de la classe \code{ParamTableMort} designant la table de mortalite de la tete principale.}

\item{age}{la valeur \code{integer} de l'age de la tete principale.}

\item{gen}{la valeur \code{integer} de la generation de la tete principale.}
}
\value{
Le coefficient actuariel calcule.
}
\description{
\code{calc_ax} est une methode permettant de calculer les coefficients actuariels
pour des rentes versant potentiellement plusieurs flux dans l'annee.
}
\seealso{
Le calcul des probabilites de survies \code{\link{calc_proba_survie}}.
}
\author{
Prim'Act
}
