% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamTableMort-calc_coef_rente.r
\docType{methods}
\name{calc_axy}
\alias{calc_axy}
\title{Calcule la valeur d'un coefficient actuariel d'une rente avec reversion.}
\usage{
calc_axy(tx_tech, tx_rvs, freq_rente, echu, table_mort_1, age_1, gen_1,
  table_mort_2, age_2, gen_2)
}
\arguments{
\item{tx_tech}{un \code{numeric} correspondant au taux technique utilise pour l'actualisation.}

\item{tx_rvs}{un \code{numeric} correspondant au taux de reversion de la rente.}

\item{freq_rente}{un \code{integer} correspondant au nombre de versements dans l'annee.}

\item{echu}{un \code{logical} valant \code{TRUE} si la rente est a terme echue et \code{FALSE}
si la rente est a terme echoir.}

\item{table_mort_1}{un objet de la classe \code{ParamTableMort} designant la table de mortalite de la tete principale.}

\item{age_1}{la valeur \code{integer} de l'age de la tete principale.}

\item{gen_1}{la valeur \code{integer} de la generation de la tete principale.}

\item{table_mort_2}{un objet de la classe \code{ParamTableMort} designant la table de mortalite de la tete secondaire.}

\item{age_2}{la valeur \code{integer} de l'age de la tete secondaire.}

\item{gen_2}{la valeur \code{integer} de la generation de la tete secondaire.}
}
\value{
Le coefficient actuariel calcule.
}
\description{
\code{calc_axy} est une methode permettant de calculer les coefficients actuariels
pour des rentes avec reversion versant potentiellement plusieurs flux dans l'annee.
}
\seealso{
Le calcul des probabilites de survies \code{\link{calc_proba_survie}}.
}
\author{
Prim'Act
}
