% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChocSolvabilite2_calc_bscr.R
\docType{methods}
\name{calc_bscr}
\alias{calc_bscr}
\title{Permet de calculer le BSCR selon la formule standard.}
\usage{
calc_bscr(table_choc, racine, pre_on, parallel, nb_coeur = 0L, ecriture_base)
}
\arguments{
\item{table_choc}{un objet de la classe \code{\link{ChocSolvabilite2}}.}

\item{racine}{un objet de la classe \code{\link{Initialisation}}.}

\item{pre_on}{une valeur \code{logical} qui lorsqu'elle vaut \code{TRUE} prend en compte la variation
de PRE dans le resultat technique utilisee pour le calcul de la participation aux benefices reglementaires.}

\item{parallel}{une valeur \code{logical} qui indique si les calculs seront parallelises.}

\item{nb_coeur}{une valeur \code{integer} qui indique le nombre de coeurs utilises dans le cas ou les calculs sont
parallelises. Par defaut cette valeur est egale a 0.}

\item{ecriture_base}{un \code{logical} qui indique si les output doivent etre ecrits dans la base \code{SQLite}.}
}
\value{
une liste de \code{matrix} contenant les valeurs des differents sous-SCR ainsi que le BSCR.
}
\description{
\code{calc_bscr} est une methode permettant de lancer
l'ensemble des projections necessaires au calcul du BSCR.
au sens de la formule standard de la directive Solvabilite 2.
}
\author{
Prim'Act
}
