% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_flux.R
\docType{methods}
\name{calc_coupon}
\alias{Oblig}
\alias{calc_coupon}
\title{Calcul le coupon des models points constituant le portefeuille obligataire.}
\usage{
calc_coupon(x)
}
\arguments{
\item{x}{un objet de la classe Oblig.}
}
\value{
Un vecteur dont chaque element correspond a la valeur du coupon de l'obligation consideree : tx_coupon * parite * nominal * nb_unit.
Le vecteur renvoye a autant d'elements que le portefeuille obligataire en input a de lignes.
}
\description{
\code{calc_coupon} est une methode permettant de calculer les valeurs de coupon de l'ensemble des obligations
composant un portefeuille obligataire.
}
\author{
Prim'Act
}

