% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppb_dotation_reprise.R
\docType{methods}
\name{calc_dotation_ppb}
\alias{calc_dotation_ppb}
\title{Dote la valeur de la PPB}
\usage{
calc_dotation_ppb(x, montant)
}
\arguments{
\item{x}{objet de la classe \code{\link{Ppb}}.}

\item{montant}{une valeur \code{numeric} a doter.}
}
\value{
\code{ppb} l'objet \code{x} mis a jour.

\code{dotation} le montnant de la dotation effectuee.
}
\description{
\code{calc_dotation_ppb} est une methode permettant de doter la PPB. La dotation est effectuee si les
limites de dotation de la PPB sur l'annee ne sont pas atteintes. La valeur de cette limite est mise a jour suite a la dotation.
}
\author{
Prim'Act
}
