% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Canton_calc_fin_proj.R
\docType{methods}
\name{calc_fin_proj}
\alias{calc_fin_proj}
\title{calcule le flux et les resultats ajustes en fin de projection.}
\usage{
calc_fin_proj(x, resultat_fin, result_tech, pm_fin_ap_pb, tx_pb, tx_enc_moy)
}
\arguments{
\item{x}{est un objet de la classe \code{\link{Canton}}.}

\item{resultat_fin}{est la valeur \code{numeric} du resultat financier avant fin de projection.}

\item{result_tech}{est la valeur \code{numeric} du resultat technique avant fin de projection.}

\item{pm_fin_ap_pb}{est un vecteur \code{numeric} par produit
correspond au PM de fin avant application de la fin de projection.}

\item{tx_pb}{est un vecteur \code{numeric} par produit
correspond au taux de PB contractuel.}

\item{tx_enc_moy}{est un vecteur \code{numeric} par produit
correspond au taux chargement sur encours moyens.}
}
\value{
\code{flux_fin_passif} un vecteur de flux de fin par produit.

\code{result_tech} le montant de resultat technique en fin de projection.

\code{result_fin} le montant de resultat finanacier en fin de projection.

\code{result_brut} le montant de resultat brut d'impot en fin de projection.

\code{result_net} le montant de resultat net d'impot en fin de projection.

\code{impot} le montant d'impot sur le resultat en fin de projection.
}
\description{
\code{calc_fin_proj} est une methode permettant de calculer au niveau du canton les resultats financier, technique,
brut et net d'impot, ainsi que le flux de passifs soldant une projection.
}

