% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_flux_annee.R
\docType{methods}
\name{calc_flux_annee}
\alias{Oblig}
\alias{calc_flux_annee}
\title{Calcul les flux percus dans l'annee du fait de la detention des obligations du portefeuille obligataire.}
\usage{
calc_flux_annee(x)
}
\arguments{
\item{x}{un objet de la classe Oblig.}
}
\value{
Une liste composée de deux vecteurs:
\describe{
\item{\code{tombee_coupon} : }{Chaque element correspond aux tombees de coupon pour l'annee a venir. Ce vecteur a autant d'elements
que le portefeuille obligataire d'inputs a de lignes.}
\item{\code{tombee_echeance} : }{Chaque element correspond aux tombees d echeances pour l'annee a venir. Ce vecteur a autant d'elements
que le portefeuille obligataire d'inputs a de lignes.}
}
}
\description{
\code{calc_flux_annee} est une methode permettant de calculer les valeurs de nominal de l'ensemble des obligations
composant un portefeuille obligataire.
}
\author{
Prim'Act
}

