% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraisPassif-calc_frais.R
\docType{methods}
\name{calc_frais}
\alias{calc_frais}
\title{Calcule des frais de passif.}
\usage{
calc_frais(x, type, nom_prod, nb, mt, coef_inf)
}
\arguments{
\item{x}{objet de la classe \code{\link{FraisPassif}}.}

\item{type}{un \code{character} designant le type de frais applique.}

\item{nom_prod}{est le nom de produit de type \code{character}.}

\item{nb}{correspond a un nombre de contrats, utilise comme assiette de frais fixe par contrat.}

\item{mt}{correspond a un montant, utilise comme assiette de frais variable.}

\item{coef_inf}{correspond au coefficient d'inflation applique.}
}
\value{
Une liste contenant les montants de frais fixes et de frais variables.
}
\description{
\code{calc_frais} est une methode generique permettant de calculer les frais sur prestations, sur primes
et sur encours.
}
\details{
Le type du contrat prend pour valeur \code{prime} pour les frais sur primes, \code{prest} pour les frais
sur prestations et \code{enc} pour les frais sur encours.
}
\author{
Prim'Act
}
