% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraisFin_calc.R
\docType{methods}
\name{calc_frais_fin}
\alias{calc_frais_fin}
\title{Calcul des frais financier.}
\usage{
calc_frais_fin(x, vm_moy, coef_inflation)
}
\arguments{
\item{x}{est un objet de type \code{FraisFin} contenant les parametres des frais financiers associes a un canton.}

\item{vm_moy}{est un objet de type \code{numeric} correspondant a la valeur moyenne de l'actif en valeur
de marche.}

\item{coef_inflation}{est un objet de type \code{numeric} correspondant au coefficient d'inflation des frais.}
}
\value{
La valeur des frais financiers : un reel de type \code{numeric}.
}
\description{
\code{calc_frais_fin} est une methode permettant de calculer les frais financiers.
}
\author{
Prim'Act
}

