% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppb_calc_ppb_attrib.R
\docType{methods}
\name{calc_pb_attrib}
\alias{calc_pb_attrib}
\title{Calcule le montant de PB attribuee depuis le debut d'annee.}
\usage{
calc_pb_attrib(x)
}
\arguments{
\item{x}{un objet de la classe \code{\link{Ppb}}.}
}
\value{
la valeur \code{numeric} correspondant au montant de la pb attribuee.
}
\description{
\code{calc_pb_attrib} est une methode permettant de calculer le montant de PB attribuee au cours d'une annee.
}
\author{
Prim'Act
}
