% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortFin_calc_value.R
\docType{methods}
\name{calc_pmvl}
\alias{PortFin}
\alias{calc_pmvl}
\title{Mets a jour les sous totaux de d'actions et immobilier en plus ou moins value latente.}
\usage{
calc_pmvl(x)
}
\arguments{
\item{x}{objet de la classe PortFin.}
}
\value{
L'objet PortFin dont la somme des composantes en PVL et en MVL a ete mise a jour
}
\description{
\code{calc_pmvl} est une methode permettant de calculer les valeurs de marche.
}
\author{
Prim'Act
}

