% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action_calc_pmvl.R
\docType{methods}
\name{calc_pmvl_action}
\alias{calc_pmvl_action}
\title{Calcule les plus et moins-values action.}
\usage{
calc_pmvl_action(x)
}
\arguments{
\item{x}{objet de la classe \code{\link{Action}} (decrivant le portefeuille d'action).}
}
\value{
\code{pvl} correspondant a la somme des plus-values latentes actions.

\code{mvl} correspondant a la somme des moins-values latentes actions.
}
\description{
\code{calc_pmvl_action} est une methode permettant de calculer les plus et moins-values du portefeuille action.
}
\author{
Prim'Act
}
