% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immo_calc_pmvl.R
\docType{methods}
\name{calc_pmvl_immo}
\alias{calc_pmvl_immo}
\title{Calcule les plus et moins-values immobilier.}
\usage{
calc_pmvl_immo(x)
}
\arguments{
\item{x}{objet de la classe \code{\link{Immo}} (decrivant le portefeuille d'immobilier).}
}
\value{
\code{pvl} correspondant a la somme des plus-values latentes immobilier.

\code{mvl} correspondant a la somme des moins-values latentes immobilier.
}
\description{
\code{calc_pmvl_immo}  est une methode permettant de calculer les plus et moins-values du portefeuille immobilier.
}
\author{
Prim'Act
}
