% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_calc_pmvl.R
\docType{methods}
\name{calc_pmvl_oblig}
\alias{calc_pmvl_oblig}
\title{Calcule les plus et moins-values obligataires.}
\usage{
calc_pmvl_oblig(x)
}
\arguments{
\item{x}{objet de la classe \code{\link{Oblig}} (decrivant le portefeuille d'obligations).}
}
\value{
\code{pvl} correspondant a la somme des plus-values latentes obligataires.

\code{mvl} correspondant a la somme des moins-values latentes obligataires.
}
\description{
\code{calc_pmvl_oblig} est une methode permettant de calculer les plus et moins-values du portefeuille obligataire.
}
\author{
Prim'Act
}

