% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_calc.R
\docType{methods}
\name{calc_PRE}
\alias{PRE}
\alias{calc_PRE}
\title{Calcul de la PRE.}
\usage{
calc_PRE(x, pmvl_action_immo)
}
\arguments{
\item{x}{objet de la classe \code{PRE}, necessaire pour connaitre le stock de PRE initial.}

\item{pmvl_action_immo}{est un \code{numeric} correspondant au montant global de plus ou moins values latentes de actifs actions et immobiliers.
En cas de moins value latente, la PRE est abondee.
En cas de plus value latente, la PRE est integralement reprise.}
}
\value{
Le format de la liste renvoyee est :
\describe{
\item{\code{pre_courante} : }{ valeur de la pre calculee a partir des inputs transmis}
\item{\code{var_pre} : }{ variation de la pre courante}
}
}
\description{
\code{calc_PRE} est une methode permettant de calculer le montant de PRE.
}
\author{
Prim'Act
}

