% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpEuroInd-calc_pm.R, R/EpEuroInd-calc_prest.R
\docType{methods}
\name{calc_prest}
\alias{EpEuroInd}
\alias{calc_prest}
\title{Calcul les PM d'un model point epargne en euros.}
\usage{
calc_pm(x, tab_prime, tab_prest, tx_cible, tx_min, an, method, tx_soc)

calc_prest(x, tx_sortie, tx_min, an, method, tx_soc)
}
\arguments{
\item{x}{un objet de la classe \code{EpEuroInd} contenant les model points epargne euros.}

\item{tx_min}{un data.frame contenant le taux de revalorisation minimum associes a chaque ligne de model points :
\describe{
\item{\code{tx_min_an : }}{un vecteur de taux annuels}
\item{\code{tx_min_se : }}{un vecteur de taux semestriels}
}}

\item{an}{un numeric represantant l'annee de projection courante}

\item{method}{un character prenant pour valeur \code{normal} pour le calcul des flux avec application de la revalorisation au titre
participation aux benefices, et la valeur \code{gar} pour le calcul avec uniquement les flux garanti (calcul de la FDB).}

\item{tx_sortie}{une liste contenant les taux de sortie associes a chaque ligne de model points :
\describe{
\item{\code{qx_rach_tot : }}{un vecteur avec les taux de rachat total structurel}
\item{\code{qx_rach_tot_dyn : }}{un vecteur avec les taux de rachat total dynamique}
\item{\code{qx_dc : }}{un vecteur avec les taux de deces}
\item{\code{qx_rach_part : }}{un vecteur avec les taux de rachats partiel structurel}
\item{\code{qx_rach_part_dyn : }}{un vecteur avec les taux de rachats partiel dynamique}
}}

\item{x}{un objet de la classe \code{EpEuroInd} contenant les model points epargne euros.}

\item{tx_sortie}{une liste contenant les taux de sortie associes a chaque ligne de model points :
\describe{
\item{\code{qx_rach_tot : }}{un vecteur avec les taux de rachat total structurel}
\item{\code{qx_rach_tot_dyn : }}{un vecteur avec les taux de rachat total dynamique}
\item{\code{qx_dc : }}{un vecteur avec les taux de deces}
\item{\code{qx_rach_part : }}{un vecteur avec les taux de rachats partiel structurel}
\item{\code{qx_rach_part_dyn : }}{un vecteur avec les taux de rachats partiel dynamique}
}}

\item{tx_min}{un data.frame contenant le taux de revalorisation minimum associes a chaque ligne de model points :
\describe{
\item{\code{tx_min_an : }}{un vecteur de taux annuels}
\item{\code{tx_min_se : }}{un vecteur de taux semestriels}
}}

\item{an}{un numeric represantant l'annee de projection courante}

\item{method}{un character prenant pour valeur \code{normal} pour le calcul des flux avec application de la revalorisation au titre
participation aux benefices, et la valeur \code{gar} pour le calcul avec uniquement les flux garanti (calcul de la FDB).}
}
\value{
Une liste contenant :
\describe{
\item{\code{method} : }{la valeur de l'argument \code{method}}
\item{\code{flux} : }{un data.frame comprenant les flux de l'annee}
\item{\code{nb} : }{un data.frame comprenant les nombres de sorties}
}
Le format de \code{flux} est :
\describe{
\item{\code{ech} : }{un vecteur contenant les flux de sortie en echeance de l'annee}
\item{\code{rach_tot} : }{un vecteur contenant les flux de rachat totaux de l'annee}
\item{\code{dc} : }{un vecteur contenant les flux de deces de l'annee}
\item{\code{rach_part} : }{un vecteur contenant les flux de rachat partiel de l'annee}
\item{\code{prest} : }{un vecteur contenant les flux prestations de l'annee}
\item{\code{rev_ech} : }{un vecteur contenant la revalorisation des echeances de l'annee}
\item{\code{rev_rach_tot} : }{un vecteur contenant la revalorisation des rachats totaux de l'annee}
\item{\code{rev_dc} : }{un vecteur contenant la revalorisation des deces de l'annee}
\item{\code{rev_rach_part} : }{un vecteur contenant la revalorisation des rachats partiels de l'annee}
\item{\code{rev} : }{un vecteur contenant la revalorisation des prestations de l'annee}
}
Le format de \code{nb} est :
\describe{
\item{\code{nb_ech : }}{un vecteur contenant le nombre de sorties en echeance de l'annee}
\item{\code{nb_rach_tot : }}{un vecteur contenant le nombre de rachats totaux de l'annee}
\item{\code{nb_dc : }}{un vecteur contenant le nombre de deces de l'annee}
\item{\code{nb_sortie : }}{un vecteur contenant le nombre de sorties de l'annee}
\item{\code{nb_contr_fin : }}{un vecteur contenant le nombre de contrats en cours en fin d'annee}
}

Une liste contenant :
\describe{
\item{\code{method} : }{la valeur de l'argument \code{method}}
\item{\code{flux} : }{un data.frame comprenant les flux de l'annee}
\item{\code{nb} : }{un data.frame comprenant les nombres de sorties}
}
Le format de \code{flux} est :
\describe{
\item{\code{ech} : }{un vecteur contenant les flux de sortie en echeance de l'annee}
\item{\code{rach_tot} : }{un vecteur contenant les flux de rachat totaux de l'annee}
\item{\code{dc} : }{un vecteur contenant les flux de deces de l'annee}
\item{\code{rach_part} : }{un vecteur contenant les flux de rachat partiel de l'annee}
\item{\code{prest} : }{un vecteur contenant les flux prestations de l'annee}
\item{\code{rev_ech} : }{un vecteur contenant la revalorisation des echeances de l'annee}
\item{\code{rev_rach_tot} : }{un vecteur contenant la revalorisation des rachats totaux de l'annee}
\item{\code{rev_dc} : }{un vecteur contenant la revalorisation des deces de l'annee}
\item{\code{rev_rach_part} : }{un vecteur contenant la revalorisation des rachats partiels de l'annee}
\item{\code{rev_prest} : }{un vecteur contenant la revalorisation des prestations de l'annee}
\item{\code{rev_prest_nette} : }{un vecteur contenant la revalorisation des prestations nette de l'annee}
\item{\code{enc_charg} : }{un vecteur contenant les chargements sur l'encours de l'annee}
\item{\code{rach_charg} : }{un vecteur contenant les chargements sur les rachats de l'annee}
\item{\code{soc_prest} : }{un vecteur contenant les prelevements sociaux sur prestations de l'annee}
}
Le format de \code{nb} est :
\describe{
\item{\code{nb_ech : }}{un vecteur contenant le nombre de sorties en echeance de l'annee}
\item{\code{nb_rach_tot : }}{un vecteur contenant le nombre de rachats totaux de l'annee}
\item{\code{nb_dc : }}{un vecteur contenant le nombre de deces de l'annee}
\item{\code{nb_sortie : }}{un vecteur contenant le nombre de sorties de l'annee}
\item{\code{nb_contr_fin : }}{un vecteur contenant le nombre de contrats en cours en fin d'annee}
}
}
\description{
\code{calc_prest} est une methode permettant de calculer les PM sur une annee.

\code{calc_prest} est une methode permettant de calculer les flux de prestations,
les chargements sur encours relatifs a ces prestations et les nombres de sorties sur une annee.
}
\author{
Prim'Act

Prim'Act
}

