% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpEuroInd-calc_primes.R
\docType{methods}
\name{calc_primes}
\alias{EpEuroInd}
\alias{calc_primes}
\title{Calcul les flux de primes d'un model point epargne en euros.}
\usage{
calc_primes(x)
}
\arguments{
\item{x}{un objet de la classe \code{EpEuroInd} contenant les model points epargne euros}
}
\value{
Un data.frame contenant :
\describe{
\item{\code{pri_brut} : }{un vecteur contenant les flux de primes brutes de chargements de l'annee}
\item{\code{pri_net} : }{un vecteur contenant les flux de primes nettes de chargements de l'annee}
\item{\code{pri_chgt} : }{un vecteur contenant les flux de chargements de l'annee}
}
}
\description{
\code{calc_primes} est une methode permettant de calculer les flux de primes sur
 une annee.
}
\author{
Prim'Act
}

