% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamTableMort-calc_proba_deces.R
\docType{methods}
\name{calc_proba_deces}
\alias{calc_proba_deces}
\title{Calcule le vecteur des probabilites cumulees de deces}
\usage{
calc_proba_deces(table_mort, age, gen, n_periodes)
}
\arguments{
\item{table_mort}{un objet de la classe \code{\link{ParamTableMort}} contenant la table de mortalite.}

\item{age}{une valeur \code{integer} correspondant a l'age.}

\item{gen}{une valeur \code{integer} correspondant a la generation.}

\item{n_periodes}{une valeur \code{integer} correspondant a la borne de sommation.}
}
\value{
La valeur du taux de deces calcule.
}
\description{
\code{calc_proba_deces} est une methode permettant de calculer le vecteur des probabilites cumulees de deces.
}
\author{
Prim'Act
}
