% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpEuroInd-calc_proba_dyn.R
\docType{methods}
\name{calc_proba_dyn}
\alias{calc_proba_dyn}
\title{Calcul les probabilites dynamiques de mouvement de flux pour des contrats epargne en euros et de retraite.}
\usage{
calc_proba_dyn(x, ht)
}
\arguments{
\item{x}{un objet de la classe \code{\link{EpEuroInd}} contenant les model points epargne euros.}

\item{ht}{un objet de la classe \code{\link{HypTech}} contenant differentes tables de mortalite et differentes
lois de rachat.}
}
\value{
Une matrice contenant pour chaque model points en ligne :
\describe{
\item{\code{qx_rach_tot_dyn} : }{un vecteur contenant les taux de rachats totaux dynamiques}
\item{\code{qx_rach_part_dyn} : }{un vecteur contenant les taux de rachats partiels dynamiques}.
}
}
\description{
\code{calc_proba_dyn} est une methode permettant de calculer les differents taux de sortie dynamique
  sur une periode.
}
\seealso{
La recuperation des taux de rachat dynamique : \code{\link{get_rach_dyn}}.
}
\author{
Prim'Act
}
