% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamTableMort-calc_proba_survie.R
\docType{methods}
\name{calc_proba_survie}
\alias{calc_proba_survie}
\title{Calcule le vecteur des probabilites cumulees de survies.}
\usage{
calc_proba_survie(table_mort, age, gen, n_periodes)
}
\arguments{
\item{table_mort}{un objet de la classe \code{\link{ParamTableMort}} contenant la table de mortalite.}

\item{age}{une valeur \code{integer} correspondant a l'age.}

\item{gen}{une valeur \code{integer} correspondant a la generation.}

\item{n_periodes}{une valeur \code{integer} correspondant a la borne de sommation.}
}
\value{
Le vecteur des taux de deces calcules.
}
\description{
\code{calc_proba_survie} est une methode permettant de calculer le vecteur des probabilites cumulees de survie.
}
\author{
Prim'Act
}
