% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamTableMort-calc_qx.R
\docType{methods}
\name{calc_qx}
\alias{ParamTableMort}
\alias{calc_qx}
\title{Calcule le taux de mortalite}
\usage{
calc_qx(table_mort, age, gen)
}
\arguments{
\item{table_mort}{un objet de la classe \code{ParamTableMort} contenant les tables de mortalite.}

\item{age}{: en numerique}

\item{gen}{: generation en numerique}
}
\value{
le taux de mortalite
}
\description{
\code{calc_qx} est une methode permettant de calculer le taux de mortalite.
}
\author{
Prim'Act
}

