% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamTableRach-calc_rach.R
\docType{methods}
\name{calc_rach}
\alias{ParamTableRach}
\alias{calc_rach}
\title{Calcule le taux de rachat total ou partiel}
\usage{
calc_rach(table_rach, age, anc)
}
\arguments{
\item{table_rach}{un objet de la classe \code{ParamTableRach} contenant les tables de rachat.}

\item{age}{: en numerique}

\item{anc}{: anciennete en numerique}
}
\value{
le taux de rachat total ou partiel
}
\description{
\code{calc_rach} est une methode permettant de calculer le taux de rachat total ou partiel.
}
\author{
Prim'Act
}

