% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamRachDyn-calc_rach_dyn.R
\docType{methods}
\name{calc_rach_dyn}
\alias{ParamRachDyn}
\alias{calc_rach_dyn}
\title{Calcule la composante rachats dynamique}
\usage{
calc_rach_dyn(p, tx_cible, tx_serv)
}
\arguments{
\item{p}{un objet de la classe \code{ParamRachDyn} contenant les parametres de rachats dynamiques.}

\item{tx_cible}{: taux cible en numerique}

\item{tx_serv}{: taux servi en numerique}
}
\value{
la composante rachats dynamique en numerique
}
\description{
\code{calc_rach_dyn} est une methode permettant de calculer la composante rachats dynamique.
}
\author{
Prim'Act
}

