% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortFin_calc_rdt.R
\docType{methods}
\name{calc_rdt}
\alias{PortFin}
\alias{calc_rdt}
\title{Calcul les rendements de chacune des composante des sous-portefeuilles action et immobilier du portefeuille PortFin.}
\usage{
calc_rdt(x, mp_ESG)
}
\arguments{
\item{x}{objet de la classe PortFin.}

\item{mp_ESG}{objet de la classe ModelPointESG decrivant les conditions de l'annee n ( ainsi que l'annee n-1 pour les indices actions & immo).}
}
\value{
Un data frame composé de deux colonnes et autant de lignes que le portefeuille action a de lignes.
}
\description{
\code{calc_rdt} est une methode permettant de calculer les rendements des portfeuilles Action et Immo d'un objet PortFin.
}
\author{
Prim'Act
}

