% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortPassif-calc_rdt_marche_ref.R
\docType{methods}
\name{calc_rdt_marche_ref}
\alias{calc_rdt_marche_ref}
\title{Calcul du taux de rendement de reference au niveau du marche}
\usage{
calc_rdt_marche_ref(x, mp_esg)
}
\arguments{
\item{mp_esg}{est un objet de type \code{ModelPointESG}, qui represente la situation courante
en annee et simulations des valeurs de l'ESG.}

\item{param_comport}{un objet de la classe \code{ParamComport}.}
}
\value{
Une liste contenant les rendements de reference du marche.
}
\description{
\code{calc_rdt_marche_ref} est une methode permettant de calculer un taux cible.
}
\author{
Prim'Act
}
