% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppb_dotation_reprise.R
\docType{methods}
\name{calc_reprise_ppb}
\alias{calc_reprise_ppb}
\title{Reprend sur la valeur de la PPB}
\usage{
calc_reprise_ppb(x, montant)
}
\arguments{
\item{x}{un objet de la classe \code{Ppb}.}

\item{montant}{la valeur \code{numeric} de la reprise.}
}
\value{
\code{ppb} l'objet \code{x} mis a jour

\code{reprise} le montnant de la reprise effectuee.
}
\description{
\code{calc_reprise_ppb} est une methode permettant de reprendre sur la PPB.
La reprise est effectuee si les limites de reprise de la PPB sur l'annee ne sont pas atteintes.
La valeur de cette limite est mise a jour suite a la reprise
}
\author{
Prim'Act
}
