% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevaloEngine_calc_result_technique.R
\docType{methods}
\name{calc_result_technique}
\alias{RevaloEngine}
\alias{calc_result_technique}
\title{calcule le resultat technique}
\usage{
calc_result_technique(passif_av_pb, var_pre)
}
\arguments{
\item{passif_av_pb}{est une liste produit par la methode \code{viellissement_av_pb}
appliquee a un portefeuille de passif.}

\item{var_pre}{est une valeur \code{numeric} correspondant a la variation de PRE.}
}
\value{
Le resulat technique
}
\description{
\code{calc_result_technique} est une methode permettant de calculer le resultat technique avant attribution de
participation aux benefices.
}

