% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Canton_calc_result_technique_ap_pb.R
\docType{methods}
\name{calc_result_technique_ap_pb}
\alias{calc_result_technique_ap_pb}
\title{calcule le resultat technique apres prise en compte de la participation aux benefices.}
\usage{
calc_result_technique_ap_pb(passif_av_pb, passif_ap_pb, ppb, var_pre)
}
\arguments{
\item{passif_av_pb}{est une liste produit par la methode \code{\link{viellissement_av_pb}}.}

\item{passif_ap_pb}{est une liste produit par la methode \code{\link{viellissement_ap_pb}}.}

\item{ppb}{est un objet de la classe \code{\link{Ppb}} qui renvoie l'etat courant de la PPB.}

\item{var_pre}{est une valeur \code{numeric} correspondant a la variation de PRE.}
}
\value{
Le resultat technique apres participation aux benefices.
}
\description{
\code{calc_result_technique_ap_pb} est une methode permettant de calculer le resultat technique
apres attribution de participation aux benefices.
}
