% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_calc_sur_dec.R
\docType{methods}
\name{calc_sur_dec}
\alias{calc_sur_dec}
\title{Calcul les surcotes/decotes de chaque composante d'un portefeuille obligataire.}
\usage{
calc_sur_dec(x)
}
\arguments{
\item{x}{objet de la classe \code{\link{Oblig}} (decrivant le portefeuille obligataire).}
}
\value{
Un vecteur contenant les surcotes decotes.
}
\description{
\code{calc_sur_dec} est une methode permettant de calculer les surcotes/decotes de chaque composante d'un portefeuille obligataire.
}
\author{
Prim'Act
}

