% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortFin_calc_tra.R
\docType{methods}
\name{calc_tra}
\alias{calc_tra}
\title{Calcul du taux de rendement financier}
\usage{
calc_tra(plac_moy, res_fin)
}
\arguments{
\item{plac_moy}{est un objet de type \code{numeric}, qui fournit la valeur moyenne
des placements de l'annee en valeur nette comptable.}

\item{res_fin}{est un objet de type \code{numeric}, qui fournit le resultat financier du porfeuille.}
}
\value{
La valeur du taux de rendement de l'actif.
}
\description{
\code{calc_tra} est une methode permettant de calculer le taux de rendement financier du portefeuille.
}
\author{
Prim'Act
}
