% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpEuroInd-calc_tx_cible.R
\docType{methods}
\name{calc_tx_cible}
\alias{EpEuroInd}
\alias{calc_tx_cible}
\title{Calcul des differents taux cible pour chaque model point de la classe EpEuroInd}
\usage{
calc_tx_cible(epi, ht, list_rd)
}
\arguments{
\item{epi}{un objet de la classe \code{EpEuroInd} contenant les model points epargne euros.}

\item{ht}{un objet de la classe \code{HypTech} contenant differentes tables en entree.}

\item{list_rd}{une liste contenant les rendements de reference.}
}
\value{
Une liste contenant le taux cible annuel et le taux cible semestriel par model points.
}
\description{
\code{calc_tx_cible} est une methode permettant de calculer un taux cible
  pour chaque model point de la classe EpEuroInd.
}
\author{
Prim'Act
}

