% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamComport-calc_tx_cible.R
\docType{methods}
\name{calc_tx_cible_ref_marche}
\alias{ParamComport}
\alias{calc_tx_cible_ref_marche}
\title{Calcul des differents taux cible pour chaque model point de la classe EpEuroInd}
\usage{
calc_tx_cible_ref_marche(param_comport, list_rd, tx_cible_prec)
}
\arguments{
\item{param_comport}{un objet de la classe \code{ParamComport}.}

\item{list_rd}{une liste qui comprend les taux de rendement de reference.}

\item{tx_cible_prec}{le taux cible individuel de l'annee precedente.}
}
\value{
Le taux cible annuel.
}
\description{
\code{calc_tx_cible_ref_marche} est une methode permettant de calculer un taux cible.
}
\author{
Prim'Act
}

