% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpEuroInd-calc_tx_min.R
\docType{methods}
\name{calc_tx_min}
\alias{EpEuroInd}
\alias{calc_tx_min}
\title{Calcul des differents taux  revalorisation minimum d'un model point.}
\usage{
calc_tx_min(x, an)
}
\arguments{
\item{x}{un objet de la classe \code{EpEuroInd} contenant les model points epargne euros.}

\item{an}{un numeric represant l'annee de projection courante}
}
\value{
Une data.frame contenant :
\describe{
\item{\code{tx_tech_an} : }{un vecteur contenant les taux de technique de l'annee}
\item{\code{tx_tech_an} : }{un vecteur contenant les taux de technique de l'annee sur base semestrielle}
\item{\code{tx_an} : }{un vecteur contenant les taux de revalorisation minimum de l'annee}
\item{\code{tx_se} : }{un vecteur contenant les taux de revalorisation minimum de l'annee exprimes en semestriel}
}
}
\description{
\code{calc_tx_min} est une methode permettant de calculer le taux de revalorisation minimum sur
 une annee.
}
\author{
Prim'Act
}

