% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpEuroInd-calc_tx_sortie.R
\docType{methods}
\name{calc_tx_sortie}
\alias{EpEuroInd}
\alias{calc_tx_sortie}
\title{Calcul des differents taux de sortie pour chaque model point de la classe EpEuroInd}
\usage{
calc_tx_sortie(epi, ht)
}
\arguments{
\item{epi}{un objet de la classe \code{EpEuroInd} contenant les model points epargne euros.}

\item{ht}{un objet de la classe \code{HypTech} contenant differentes tables en entree.}
}
\value{
Une data.frame contenant :
\describe{
\item{\code{qx_rach_tot} : }{un vecteur contenant les taux de rachats totaux}
\item{\code{qx_rach_tot_dyn} : }{un vecteur contenant les taux de rachats totaux dynamiques}
\item{\code{qx_dc} : }{un vecteur contenant les taux de deces}
\item{\code{qx_rach_part} : }{un vecteur contenant les taux de rachats partiels}
\item{\code{qx_rach_part_dyn} : }{un vecteur contenant les taux de rachats partiels dynamiques}
}
}
\description{
\code{calc_tx_sortie} est une methode permettant de calculer les differents taux de sortie
  pour chaque model point de la classe EpEuroInd.
}
\author{
Prim'Act
}

