% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action_calc_vm.R
\docType{methods}
\name{calc_vm_action}
\alias{calc_vm_action}
\title{Calcule les valeurs de marches du portefeuille action.}
\usage{
calc_vm_action(x, rdt)
}
\arguments{
\item{x}{objet de la classe \code{\link{Action}} (decrivant le portefeuille d'action).}

\item{rdt}{vecteur de type \code{numeric} decrivant le rendement de chacune des lignes du portefeuille action
 de l'assureur.
Contient autant d'elements que le portefeuille action a de lignes.}
}
\value{
L'objet \code{x} dont les valeurs de marche ont ete mises a jour.
}
\description{
\code{calc_vm_action} est une methode permettant de calculer les valeurs de marche du portefeuille action.
}
\author{
Prim'Act
}
