% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immo_calc_value.R
\docType{methods}
\name{calc_vm_immo}
\alias{Treso}
\alias{calc_vm_immo}
\title{Calcul les valeurs de marches de chaque composante du portefeuille immo.}
\usage{
calc_vm_immo(x, rdt)
}
\arguments{
\item{x}{objet de la classe Immo (decrivant le portefeuille d'immo).}

\item{rdt}{vecteur decrivant le rendement de chacun des biens immo du ptf.
Contient autant d'elements que le portefeuille a de lignes.}
}
\value{
L'objet x dont les valeurs de marche ont ete mises a jour.
}
\description{
\code{calc_vm_immo} est une methode permettant de calculer les valeurs de marche.
}
\author{
Prim'Act
}

