% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_calc_value.R
\docType{methods}
\name{calc_vm_oblig}
\alias{Oblig}
\alias{calc_vm_oblig}
\title{Calcul les valeurs de marches de chaque composante d'un portefeuille obligataire.}
\usage{
calc_vm_oblig(x, yield_curve)
}
\arguments{
\item{x}{objet de la classe Oblig (decrivant le portefeuile obligataire).}

\item{yield_curve}{vecteur decrivant la courbe de taux sans risque retenue.)}
}
\value{
Un vecteur dont chaque element correspond a la valeur de marche de l'obligation du portefeuille obligataire.
Ce vecteur a autant d'elements que le portefeuille obligataire a de lignes.
}
\description{
\code{calc_vm_oblig} est une methode permettant de calculer les valeurs de marche.
}
\author{
Prim'Act
}

