% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Treso_calc_value.R
\docType{methods}
\name{calc_vm_treso}
\alias{Treso}
\alias{calc_vm_treso}
\title{Calcul les valeurs de marches de chaque composante du portefeuille treso.}
\usage{
calc_vm_treso(x, rdt, flux_milieu, flux_fin)
}
\arguments{
\item{x}{objet de la classe treso (decrivant le portefeuille de treso).}

\item{rdt}{vecteur decrivant le rendement de chacune des lignes treso du ptf.
Contient autant d'elements que le portefeuille a de lignes.}

\item{flux_milieu}{vecteur decrivant les flux ( percus)entrants : positif, sortants : negatifs) en milieu d'annee, ventile selon chacune des lignes de cash.}

\item{flux_fin}{vecteur decrivant les flux (entrants : positifs, sortants : negatifs) en fin d'annee, ventile selon chacune des lignes de cash.}
}
\value{
L'objet x dont les valeurs de marche ont ete mises a jour.
}
\description{
\code{calc_vm_treso} est une methode permettant de calculer les valeurs de marche de chaque ligne du portefeuille treso.
}
\author{
Prim'Act
}

