% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_calc_vnc.R
\docType{methods}
\name{calc_vnc}
\alias{calc_vnc}
\title{Calcul les valeurs nettes comptables de chaque composante portefeuille obligataires.}
\usage{
calc_vnc(x, sd_unitaire)
}
\arguments{
\item{x}{objet de la classe \code{\link{Oblig}} (decrivant le portefeuille d'obligation).}

\item{sd_unitaire}{vecteur de type \code{numeric} decrivant la surcote decote de chacune des lignes d'obligation du portefeuille obligation de l'assureur.
Contient autant d'elements que le portefeuille a de lignes.}
}
\value{
L'objet \code{x} dont les valeurs nettes comptables ont ete mises a jour.
}
\description{
\code{calc_vnc} est une methode permettant de calculer les valeurs de marche.
}
\author{
Prim'Act
}
