% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_calc_z_spread.R
\docType{methods}
\name{calc_z_spread}
\alias{Oblig}
\alias{calc_z_spread}
\title{Calcul les zeros spreads de chaque composante d'un portefeuille obligataire.}
\usage{
calc_z_spread(x, yield_curve)
}
\arguments{
\item{x}{objet de la classe Oblig (decrivant le portefeuile obligataire).}

\item{yield_curve}{vecteur decrivant la courbe de taux sans risque retenue.)}
}
\value{
Un vecteur dont chaque element correspond a la valeur du zero spread de l'obligation du portefeuille obligataire.
Ce vecteur a autant d'elements que le portefeuille obligataire a de lignes.
}
\description{
\code{calc_z_spread} est une methode permettant de calculer les zeros spread de chaque composante d'un portefeuille obligataire.
}
\author{
Prim'Act
}

