% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Canton_data.R
\docType{data}
\name{Canton_DataSet}
\alias{Canton_DataSet}
\alias{canton}
\title{Dataset Canton.}
\format{Objet de la classe \code{\link{Canton}}.}
\usage{
data(canton)
}
\description{
Dataset de la classe \code{\link{Canton}} contenant un portefeuille de passif, d'actif ainsi que plusieurs hypothhses.
}
\note{
Il s'agit de donnees simulees, elles ont ete creees afin de tester les differents fonctions du package.
}
\examples{
data("canton")
portefeuille_passif <- canton@ptf_passif
}
\author{
Prim'Act
}
\keyword{datasets}
