% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESG_chargement.R
\docType{methods}
\name{chargement_ESG}
\alias{chargement_ESG}
\title{Cette methode charge les tables de simulations d'un ESG.}
\usage{
chargement_ESG(folder_ESG_address, nb_simu, nb_annee_proj)
}
\arguments{
\item{folder_ESG_address}{est un \code{character}. Il correspond au chemin de reference du dossier contenant
les extractions de l'ESG Prim'Act.}

\item{nb_simu}{est une valeur de type \code{integer} correspondant au nombre de trajectoire
simulees par l'ESG Prim'Act.}

\item{nb_annee_proj}{est une valeur de type \code{integer} correspondant au nombre d'annees de projection
des sorties de l'ESG Prim'Act.}
}
\value{
\code{x} l'objet de la classe \code{\link{ESG}} construit.
}
\description{
\code{chargement_ESG} est une methode permettant de charger les trajectoires simulees par le generateur de
scenarios economiques (ESG) de Prim'Act et d'alimenter un objet \code{\link{ESG}}.
}
\details{
Les differentes adresses potentielles pour les differents ESG employes (central, hausse de taux, baisse de taux)
sont construites par la fonction \code{\link{set_architecture}} de la classe \code{\link{Initialisation}}.
}
\author{
Prim'Act
}

