% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortFin_chargement_reference.R
\docType{methods}
\name{chargement_PortFin_reference}
\alias{chargement_PortFin_reference}
\title{Charge le PortFin de reinvestissement a partir des donnees renseignees par l'utilisateur.}
\usage{
chargement_PortFin_reference(folder_PortFin_reference_address, mp_ESG)
}
\arguments{
\item{folder_PortFin_reference_address}{est un chemin de type \code{character}, cf la methode \code{\link{set_architecture}}}

\item{mp_ESG}{est un objet de la classe \code{\link{ModelPointESG}}, qui fournit le resultat financier du porfeuille.}
}
\value{
L'objet \code{\link{PortFin}} correspondant au portefeuille financier de reinvestissement
tel que precise par les donnees initiales et les parametres renseignes par l'utilisateur.
}
\description{
\code{chargement_PortFin_reference} est une methode permettant de creer un objet \code{PortFin} correspondant au portefeuille finanicer de reinvestissement
a partir des donnees renseignees par l'utilisateur.
}
\author{
Prim'Act
}

