% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taux_period-function.R
\docType{methods}
\name{chgt_period}
\alias{chgt_period}
\title{Fonction permettant de calculer le taux de chargement sur une periode.}
\usage{
chgt_period(x, period)
}
\arguments{
\item{x}{un vecteur de la classe \code{numeric} representant le taux annuel.}

\item{period}{un \code{character} representant la peridocite souhaitee. Cette variable prend pour valeur :
\describe{
\item{\code{an} : }{annuel ;}
\item{\code{se} : }{semestriel ;}
\item{\code{trim} : }{trimestriel ;}
\item{\code{mens} : }{mensuel.}}}
}
\value{
Le taux periodique.
}
\description{
\code{chgt_period} permet de calculer le taux de chargement sur une periode.
}
\author{
Prim'Act
}
