% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypTech-convert_table.R
\docType{methods}
\name{convert_table}
\alias{convert_table}
\title{Tranforme une table contenant des "l_x" en une table contenant des "q_x"}
\usage{
convert_table(table_mort, type)
}
\arguments{
\item{table_mort}{un objet de la classe \code{data.frame} contenant la table de mortalite.}

\item{type}{un objet de la classe \code{character} correspondant au type de la table de mortalite passee en parametre (qx ou lx).}
}
\value{
La table transformee.
}
\description{
\code{convert_table} est une methode permettant de calculer le vecteur des probabilites cumulees de survie.
}
\author{
Prim'Act
}
