% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlmEngine_create_ptf_bought_action.R
\docType{methods}
\name{create_ptf_bought_action}
\alias{create_ptf_bought_action}
\title{Ajuste les quantites d'actions a acheter.}
\usage{
create_ptf_bought_action(x, coefficient)
}
\arguments{
\item{x}{objet de la classe \code{\link{Action}}, correspondant au portefeuille actions de reinvestissement.
Ce portefeuille est unitaire.}

\item{coefficient}{un vecteur de type \code{numeric} qui a autant d'elements
que le portefeuille de reinvestissement action a de lignes.
 Il correspond au coefficient a appliquer au portefeuille de reinvestissement action
 pour effectuer l'achat desire.}
}
\value{
\code{x} un objet de la classe \code{\link{Action}} correspondant a une
 proportion precise du portefeuille de reinvestissement action.
}
\description{
\code{create_ptf_bought_action} est une methode permettant d'ajuster d'un coefficient les quantites
d'actions a acheter. Cette methode est utilisee pour l'achat de nouvelles actions.
}
\author{
Prim'Act
}
\seealso{
La classe \code{\link{Action}}.
}

