% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlmEngine_create_ptf_bought_immo.R
\docType{methods}
\name{create_ptf_bought_immo}
\alias{AlmEngine}
\alias{create_ptf_bought_immo}
\title{Ajuste les quantites d'immobilier a acheter.}
\usage{
create_ptf_bought_immo(x, coefficient)
}
\arguments{
\item{x}{objet de la classe \code{\link{Immo}}, correspondant au portefeuille immobilier de reinvestissement.
Ce portefeuille est unitaire.}

\item{coefficient}{est un vecteur de type \code{numeric} qui a autant d'elements
que le portefeuille de reinvestissement immo a de lignes. Il correspond au coefficient
 a appliquer au portefeuille de reinvestissement immo pour effectuer l'achat desire.}
}
\value{
\code{x} un objet de la classe \code{\link{Immo}} correspondant a une
proportion precise du portefeuille de reinvestissement immo.
}
\description{
\code{create_ptf_bought_immo} est une methode permettant d'ajuster d'un coefficient les quantites
d'immobilier a acheter. Cette methode est utilisee pour l'achat de nouveaux titres immobilier.
}
\author{
Prim'Act
}
\seealso{
La classe \code{\link{Immo}}.
}

