% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlmEngine_create_ptf_bought_oblig.R
\docType{methods}
\name{create_ptf_bought_oblig}
\alias{AlmEngine}
\alias{create_ptf_bought_oblig}
\title{Ajuste les quantites d'obligations a acheter.}
\usage{
create_ptf_bought_oblig(x, coefficient)
}
\arguments{
\item{x}{objet de la classe \code{\link{Oblig}}, correspondant au portefeuille
obligataire de reinvestissement. Ce portefeuille est unitaire.}

\item{coefficient}{est un vecteur de type \code{numeric} qui a autant d'elements
que le portefeuille de reinvestissement obligataire a de lignes. Il correspond au coefficient
 a appliquer au portefeuille de reinvestissement obligataire pour effectuer l'achat desire.}
}
\value{
\code{x} un objet de la classe \code{\link{Oblig}} correspondant a une proportion
 precise du portefeuille de reinvestissement obligataire.
}
\description{
Cette methode permet d'ajuster d'un coefficient les quantites
d'obligations a acheter. Cette methode est utilisee pour l'achat de nouveaux titres obligataires.
}
\author{
Prim'Act
}
\seealso{
La classe \code{\link{Oblig}}.
}

