% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBase_class.R
\docType{class}
\name{DataBase}
\alias{DataBase}
\title{La classe \code{DataBase}.}
\description{
Une classe permettant de stocker les donnes d'output de la fonction \code{\link{proj_an}}.
}
\section{Slots}{

\describe{
\item{\code{database}}{une base de donnees \code{SQLite}}

\item{\code{ecriture_base}}{un \code{logical} qui indique si les output doivent etre ecrits dans la base \code{SQLite}.}
}}

\author{
Prim'Act
}
\keyword{classes}
