% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlmEngine_do_calc_nb_sold_immo.R
\docType{methods}
\name{do_calc_nb_sold_immo}
\alias{do_calc_nb_sold_immo}
\title{Calcule le nombre de titres immobilier a vendre.}
\usage{
do_calc_nb_sold_immo(x, montant_vente, method_vente)
}
\arguments{
\item{x}{objet de la classe \code{\link{Immo}}, correspondant au portefeuille immo de l'assureur.}

\item{montant_vente}{est un reel de type \code{numeric} correspondant a la vente totale de vm immo
que l'assureur souhaite effectuer.}

\item{method_vente}{est un element de type \code{character} correspondant a methode de vente retenue
(seule la methode proportionnelle est implementee actuellement).}
}
\value{
\code{data.frame} contenant deux colonnes \code{(num_mp, nb_sold)} correspondant respectivement
 au numero de model point de chaque ligne immo du portefeuille et du nombre d'unite a vendre pour chacune
  d'entre elles.
}
\description{
Cette methode permet de calculer pour chaque ligne d'un portefeuille
immobilier d'un assureur le nombre d'unites a vendre afin de realiser un certain montant de vente immo.
}
\seealso{
La classe \code{\link{Immo}}.
}
\author{
Prim'Act
}
