% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChocSolvabilite2_do_choc_action_type1.R
\docType{methods}
\name{do_choc_action_type1}
\alias{do_choc_action_type1}
\title{Permet a partir d'un canton initial de creer un canton choque action.}
\usage{
do_choc_action_type1(x, canton)
}
\arguments{
\item{x}{objet de la classe \code{\link{ChocSolvabilite2}}.}

\item{canton}{un objet de la classe \code{\link{Canton}}. Il correspond au canton non choque (i.e. central) de l'assureur.}
}
\value{
\code{canton} l'objet  de la classe \code{\link{Canton}} correspondant au scenario choque action
 au sens de la formule standard Solvabilite 2.
}
\description{
\code{do_choc_action_type1} est une methode permettant d'appliquer le choc action type 1 de la formule standard Solvabilite 2 a un canton.
}
\note{
Il est possible d'appliquer des chocs actions distincts a chaque action selon l'index.
Cette parametrisation est effectuee dans les fichiers d'inputs utilisateurs.
}
\author{
Prim'Act
}
