% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChocSolvabilite2_do_choc_currency.R
\docType{methods}
\name{do_choc_currency}
\alias{do_choc_currency}
\title{Permet a partir d'un canton initial de creer un canton choque de device.}
\usage{
do_choc_currency(x, nom_currency, direction, canton)
}
\arguments{
\item{x}{objet de la classe \code{\link{ChocSolvabilite2}}.}

\item{nom_currency}{object \code{character} correspondant au nom de la devise du choc.}

\item{direction}{object \code{character} prenant pour valeur "up" pour le choc a la hausse et "down" pour le choc a la baisse.}

\item{canton}{est un objet de la classe \code{\link{Canton}}. Il correspond au canton non choque (i.e. central)
de l'assureur.}
}
\value{
\code{canton} l'objet  de la classe \code{\link{Canton}} correspondant au scenario choque
 currency au sens de la formule standard Solvabilite 2.
Cette parametrisation est effectuee dans les fichiers d'inputs utilisateurs.
}
\description{
\code{do_choc_currency} est une methode permettant d'appliquer le choc currency de la formule standard Solvabilite 2
 a un canton. Cette methode s'applique uniquement aux obligations, actions et titres en immobilier.
}
\author{
Prim'Act
}
