% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChocSolvabilite2_do_choc_frais.R
\docType{methods}
\name{do_choc_frais}
\alias{do_choc_frais}
\title{Permet a partir d'un canton initial de creer un canton choque frais.}
\usage{
do_choc_frais(x, canton, autres_passifs_choc)
}
\arguments{
\item{x}{objet de la classe \code{\link{ChocSolvabilite2}}.}

\item{canton}{est un objet de la classe \code{\link{Canton}}. Il correspond au canton non choque (i.e. central)
de l'assureur.}

\item{autres_passifs_choc}{est un objet de la classe \code{\link{AutresPassifs}}, il correspond au chargement
 des autres passifs choques.
Ces derniers ont ete renseignes par l'utilisateur en donnees.}
}
\value{
\code{canton} l'objet  de la classe \code{\link{Canton}} correspondant au scenario choque frais
 au sens de la formule standard Solvabilite 2.
}
\description{
\code{do_choc_frais} est une methode permettant d'appliquer le choc frais de la formule standard Solvabilite 2
 a un canton.
}
\note{
La parametrisation des chocs de frais est effectuee dans les fichiers d'inputs utilisateurs.
}
\author{
Prim'Act
}

