% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChocSolvabilite2_do_choc_longevite.R
\docType{methods}
\name{do_choc_longevite}
\alias{do_choc_longevite}
\title{Permet a partir d'un canton initial de creer un canton choque longevite.}
\usage{
do_choc_longevite(x, canton, autres_passifs_choc)
}
\arguments{
\item{x}{objet de la classe \code{\link{ChocSolvabilite2}}.}

\item{canton}{est un objet de la classe \code{\link{Canton}}. Il correspond au canton non choque (i.e. central) de l'assureur.}

\item{autres_passifs_choc}{est un objet de la classe \code{\link{AutresPassifs}}, il correspond au chargement des autres passifs choques en longevite.
Ces derniers ont ete renseignes par l'utilisateur en donnees.}
}
\value{
\code{canton} l'objet  de la classe \code{\link{Canton}} correspondant au scenario choque longevite
au sens de la formule standard Solvabilite 2.
}
\description{
\code{do_choc_longevite} est une methode permettant d'appliquer le choc longevite de la formule standard
 Solvabilite 2 a un canton.
}
\note{
La parametrisation des chocs de longevite est effectuee dans les fichiers d'inputs utilisateurs.
}
\author{
Prim'Act
}

