% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortFin_do_update_pmvl.R
\docType{methods}
\name{do_update_pmvl}
\alias{PortFin}
\alias{do_update_pmvl}
\title{Met a jour l'ensemble des attributs pvl et pml d'un objet PortFin}
\usage{
do_update_pmvl(x)
}
\arguments{
\item{x}{est un objet de la classe \code{PortFin},}
}
\value{
L'objet \code{x} de la classe \code{PortFin} dont les plus values et moins values ont ete recalculees avec les elements du \code{PortFin} renseigne en input.
}
\description{
\code{do_update_pmvl} est une methode permettant de calculer le taux de rendement financier du portefeuille.
}
\author{
Prim'Act
}

