% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_do_update_val_courante.R
\docType{methods}
\name{do_update_PRE_val_courante}
\alias{PRE}
\alias{do_update_PRE_val_courante}
\title{Mise a jour de la valeur courante de PRE.}
\usage{
do_update_PRE_val_courante(x, val_courante)
}
\arguments{
\item{x}{objet de la classe \code{PRE}correspondant a la PRE avant mise a jour.}

\item{val_courante}{est un \code{numeric} correspondant au montant de PRE calcule par la fonction \code{\link{calc_PRE}}.}
}
\value{
L'objet \code{PRE} mis a jour de la nouvelle valeur courante de \code{PRE}
}
\description{
\code{do_update_PRE_val_courante} est une methode permettant de calculer le montant de PRE.
}
\author{
Prim'Act
}
\seealso{
La methode de calcul de la PRE \code{\link{calc_PRE}}
}

