% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC_do_update_val_debut.R
\docType{methods}
\name{do_update_RC_val_debut}
\alias{do_update_RC_val_debut}
\title{Mise a jour de la valeur initiale de RC}
\usage{
do_update_RC_val_debut(x, val_debut)
}
\arguments{
\item{x}{objet de la classe \code{\link{RC}} correspondant a la RC avant mise a jour.}

\item{val_debut}{est un \code{numeric} correspondant au montant de debut de periode de RC.}
}
\value{
L'objet \code{x} mis a jour de la nouvelle valeur de debut de \code{\link{RC}}.
}
\description{
\code{do_update_RC_val_debut} est une methode permettant de mettre a jour le montant de debut de periode de RC.
}
\author{
Prim'Act
}
\seealso{
La methode de calcul de la RC \code{\link{calc_RC}}
}

