% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{echeancier}
\alias{echeancier}
\title{Calcule les flux obligataires.}
\arguments{
\item{coupon}{un vecteur contenant les taux de coupons de chaque obligation.}

\item{maturite}{un vecteur d'entiers contenant les maturites residuelles de chaque obligation.}

\item{zspread}{un vecteur contenant les zero-spreadsde chaque obligation.}

\item{nominal}{un vecteur contenant les valeurs nominales de chaque obligation.}

\item{yield}{un vecteur contenant la courbe de taux consideree (peut-etre vide).}
}
\value{
Une matrice contenant :
\describe{
\item{\code{grid_flux} : }{la matrice d'ecoulement des flux. Cette matrice a autant de colonnes
que le max du vecteur de maturite residuelle, et autant de lignes que les vecteurs d'input
\code{coupon, maturite, zspread, nominal}.
Chaque ligne decrit les flux annuels a venir pour l'actif obligataire dont les caracteristiques sont
renseignees en input.}
}
}
\description{
\code{echeancier} est une methode permettant de calculer les flux jusqu'a maturite residuelle d'un model point ou
d'un ensemble de model points obligataires.
}
\author{
Prim'Act
}
