% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpEuroInd-class.R
\docType{class}
\name{EpEuroInd}
\alias{EpEuroInd}
\title{La classe EpEuroInd}
\description{
Classe pour les passifs epargne en euros
}
\section{Slots}{

\describe{
\item{\code{mp}}{un objet data.frame au format fige contenant l'ensemble de model points epargne en euros}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

