% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESG_class.R
\docType{class}
\name{ESG}
\alias{ESG}
\title{La classe ESG}
\description{
Classe pour les donnees globales d'actif
}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

