% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevaloEngine_finance_cible_marge.R
\docType{methods}
\name{finance_cible_marge}
\alias{RevaloEngine}
\alias{finance_cible_marge}
\title{Evalue le financement d'une revalorisation au taux cible par la marge de l'assureur}
\usage{
finance_cible_marge(marge_fin, bes_cible, rev_stock_nette, marge_min)
}
\arguments{
\item{marge_fin}{une valeur \code{numeric} donnant le montant courant de la marge financiere de l'assureur.}

\item{bes_cible}{un vecteur \code{numeric} correspondant au besoin de financement necessaire pour atteindre
le taux cible part produit.}

\item{rev_stock_nette}{un vecteur \code{numeric} comprenant par produit
le montant de revalorisation nette au titre de le PB atteint.}

\item{marge_min}{est une valeur \code{numeric} correspondant
au montant minimum de marge financiere souhaite par l'assureur.}
}
\value{
\code{rev_stock_nette} la valeur de la revalorisation nette servie par produit apres reduction de marge.

\code{marge_fin} le montant de marge de l'assureur apres reduction.
}
\description{
\code{finance_cible_marge} est une methode permettant de
 determiner le financement d'une revalorisation au taux cible en comprimant la marge financiere
 de l'assureur
}
\author{
Prim'Act
}

