% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevaloEngine_finance_cible_pmvl.R
\docType{methods}
\name{finance_cible_pmvl}
\alias{RevaloEngine}
\alias{finance_cible_pmvl}
\title{Evalue le financement d'un taux cible par des cessions de plus-values latentes.}
\usage{
finance_cible_pmvl(bes_cible, rev_stock_nette, base_fin, seuil_pmvl, tx_pb)
}
\arguments{
\item{bes_cible}{est un vecteur de type \code{numeric} comprenant par produit le besoin de financement aux taux cible.}

\item{rev_stock_nette}{est un vecteur de type \code{numeric} comprenant par produit le montant de revalorisation nette atteint.}

\item{base_fin}{est un vecteur de type \code{numeric} comprenant par produit la base de produits financiers.}

\item{seuil_pmvl}{est une valeur \code{numeric} comprenant au montant de plus-values latentes qui peut etre liquidee. Ce montant
doit etre exprime en tenant compte de l'abattement realise pour rapport les plus-values latentes de l'actiff general au passif.}

\item{tx_pb}{est un vecteur de type \code{numeric} comprenant par produit les taux de particippation aux benefices contractuels.}
}
\value{
Une liste avec la valeur de la revalorisation nette servie par produit et le montant de plus-values a liquider, ramene a
la valeur du passif pour financer la revalorisation.
}
\description{
\code{finance_cible_pmvl} est une methode permettant de
 determiner le financement de taux cibles par des cessions de plus-values latentes.
}
\details{
Lorsque la revalorisation nette est superieure au besoin de financement des taux cibles, on sert le taux cible
et on partage le surplus. A l'inverse, les taux cible sont finances par
les compensations entre produits lorsque certains prevoient une revalorisation superieure au taux cible, et
par une liquidation de plus-values latentes.
}
\author{
Prim'Act
}

