% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevaloEngine_finance_cible_ppb.R
\docType{methods}
\name{finance_cible_ppb}
\alias{RevaloEngine}
\alias{finance_cible_ppb}
\title{Evalue le financement d'une revalorisation au taux cible par une reprise de PPB.}
\usage{
finance_cible_ppb(bes_cible, rev_stock_nette, ppb)
}
\arguments{
\item{bes_cible}{un vecteur \code{numeric} correspondant au besoin de financement necessaire pour atteindre
le taux cible par produit.}

\item{rev_stock_nette}{un vecteur \code{numeric} comprenant par produit
le montant de revalorisation nette au titre de le PB atteint.}

\item{ppb}{un objet de la classe \code{\link{Ppb}} qui renvoie l'etat courant de la PPB.}
}
\value{
\code{rev_stock_nette} la valeur de la revalorisation nette servie apres une eventuelle reprise de PPB.

\code{dotation} le montant de dotation a la PPB.

\code{reprise} le montant de reprise sur la PPB.

\code{ppb} l'objet \code{ppb} mis a jour.
}
\description{
\code{finance_cible_ppb} est une methode permettant de
 determiner le financement d'une revalorisation au taux cible par la reprise de provision pour participation
  aux benefices (PPB). Cette methode evalue egalement si une dotation est effectue.
}
\details{
Lorsque la revalorisation nette est superieure au besoin de financement des taux cibles,
on sert le taux cible et on dote le reste a la PPB dans la limite du plafond de dotation annuel.
A l'inverse, les taux cible sont finances par les compensations entre produits lorsque certains
prevoient une revalorisation superieure au taux cible, puis
par une reprise sur PPB.
}
\author{
Prim'Act
}

