% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevaloEngine_finance_contrainte_legale.R
\docType{methods}
\name{finance_contrainte_legale}
\alias{RevaloEngine}
\alias{finance_contrainte_legale}
\title{Applique la contrainte legale de participation aux benefices}
\usage{
finance_contrainte_legale(base_fin, base_fin_etendu, result_tech, it_stock,
  rev_stock_nette, rev_prest_nette, dot_ppb, marge_fin, ppb, param_revalo)
}
\arguments{
\item{base_fin}{est un vecteur de type \code{numeric} comprenant par produit la base de produits financiers.}

\item{result_tech}{est une valeur de type \code{numeric} comprenant le resultat technique.}

\item{it_stock}{est un vecteur de type \code{numeric} comprenant par produit les interets techniques affectes au stock.}

\item{rev_stock_nette}{est un vecteur de type \code{numeric} comprenant par produit la revalorisaton nette
affectee au stock.}

\item{rev_prest_nette}{est un vecteur de type \code{numeric} comprenant par produit a revalorisaton nette
affectee aux prestations.}

\item{dot_ppb}{est une valeur de type \code{numeric} comprenant la dotation de PPB financant la revalorisation
sur stock, hors TMG.}

\item{marge_fin}{est une valeur de type \code{numeric} comprenant la marge financiere courante de l'assureur.}

\item{ppb}{est un objet de la classe \code{Ppb} qui renvoie l'etat courant de la PPB.}

\item{param_revalo}{est un objet de la classe \code{ParamRevaloEngine}
comprenant les parametres de revalorisation.}

\item{base_fin_entendu}{est une valeur \code{numeric} comprenant la base totale de produits financiers,
y compris passifs non modelises.}
}
\value{
Une liste avec la revalorisation nette affectee au stock
la marge financiere de l'assureur apres prise en compte de la contrainte legale, et la PPB.
}
\description{
\code{finance_contrainte_legale} est une methode permettant de calculer la contrainte legale
de participation aux benefices et de l'appliquer si necessaire pour accroitre la revalorisation.
}
\author{
Prim'Act
}

