% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevaloEngine_finance_tmg.R
\docType{methods}
\name{finance_tmg}
\alias{RevaloEngine}
\alias{finance_tmg}
\title{Calcule la contribution de la PPB au financement des taux minimum garanti.}
\usage{
finance_tmg(bes_tmg_prest, bes_tmg_stock, ppb)
}
\arguments{
\item{bes_tmg_prest}{est un vecteur de type \code{numeric} comprenant
par produit le besoin de finance des TMG sur prestations.
@param bes_tmg_stock est un vecteur de type \code{numeric} comprenant
par produit le besoin de finance des TMG sur le stock.}

\item{ppb}{est un objet de la classe \code{\link{Ppb}} qui renvoie l'etat courant de la PPB.}
}
\value{
Une liste comprenant la valeur de la contribution au financement des TMG sur prestations et sur stock.
Cette liste comprend egalement la PPB mise a jour.
}
\description{
\code{finance_tmg} est une methode permettant d'evaluer le contribution de la PPB
au financement des taux minimum garanti sur prestations et sur stock.
}
\details{
Cette fonction priorise le financement des TMG prestations. Elle met a jour la PPB des reprises effectuees.
}
\author{
Prim'Act
}

