% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraisFin_load.R
\docType{methods}
\name{frais_fin_load}
\alias{frais_fin_load}
\title{Methode permettant de charger la valeur initiale des frais financiers dans un objet de type FraisFin.}
\usage{
frais_fin_load(file_frais_fin_address)
}
\arguments{
\item{file_frais_fin_address}{est un \code{character} contenant l'adresse exacte du fichier d'input utilisateur permettant de renseigner les Frais
financier.}
}
\value{
L'objet de la classe \code{FraisFin} construit a partir des inputs renseignes par l'utilisateur.
}
\description{
\code{frais_fin_load} est une methode permettant de charger les frais financiers.
}
\seealso{
La classe \code{\link{Initialisation}} et sa methode \code{\link{set_architecture}} pour renseigner l'input.
}
\author{
Prim'Act
}
