% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraisPassif-load.R
\docType{methods}
\name{frais_passif_load}
\alias{frais_passif_load}
\title{Methode permettant de charger la valeur des frais de passif.}
\usage{
frais_passif_load(file_frais_passif_address)
}
\arguments{
\item{file_frais_passif_address}{est un \code{character} contenant l'adresse exacte du fichier d'input utilisateur
permettant de renseigner un objet \code{\link{FraisPassif}}.}
}
\value{
L'objet de la classe \code{\link{FraisPassif}} construit a partir des inputs renseignes par l'utilisateur.
}
\description{
\code{frais_passif_load} est une methode permettant de charger les donnees associees a un
objet de classe \code{\link{FraisPassif}}.
}
\seealso{
La classe \code{\link{Initialisation}} et sa methode \code{\link{set_architecture}}
 pour renseigner l’input.
}
\author{
Prim'Act
}
