% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraisFin_class.R
\docType{class}
\name{FraisFin}
\alias{FraisFin}
\title{La classe FraisFin}
\description{
Classe pour les parametres de frais financiers d'un assureur.
}
\section{Slots}{

\describe{
\item{\code{tx_chargement}}{est une valeur \code{numeric} correspondant au taux de frais de gestion financiere.}

\item{\code{indicatrice_inflation}}{est un objet de type \code{logical}, qui permet d'indiquer si une inflation
doit etre applique.}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

