% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraisPassif-class.R
\docType{class}
\name{FraisPassif}
\alias{FraisPassif}
\title{La classe FraisPassif}
\description{
Classe pour pour les frais de passif
}
\section{Slots}{

\describe{
\item{\code{mp}}{data frame contenant les paramètres des taux de passif par produit}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

