% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESG-get_choc_inflation_frais.R
\docType{methods}
\name{get_choc_inflation_frais}
\alias{get_choc_inflation_frais}
\title{Applique le choc frais de la formule standard a la table de simulation pour l'indice inflation.}
\usage{
get_choc_inflation_frais(x, choc)
}
\arguments{
\item{x}{un objet de la classe \code{\link{ESG}}.}

\item{choc}{une valeur \code{numeric} correspondant au coefficient de
choc a appliquer en additif au taux d'inflation.}
}
\value{
L'objet \code{x} mis a jour.
}
\description{
\code{get_choc_inflation_frais} est une methode permettant d'appliquer le choc frais
 de la formule standard a la table de simulation pour l'indice inflation.
}
\note{
L'inflation comprise dans l'ESG est suppose etre deja capitalise, i.e.
\eqn{indice_inflation = (1 + tx inflation)^{annee}}. Il ne s'agit pas du taux d'inflation.
}
\author{
Prim'Act
}

