% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypTech-get_choc_table.R
\docType{methods}
\name{get_choc_table}
\alias{HypTech}
\alias{get_choc_table}
\title{donne le résulat d'un choc sur toutes les tables d'un objet HypTech}
\usage{
get_choc_table(x, choc)
}
\description{
\code{get_choc_table}
}
\author{
Prim'Act
}

