% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetraiteEuroRest-get_coef_rente.R
\docType{methods}
\name{get_coef_rente}
\alias{get_coef_rente}
\title{Recupere le coefficient d'ax ou d'axy}
\usage{
get_coef_rente(x, ht)
}
\arguments{
\item{x}{un objet de la classe \code{\link{RetraiteEuroRest}}.}

\item{ht}{un objet de la classe \code{\link{HypTech}}.}
}
\value{
Le coefficient \code{numeric} d'ax ou axy selon le statut du model point dont les caracteristiques sont precisees en input.
}
\description{
\code{get_coef_rente} est une methode permettant d'obtenir le coefficient de survie esperee actualise au taux technique de rente.
}
\author{
Prim'Act
}
