% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypTech-get_comport.R
\docType{methods}
\name{get_comport}
\alias{HypTech}
\alias{get_comport}
\title{Calcule le taux cible.}
\usage{
get_comport(x, nom_table, list_rd, tx_cible_prec)
}
\arguments{
\item{x}{un objet de la classe \code{HypTech}.}

\item{nom_table}{un nom de la table de parametres de taux cible.}

\item{list_rd}{une liste de taux de rendement.}

\item{tx_cible_prec}{un taux cible en numerique.}
}
\value{
La valeur du taux cible.
}
\description{
\code{get_comport} est une methode permettant de calculer le taux cible.
}
\author{
Prim'Act
}

