% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypTech-get_proba_paye_retraite.R
\docType{methods}
\name{get_proba_paye_retraite}
\alias{get_proba_paye_retraite}
\title{Recupere la probabilite de payer un flux de rente pour un produit de retraite.}
\usage{
get_proba_paye_retraite(x, nom_table_1, age_1, gen_1, statut, tx_rvs,
  nom_table_2, age_2, gen_2)
}
\arguments{
\item{x}{un objet de la classe \code{\link{HypTech}}.}

\item{nom_table_1}{un \code{character} designant le nom de la table de mortalite de la tete principale.}

\item{age_1}{une valeur \code{integer} correspondant a l'age de la tete principale.}

\item{gen_1}{une valeur \code{integer} correspondant a la generation de la tete principale.}

\item{statut}{l'indicatrice \code{integer} permettant de determiner le statut de la rente
(1 : une tete, 2 : une tete principale et un reversataire, 3 : seul le reversataire est en vie).}

\item{tx_rvs}{une valur \code{numeric} correspondant au taux de reversion.}

\item{nom_table_2}{un \code{character} designant le nom de la table de mortalite de la tete secondaire.}

\item{age_2}{une valeur \code{integer} correspondant a l'age de la tete secondaire (renseigner la meme valeur que pour la tete principale en l'absence de tete secondaire).}

\item{gen_2}{une valeur \code{integer} correspondant a la generation de la tete secondaire (renseigner la meme valeur que pour la tete principale en l'absence de tete secondaire).}
}
\value{
La probabilite de payer un flux de rente, ponderee du taux de reversion si le staut est 2 ou 3.
}
\description{
\code{get_proba_paye_retraite} est une methode qui calcule un flux de rente pour un produit de retraite
en tenant compte de la possible reversion.
}
\seealso{
Recuperer le taux de deces \code{\link{get_qx_mort}}.
}
\author{
Prim'Act
}
