% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypTech-get_proba_sortie_retraite.R
\docType{methods}
\name{get_proba_sortie_retraite}
\alias{get_proba_sortie_retraite}
\title{Recupere la probabilite de sortie d'un contrat retraite.}
\usage{
get_proba_sortie_retraite(x, nom_table_1, age_1, gen_1, statut, nom_table_2,
  age_2, gen_2)
}
\arguments{
\item{x}{un objet de la classe \code{\link{HypTech}}.}

\item{nom_table_1}{un \code{character} designant le nom de la table de mortalite de la tete principale.}

\item{age_1}{la valeur \code{integer} de l'age de la tete principale.}

\item{gen_1}{la valeur \code{integer} de la generation de la tete principale.}

\item{statut}{l'indicatrice \code{integer} permettant de determiner le statut de la rente
(1 : une tete, 2 : une tete principale et un reversataire, 3 : seul le reversataire est en vie).}

\item{nom_table_2}{un \code{character} designant le nom de la table de mortalite de la tete secondaire.}

\item{age_2}{la valeur \code{integer} de l'age de la tete secondaire.}

\item{gen_2}{la valeur \code{integer} de la generation de la tete secondaire.}
}
\value{
La probabilite de fin de contrat calculee.
}
\description{
\code{get_proba_sortie_retraite} est une methode permettant de calculer la probabilite de sortie
pour des contrats avec ou sans reversion.
}
\seealso{
Recuperer le taux de deces \code{\link{get_qx_mort}}.
}
\author{
Prim'Act
}
