% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypTech-get_qx_mort.R
\docType{methods}
\name{get_qx_mort}
\alias{HypTech}
\alias{get_qx_mort}
\title{Recuperer les taux de deces calcules.}
\usage{
get_qx_mort(x, nom_table, age, gen)
}
\arguments{
\item{x}{un objet de la classe \code{\link{HypTech}}.}

\item{nom_table}{un nom de la table de mortalite.}

\item{age}{est la valeur \code{numeric} de l'age.}

\item{gen}{est la valeur \code{numeric} de la generation.}
}
\value{
Le taux de deces.
}
\description{
\code{get_qx_mort} est une methode permettant d'executer le calcul des taux de deces.
}
\author{
Prim'Act
}
\seealso{
Le calcul du taux de deces \code{\link{calc_qx}}.
}

