% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypTech-get_rach_dyn.R
\docType{methods}
\name{get_rach_dyn}
\alias{HypTech}
\alias{get_rach_dyn}
\title{Calcule la composante rachats dynamique}
\usage{
get_rach_dyn(x, nom_table, tx_cible, tx_serv)
}
\arguments{
\item{x}{un objet de la classe \code{HypTech}.}

\item{nom_table}{: nom de la table de parametres de rachats conjoncturels consideree}

\item{tx_cible}{: taux cible en numerique}

\item{tx_serv}{: taux servi en numerique}
}
\value{
la composante rachats dynamique en numerique
}
\description{
\code{get_rach_dyn} est une methode permettant de calculer la composante rachats dynamique.
}
\author{
Prim'Act
}

