% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypCanton_load.R
\docType{methods}
\name{hyp_canton_load}
\alias{hyp_canton_load}
\title{Methode permettant de charger la valeur initiale des hypotheses du canton.}
\usage{
hyp_canton_load(file_hyp_canton_address)
}
\arguments{
\item{file_hyp_canton_address}{est un \code{character} contenant l'adresse exacte
du fichier d'input utilisateur
permettant de renseigner un objet \code{\link{HypCanton}}.}
}
\value{
L'objet de la classe \code{\link{HypCanton}} construit a partir des inputs renseignes par l'utilisateur.
}
\description{
\code{hyp_canton_load} est une methode permettant de charger les parametres associees a un
objet de classe \code{\link{HypCanton}}.
}
\author{
Prim'Act
}
\seealso{
La classe \code{\link{Initialisation}} et sa methode \code{\link{set_architecture}}
 pour renseigner l’input.
}

