% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypCanton_class.R
\docType{class}
\name{HypCanton}
\alias{HypCanton}
\title{La classe \code{HypCanton}.}
\description{
Une class de parametres pour les parametres generaux du canton.
}
\section{Slots}{

\describe{
\item{\code{tx_soc}}{une valeur \code{numeric} correspondant au taux de prelevements social.}

\item{\code{tx_import}}{une valeur \code{numeric} correspondant au taux d'impot sur le resultat.}

\item{\code{method_taux_cible}}{un \code{character} correspond au nom de la methode de calcul du taux cible.}
}}
\note{
Dans la version courante, la valeur de \code{method_taux_cible} doit etre parametree a "Meth1".
}
\author{
Prim'Act
}
\keyword{classes}

