% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immo_class.R
\docType{class}
\name{Immo}
\alias{Immo}
\title{La classe Immo}
\description{
Classe pour les actifs de type Immo
}
\section{Slots}{

\describe{
\item{\code{ptf_immo}}{est un dataframe, chaque ligne represente un actif immobilier du portefeuille d'Immobilier.}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

