% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialisation_create_folder.R
\docType{methods}
\name{init_create_folder}
\alias{init_create_folder}
\title{Creation de l'architecture de sauvegarde des scenarios et executions du code a partir de la racine renseignee.}
\usage{
init_create_folder(x)
}
\arguments{
\item{x}{objet de la classe \code{\link{Initialisation}}.}
}
\value{
En cas de bonne execution (i.e. l'ensemble des dossiers est cree ou ecrase) la methode renvoie un \code{logical}.
}
\description{
\code{init_create_folder} est une methode permettant de creer l'architecture de sauvegarde des scenarios et les executions du code a partir de la racine renseignee.
}
\note{
Il est necessaire anterieurement a l'appel de cette fonction d'avoir dans un premier temps
 cree un objet \code{\link{Initialisation}} en lui ayant affecte une racine,
puis dans un second temps d'avoir appele la methode \code{\link{set_architecture}} a ce meme objet.
}
\author{
Prim'Act
}

