% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutresReserves-init_debut_pgg_psap.R
\docType{methods}
\name{init_debut_pgg_psap}
\alias{init_debut_pgg_psap}
\title{Re-initialise un objet \code{AutresReserves} en debut d'annee.}
\usage{
init_debut_pgg_psap(x)
}
\arguments{
\item{x}{objet de la classe \code{AutresReserves}.}
}
\value{
L'objet x reinitialise.
}
\description{
\code{init_debut_pgg_psap} est une methode permettant de re-initialiser les montants
de PGG et de PSAP de debut de periode.
}
\author{
Prim'Act
}

