% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialisation_init_scenario.R
\docType{methods}
\name{init_scenario}
\alias{Initialisation}
\alias{init_scenario}
\title{Initialisation des scenarios de chocs d'un workspace.}
\usage{
init_scenario(x, alloc_cible, seuil_realisation_PVL, nb_simu, nb_annee_proj)
}
\arguments{
\item{x}{un objet de la classe \code{Initialisation}.}
}
\value{
Pas de sortie.
}
\description{
\code{init_scenario} est la methode d'initialisation.
}
\author{
Prim'Act
}

