% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialisation_initSimBEL.R
\docType{methods}
\name{init_SimBEL}
\alias{init_SimBEL}
\title{Initialisation d'un workspace.}
\usage{
init_SimBEL(x)
}
\arguments{
\item{x}{un objet de la classe \code{\link{Initialisation}}.}
}
\value{
Pas de sortie.
}
\description{
\code{init_SimBEL} est la methode d'initialisation d'un workspace.
}
\note{
Cette methode permet de creer l'objet \code{\link{Canton}} initial et de le sauvegarder dans le repertoire adequat de l'architecture.
}
\author{
Prim'Act
}
