% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBase_init_tables.R
\docType{methods}
\name{init_tables}
\alias{init_tables}
\title{Intialise les tables de la base de donnees. Cette fonction initialise les differentes tables de la base.
Chacune des tables possede 4 colonnes : num_sim, num_an, prod et flux.}
\usage{
init_tables(x)
}
\arguments{
\item{x}{est un objet de type \code{\link{DataBase}}.}
}
\description{
Intialise les tables de la base de donnees. Cette fonction initialise les differentes tables de la base.
Chacune des tables possede 4 colonnes : num_sim, num_an, prod et flux.
}
\author{
Prim'Act
}
