% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialisation_class.R
\docType{class}
\name{Initialisation}
\alias{Initialisation}
\title{La classe Initialisation}
\description{
Classe d'initialisation
}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

