% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBase_insert_tables.R
\docType{methods}
\name{insert_tables}
\alias{insert_tables}
\title{Fusionne et insere les donnees dans la \code{\link{DataBase}}}
\usage{
insert_tables(x, result_simu, ens_simu)
}
\arguments{
\item{x}{est un objet de type \code{\link{DataBase}}.}

\item{result_simu}{une \code{liste} contenant les tables a inserer dans la \code{DataBase}.}

\item{ens_simu}{un vecteur de type \code{numeric} contenant les indices des elements de la liste a inserer dans la \code{DataBase}.}
}
\description{
Fusionne et insere les donnees dans la \code{\link{DataBase}}
}
\author{
Prim'Act
}
