% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action_load.R
\docType{methods}
\name{load_action}
\alias{load_action}
\title{Methode permettant de charger le portefeuille action initital dans un objet de type \code{\link{Action}}.}
\usage{
load_action(file_action_address)
}
\arguments{
\item{file_action_address}{est un \code{character} contenant l'adresse exacte du fichier d'input utilisateur renseignant le portefeuille.}
}
\value{
L'objet de la classe \code{\link{Action}} construit a partir des inputs renseignes par l'utilisateur.
}
\description{
\code{load_action} est une methode permettant de charger le portefeuille action.
}
\seealso{
La classe \code{\link{PortFin}} et sa methode \code{\link{chargement_PortFin}}.
}
\author{
Prim'Act
}
