% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialisation_load_be.R
\docType{methods}
\name{load_be}
\alias{load_be}
\title{Recuperation de l'objet \code{\link{Be}} et affectation d'une base de donnees.}
\usage{
load_be(x, choc_name)
}
\arguments{
\item{x}{un objet de la classe \code{\link{Initialisation}}.}

\item{choc_name}{un \code{character} indiquant le nom du choc.}
}
\value{
Pas de sortie.
}
\description{
\code{load_be} est la methode permettant de recuper un objet \code{\link{Be}}.
}
\note{
Cette methode permet de creer l'objet \code{\link{Canton}} initial et de le sauvegarder dans le repertoire adequat de l'architecture.
}
\author{
Prim'Act
}
