% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpEuroInd_load.R
\docType{methods}
\name{load_epeuroind}
\alias{load_epeuroind}
\title{Methode permettant de charger un portefeuille epargne initital dans un objet de type \code{\link{EpEuroInd}}.}
\usage{
load_epeuroind(file_epeuroind_address)
}
\arguments{
\item{file_epeuroind_address}{est un \code{character} contenant l'adresse exacte du fichier d'input utilisateur permettant de renseigner le portefeuille.}
}
\value{
L'objet de la classe \code{\link{Action}} construit a partir des inputs renseignes par l'utilisateur.
}
\description{
\code{load_epeuroind} est une methode permettant de charger un portefeuille epargne.
}
\seealso{
La classe \code{\link{PortPassif}} et sa methode \code{\link{load_pp}}.
}
\author{
Prim'Act
}
