% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypTech-load.R
\docType{methods}
\name{load_ht}
\alias{HypTech}
\alias{load_ht}
\title{permet de charger et d'instancier la classe HypTech}
\usage{
load_ht(x)
}
\description{
\code{load_ht} est une methode permettant charger la class HypTech
}
\author{
Prim'Act
}

