% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immo_load.R
\docType{methods}
\name{load_immo}
\alias{load_immo}
\title{Methode permettant de charger le portefeuille action initital dans un objet de type \code{\link{Immo}}.}
\usage{
load_immo(file_immo_address)
}
\arguments{
\item{file_immo_address}{est un \code{character} contenant l'adresse exacte du fichier d'input utilisateur permettant de renseigner le portefeuille.}
}
\value{
L'objet de la classe \code{\link{Immo}} construit a partir des inputs renseignes par l'utilisateur.
}
\description{
\code{load_immo} est une methode permettant de charger le portefeuille action.
}
\seealso{
La classe \code{\link{PortFin}} et sa methode \code{\link{chargement_PortFin}}.
}
\author{
Prim'Act
}
