% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_load.R
\docType{methods}
\name{load_oblig}
\alias{load_oblig}
\title{Methode permettant de charger le portefeuille oblig initital dans un objet de type \code{\link{Oblig}}.}
\usage{
load_oblig(file_oblig_address)
}
\arguments{
\item{file_oblig_address}{est un \code{character} contenant l'adresse exacte du fichier d'input utilisateur permettant de renseigner le portefeuille.}
}
\value{
L'objet de la classe \code{\link{Oblig}} construit a partir des inputs renseignes par l'utilisateur.
}
\description{
\code{load_oblig} est une methode permettant de charger le portefeuille oblig
}
\seealso{
La classe \code{\link{PortFin}} et sa methode \code{\link{chargement_PortFin}}.
}
\author{
Prim'Act
}
