% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortPassif-load.R
\docType{methods}
\name{load_pp}
\alias{PortPassif}
\alias{load_pp}
\title{permet de charger et d'instancier la classe PortPassif}
\usage{
load_pp(x)
}
\description{
\code{load_pp} est une methode permettant charger la class PortPassif
}
\author{
Prim'Act
}

