% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetraiteEuroRest_load.R
\docType{methods}
\name{load_reteurorest}
\alias{load_reteurorest}
\title{Methode permettant de charger un portefeuille retraite initital dans un objet de type \code{\link{RetraiteEuroRest}}.}
\usage{
load_reteurorest(file_reteurorest_address)
}
\arguments{
\item{file_reteurorest_address}{est un \code{character} contenant l'adresse exacte du fichier d'input utilisateur permettant de renseigner
le portefeuille.}
}
\value{
L'objet de la classe \code{\link{Action}} construit a partir des inputs renseignes par l'utilisateur.
}
\description{
\code{load_reteurorest} est une methode permettant de charger un portefeuille retraite en phase de restitution.
}
\seealso{
La classe \code{\link{PortPassif}} et sa methode \code{\link{load_pp}}.
}
\author{
Prim'Act
}
