% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Treso_load.R
\docType{methods}
\name{load_treso}
\alias{load_treso}
\title{Methode permettant de charger le portefeuille treso initital dans un objet de type \code{\link{Treso}}.}
\usage{
load_treso(file_treso_address)
}
\arguments{
\item{file_treso_address}{est un \code{character} contenant l'adresse exacte du fichier d'input utilisateur permettant de renseigner le
portefeuille.}
}
\value{
L'objet de la classe \code{\link{Treso}} construit a partir des inputs renseignes par l'utilisateur.
}
\description{
\code{load_treso} est une methode permettant de charger le portefeuille treso
}
\seealso{
La classe \code{\link{PortFin}} et sa methode \code{\link{chargement_PortFin}}.
}
\author{
Prim'Act
}
