% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBase_merge_actifs.R
\docType{methods}
\name{merge_actifs}
\alias{merge_actifs}
\title{Aggregation des donnees relatives aux actifs.}
\usage{
merge_actifs(action, immo, oblig, treso, sim)
}
\arguments{
\item{action}{un \code{data.frame} contenant les donnees des actions a inserer dans la base..}

\item{immo}{un \code{data.frame} contenant les donnees des immo a inserer dans la base.}

\item{oblig}{un \code{data.frame} contenant les donnees des obligations a inserer dans la base.}

\item{treso}{un \code{data.frame} contenant les donnees treso a inserer dans la base.}

\item{sim}{un \code{integer} representant le numero de la simulation.}
}
\description{
Aggregation des donnees relatives aux actifs.
}
\author{
Prim'Act
}
