% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBase_merge_be.R
\docType{methods}
\name{merge_be}
\alias{merge_be}
\title{Insertion des BE dans la table \code{BE} de la base de donnees.}
\usage{
merge_be(be, sim, nom_produit)
}
\arguments{
\item{be}{une \code{matrix} a inserer dans la base.}

\item{sim}{un \code{integer} representant le numero de la simulation.}

\item{nom_produit}{une liste de \code{character} contenant les noms des differents produits.}
}
\description{
Insertion des BE dans la table \code{BE} de la base de donnees.
}
\author{
Prim'Act
}
