% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBase_merge_flux_fin.R
\docType{methods}
\name{merge_flux_fin}
\alias{merge_flux_fin}
\title{Aggregation des donnees relatives aux flux financiers.}
\usage{
merge_flux_fin(flux_fin, sim)
}
\arguments{
\item{flux_fin}{est un objet de type \code{matrix}.}

\item{sim}{un \code{integer} correspondant au numero de simulation.}
}
\description{
Aggregation des donnees relatives aux flux financiers.
}
\author{
Prim'Act
}
