% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBase_merge_pb.R
\docType{methods}
\name{merge_pb}
\alias{merge_pb}
\title{Aggregation des donnees relatives a la PB.}
\usage{
merge_pb(output_pb, sim)
}
\arguments{
\item{output_pb}{est un objet de type \code{data.frame}.}

\item{sim}{un \code{integer} correspondant au numero de simulation.}
}
\description{
Aggregation des donnees relatives a la PB.
}
\author{
Prim'Act
}
