% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBase_merge_table_be.R
\docType{methods}
\name{merge_table_be}
\alias{merge_table_be}
\title{Jointure et insertion des donnees dans la base de donnees. Les donnees sont les ouputs (\code{output_be}) de la fonction \code{\link{proj_an}}}
\usage{
merge_table_be(prime, frais, prestation, prestation_fdb, sim, nom_produit)
}
\arguments{
\item{prime}{une \code{matrix} a inserer dans la base.}

\item{frais}{une \code{matrix} a inserer dans la base.}

\item{prestation}{une \code{matrix} a inserer dans la base.}

\item{prestation_fdb}{une \code{matrix} a inserer dans la base.}

\item{sim}{un \code{integer} representant le numero de la simulation.}

\item{nom_produit}{une liste de \code{character} contenant les noms des differents produits.}
}
\description{
Jointure et insertion des donnees dans la base de donnees. Les donnees sont les ouputs (\code{output_be}) de la fonction \code{\link{proj_an}}
}
\author{
Prim'Act
}
