% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBase_merge_van_agg.R
\docType{methods}
\name{merge_van_agg}
\alias{merge_van_agg}
\title{Insertion des VAN dans la table \code{VAN_AGG} de la base de donnees.}
\usage{
merge_van_agg(result_tech, result_fin, result_brut, result_net, sim)
}
\arguments{
\item{result_tech}{une \code{matrix} contenant le resultat technique.}

\item{result_fin}{une \code{matrix} contenant le resultat financier.}

\item{result_brut}{une \code{matrix} contenant le resultat brut.}

\item{result_net}{une \code{matrix} contenant le resultat net.}

\item{sim}{un \code{integer} representant le numero de la simulation.}
}
\description{
Insertion des VAN dans la table \code{VAN_AGG} de la base de donnees.
}
\author{
Prim'Act
}
