% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_class.r
\docType{class}
\name{Oblig}
\alias{Oblig}
\title{La classe Oblig}
\description{
Classe pour les actifs de type Obligation
}
\section{Slots}{

\describe{
\item{\code{ptf_oblig}}{est un dataframe, chaque ligne represente un actif obligataire du portefeuille d'obligation.}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

