% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamAlmEngine_load.R
\docType{methods}
\name{param_alm_engine_load}
\alias{ParamAlmEngine}
\alias{param_alm_engine_load}
\title{Chargement des attributs d'un objet \code{ParamAlmEngine} a partir des donnees utilisateurs.}
\usage{
param_alm_engine_load(file_alm_address, ptf_fin_ref)
}
\arguments{
\item{file_alm_address}{un \code{character} renvoyant .}

\item{ptf_fin_ref}{un objet de la classe \code{PortFin} correspondant au portefeuille de reinvestissement.}
}
\value{
Pas de sortie.
}
\description{
\code{param_alm_engine_load} est la methode de chargement des attributs d'un objet ParamAlmEngine
a partir des donnees de l'environnement utilisateur et d'un portefeuille financier de reference (charge par la fonction \code{\link{chargement_PortFin_ref}}.
}
\author{
Prim'Act
}

