% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamRevaloEngine_load.R
\docType{methods}
\name{param_revalo_load}
\alias{param_revalo_load}
\title{Chargement des attributs d'un objet \code{ParamRevaloEngine} a partir des donnees utilisateurs.}
\usage{
param_revalo_load(file_revalo_address)
}
\arguments{
\item{file_revalo_address}{un \code{character} contenant l'adresse exacte
du fichier d'input utilisateur.}
}
\value{
L'objet de la classe \code{\link{ParamRevaloEngine}} construit a partir des inputs renseignes par l'utilisateur.
}
\description{
\code{param_revalo_load} est la methode de chargement des attributs d'un objet \code{\link{ParamRevaloEngine}}
a partir des donnees de l'environnement utilisateur.
}
\author{
Prim'Act
}

