% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamAlmEngine_class.R
\docType{class}
\name{ParamAlmEngine}
\alias{ParamAlmEngine}
\title{La classe ParamAlmEngine}
\description{
Classe pour le parametre ALM d'un canton.
}
\section{Slots}{

\describe{
\item{\code{ptf_reference}}{est un objet de type \code{PortFin}, qui represente le portefeuille
d'investissement de reference d'un canton.}

\item{\code{alloc_cible}}{Vecteur de 4 elements rendant compte du pourcentage de l'actif composant respectivement les investissements: actions, immobiliers, obligataires et de tresorerie.}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Methode principale : \code{\link{param_alm_engine_load}}
}
\keyword{classes}

