% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamBe_class.R
\docType{class}
\name{ParamBe}
\alias{ParamBe}
\title{La classe \code{ParamBe}.}
\description{
Une classe contenant le nombre d'annees de projection utilise
pour le calcul du best estimate d'un assureur.
}
\section{Slots}{

\describe{
\item{\code{nb_annee}}{un entier comprenant le nombre d'annees de projection.}
}}
\author{
Prim'Act
}
\keyword{classes}

