% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamChocMket_class.R
\docType{class}
\name{ParamChocMket}
\alias{ParamChocMket}
\title{La classe \code{ParamChocMket}.}
\description{
Une classe contenant les parametres des chocs de marche de la formule standard.
}
\section{Slots}{

\describe{
\item{\code{table_choc_action_type1}}{un \code{data.frame} contenant les parametres du choc action type 1.}

\item{\code{table_choc_action_type2}}{un \code{data.frame} contenant les parametres du choc action type 2.}

\item{\code{table_choc_immo}}{un \code{data.frame} contenant les parametres du choc immobilier.}

\item{\code{table_choc_spread}}{un \code{data.frame} contenant les parametres du choc de spread.}

\item{\code{table_choc_currency}}{un \code{data.frame} contenant les parametres du choc currency.}
}}

\author{
Prim'Act
}
\keyword{classes}
