% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamChocSousc-class.R
\docType{class}
\name{ParamChocSousc}
\alias{ParamChocSousc}
\title{La classe \code{ParamChocSousc}.}
\description{
Une classe contenant les parametres des chocs souscription de la formule standard.
}
\section{Slots}{

\describe{
\item{\code{mp}}{un \code{data.frame} contenant l'ensemble des parametres necessaires a l'application des
chocs du module Souscription Vie.}
}}

\author{
Prim'Act
}
\keyword{classes}
