% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamComport-class.R
\docType{class}
\name{ParamComport}
\alias{ParamComport}
\title{La classe ParamComport}
\description{
Classe pour pour les parametres de comportement
}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

