% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamRachDyn-class.R
\docType{class}
\name{ParamRachDyn}
\alias{ParamRachDyn}
\title{La classe ParamRachDyn}
\description{
Classe pour les tables de parametres pour les rachats dynamiques
}
\section{Slots}{

\describe{
\item{\code{vec_param}}{un data frame contenant les parametres pour les rachats dynamiques}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

