% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamRevaloEngine_class.R
\docType{class}
\name{ParamRevaloEngine}
\alias{ParamRevaloEngine}
\title{La classe \code{ParamRevaloEngine}.
Une Classe pour les parametres utilises pour la gestion de la revalorisation.}
\description{
La classe \code{ParamRevaloEngine}.
Une Classe pour les parametres utilises pour la gestion de la revalorisation.
}
\section{Slots}{

\describe{
\item{\code{taux_pb_fi}}{une valeur \code{numeric} correspondant au taux de participation applique au resultat financier.}

\item{\code{taux_pb_tech}}{une valeur \code{numeric} correspondant au taux de participation applique au resultat technique.}

\item{\code{tx_marge_min}}{une valeur \code{numeric} correspondant au taux de marge minimal auquel s'attend l'assureur.}

\item{\code{solde_pb_regl}}{une valeur \code{numeric} correspondant au solde deficitaire de participation aux
benefices reglementaire. Cette valeur doit etre negative.}
}}
\author{
Prim'Act
}
\keyword{classes}

