% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamTableRach-class.R, R/ParamTableMort-class.R
\docType{class}
\name{ParamTableMort}
\alias{ParamTableMort}
\title{La classe de parametres pour les table de rachat \code{ParamTableRach}.}
\description{
Une classe de parametres pour les tables de rachat.

Une classe de parametres pour les tables de mortalite.
}
\details{
Une table de rachat peut etre une table de rachat partiel ou une table de rachat total.
Pour une table de rachat partiel, les taux de rachat sont exprimes en pourcentage de l'encours.
Pour une table de rachat total, les taux de rachat sont exprimes en pourcentage du nombre de contrats.
}
\section{Slots}{

\describe{
\item{\code{age_min}}{un entier correspondant a l'age minimal de la table.}

\item{\code{age_max}}{un entier correspondant a l'age maximal de la table.}

\item{\code{anc_min}}{un entier correspondant a la premiere anciennete de la table.}

\item{\code{anc_max}}{un entier correspondant a la derniere anciennete de la table.}

\item{\code{table}}{un \code{data frame} contenant la table de rachat.}

\item{\code{age_min}}{un entier correspondant a l'age minimal de la table.}

\item{\code{age_max}}{un entier correspondant a l'age maximal de la table.}

\item{\code{gen_min}}{un entier correspondant a la premiere generation de la table.}

\item{\code{gen_max}}{un entier correspondant a la derniere generation de la table.}

\item{\code{table}}{un \code{data frame} contenant la table de mortalite.}
}}
\note{
Les tables de mortalite doivent contenir des effectifs sous risque par age (Lx).
}
\author{
Prim'Act

Prim'Act
}
\seealso{
Le calcul du taux de rachat \code{\link{calc_rach}}.

Le calcul du taux de deces \code{\link{calc_qx}}.
}
\keyword{classes}

