% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamTableRach-class.R
\docType{class}
\name{ParamTableRach}
\alias{ParamTableRach}
\title{La classe de parametres pour les table de rachat \code{ParamTableRach}.}
\description{
Une classe de parametres pour les tables de rachat.
}
\details{
Une table de rachat peut etre une table de rachat partiel ou une table de rachat total.
Pour une table de rachat partiel, les taux de rachat sont exprimes en pourcentage de l'encours.
Pour une table de rachat total, les taux de rachat sont exprimes en pourcentage du nombre de contrats.
}
\section{Slots}{

\describe{
\item{\code{age_min}}{un entier correspondant a l'age minimal de la table.}

\item{\code{age_max}}{un entier correspondant a l'age maximal de la table.}

\item{\code{anc_min}}{un entier correspondant a la premiere anciennete de la table.}

\item{\code{anc_max}}{un entier correspondant a la derniere anciennete de la table.}

\item{\code{table}}{un \code{data.frame} contenant la table de rachat.}
}}

\seealso{
Le calcul du taux de rachat \code{\link{calc_rach}}.
}
\author{
Prim'Act
}
\keyword{classes}
