% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortFin_class.R
\docType{class}
\name{PortFin}
\alias{PortFin}
\title{La classe PortFin}
\description{
Classe pour le portefeuille global d'actif
}
\section{Slots}{

\describe{
\item{\code{ptf_action}}{est un objet de type Action, qui represente le portefeuille d'action d'un canton.}

\item{\code{ptf_immo}}{est un objet de type Immo, qui represente le portefeuille immobilier d'un canton.}

\item{\code{ptf_oblig}}{est un objet de type Oblig, qui represente le portefeuille obligataire d'un canton.}

\item{\code{ptf_treso}}{est un objet de type Treso, qui represente le portefeuille monetaire d'un canton.}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

