% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortPassif-class.R
\docType{class}
\name{PortPassif}
\alias{PortPassif}
\title{La classe \code{PortPassif}.}
\description{
Une classe regroupant l'ensemble des donnees de passifs et les hypotheses correspondantes.
}
\section{Slots}{

\describe{
\item{\code{annee}}{une valeur entiere correspondant a l'annee de projection.}

\item{\code{eei}}{une liste d'objets de la classe \code{\link{EpEuroInd}} contenant l'ensemble
des produits de type epargne en euros.}

\item{\code{names_class_prod}}{un vecteur \code{character} indiquant les noms de classes de produits.}

\item{\code{ht}}{un objet de classe \code{\link{HypTech}} contenant les hypotheses techniques.}

\item{\code{fp}}{un objet de classe \code{\link{FraisPassif}} contenant les hypotheses de frais de passif
par produit.}

\item{\code{tx_pb}}{un objet de classe \code{\link{TauxPB}} contenant les taux contractuel de participation
aux benefices par produit.}

\item{\code{autres_passifs}}{un objet de classe \code{\link{AutresPassifs}}.}

\item{\code{autres_reserves}}{un objet de classe \code{\link{AutresReserves}}.}
}}

\seealso{
La projection des produits sur l'annee avant attributiuon de participation
aux benefices : \code{\link{proj_annee_av_pb}}.
Le vieillissement des model points de passifs avant et apres attributiuon de participation
aux benefices : \code{\link{vieillissement_av_pb}}, \code{\link{vieillissement_ap_pb}}.
}
\author{
Prim'Act
}
\keyword{classes}
