% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppb_class.R
\docType{class}
\name{Ppb}
\alias{Ppb}
\title{La classe \code{Ppb}.}
\description{
Classe pour la provision pour participation aux benefices (PPB)
}
\section{Slots}{

\describe{
\item{\code{hist_ppb}}{est le vecteur contenant les valeurs courantes \code{numeric} prise par la PPB sur les huit dernieres annees.}

\item{\code{valeur_ppb}}{est la valeur courante \code{numeric} prise par la PPB.}

\item{\code{ppb_debut}}{est la valeur prise \code{numeric} par la PPB en debut d'annee.}

\item{\code{seuil_rep}}{est une valeur \code{numeric} correspondant a la proportion de PPB de debut d'annee
que l'on peut reprendre sur une periode.}

\item{\code{seuil_dot}}{est une valeur \code{numeric} correspondant au montant maximal de dotation possible sur la PPB
sur une periode, exprimee comme une fraction de la PPB de debut d'annee.}

\item{\code{compte_rep}}{est une valeur \code{numeric} qui totalise les montants de reprises effectuees sur une periode.}

\item{\code{compte_dot}}{est une valeur \code{numeric} qui totalise les montants de dotations effectuees sur une periode.}
}}

\seealso{
La dotation et la reprise de PPB : \code{\link{calc_dotation_ppb}}, \code{\link{calc_reprise_ppb}}.
}
\author{
Prim'Act
}
\keyword{classes}
