% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppb_ppb_8ans.R
\docType{methods}
\name{ppb_8ans}
\alias{ppb_8ans}
\title{Calcule la valeur de la PPB a distribue en appliquant la regle des 8 ans.}
\usage{
ppb_8ans(x)
}
\arguments{
\item{x}{un objet de la classe \code{\link{Ppb}}.}
}
\value{
une liste contenant les parametres
\describe{
\item{\code{ppb_8} : }{la valeur \code{numeric} correspondant au montant de la ppb de l'annee t-8.}
\item{\code{ppb} : }{un objet \code{\link{Ppb}} correspondant a la PPB mise a jour.}
}
}
\description{
\code{ppb_8ans} est une methode permettant d'appliquer la regle des 8 ans.
}
\author{
Prim'Act
}
