% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppb_load.R
\docType{methods}
\name{ppb_load}
\alias{ppb_load}
\title{Methode permettant de charger les valeurs des hypotheses et des donnees de PPB}
\usage{
ppb_load(file_ppb_address)
}
\arguments{
\item{file_ppb_address}{est un \code{character} contenant l'adresse exacte du fichier d'input utilisateur
permettant de renseigner un objet \code{\link{Ppb}}.}
}
\value{
L'objet de la classe \code{\link{Ppb}} construit a partir des inputs renseignes par l'utilisateur.
}
\description{
\code{ppb_load} est une methode permettant de charger les parametres associees a un
objet de classe \code{\link{Ppb}}.
}
\seealso{
La classe \code{\link{Initialisation}} et sa methode \code{\link{set_architecture}} pour renseigner l'input.
}
\author{
Prim'Act
}
