% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_class.r
\docType{class}
\name{PRE}
\alias{PRE}
\title{La classe PRE}
\description{
Classe pour la gestion de la provision pour risque d'exigibilite (PRE).
}
\section{Slots}{

\describe{
\item{\code{val_debut}}{est une valeur \code{numeric} correspondant a la valeur de la PRE en debut d'annee.}

\item{\code{val_courante}}{est une valeur \code{numeric} correspondant a la valeur courante de la PRE.}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

