% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortFin_calc_value.R
\docType{methods}
\name{print_alloc}
\alias{PortFin}
\alias{print_alloc}
\title{Calcul le poids de chaque composante du portefeuille action.}
\usage{
print_alloc(x)
}
\arguments{
\item{x}{objet de la classe PortFin.}
}
\value{
Un data frame compose de deux colonnes et cinq lignes.
La premiere colonne decrit le montant alloue par poche et au total en valeur.
La seconde colonne decrit le montant alloue par poche en proportion.
Les lignes correspondent aux actifs : (Action / Immobilier / Obligation / Tresorerie / Actifs cumules)
}
\description{
\code{pint_alloc} est une methode permettant de calculer l'allocation absolue et relative du portefeuille.
}
\author{
Prim'Act
}

