% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortPassif-proj_annee_av_pb.R
\docType{methods}
\name{proj_annee_av_pb}
\alias{PortPassif}
\alias{proj_annee_av_pb}
\title{Calcul les flux et les PM des produits modelises}
\usage{
proj_annee_av_pb(an, x, tx_soc, coef_inf, list_rd)
}
\arguments{
\item{an}{est l'annee de projection.}

\item{x}{un objet de la classe \code{PortPassif} contenant l'ensemble des produits de passifs.}

\item{tx_soc}{le taux de charges sociales.}

\item{coef_inf}{est le coefficient d'inflation considere pour le traitement des frais.}

\item{list_rd}{est une liste contenant les rendements des actifs de references.}
}
\value{
Une liste dont le premier element designe les noms des produits, puis deux matrices de resultats aggreges : une pour les flux
et une pour le stock. Retourne egalement le portefeuille avec une mise a jour des tables de resultats intermediaires.
}
\description{
\code{proj_annee_av_pb} est une methode permettant de calculer les PM et les flux sur une annee avant PB.
Cette me
thode calcule egalement les frais sur flux et PM.
}
\details{
L'annee de projection est utilisee pour gerer les produits dont les clauses dependent de l'annee
}
\author{
Prim'Act
}

