% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC_class.R
\docType{class}
\name{RC}
\alias{RC}
\title{La classe \code{RC}.}
\description{
Une classe pour la gestion de la reserve de capitalisation (RC).
}
\section{Slots}{

\describe{
\item{\code{val_debut}}{est une valeur \code{numeric} correspondant a la valeur de la RC en debut d'annee.}

\item{\code{val_courante}}{est une valeur \code{numeric} correspondant a la valeur courante de la RC.}
}}

\seealso{
Les methodes de calcul de la RC \code{\link{calc_RC}},
et de mises a jour des RC initiales et courantes \code{\link{do_update_RC_val_courante}}, \code{\link{do_update_RC_val_debut}}.
}
\author{
Prim'Act
}
\keyword{classes}
