% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC_load.R
\docType{methods}
\name{rc_load}
\alias{RC}
\alias{rc_load}
\title{Chargement de la valeur initiale de la RC}
\usage{
rc_load(file_RC_address)
}
\arguments{
\item{file_RC_address}{est un \code{character} correspondant a l'adresse du fichier d'input renseignant les donnees de RC}
}
\value{
Un objet de la classe \code{RC} charge a partir des donnees du fichier dont le nom est RCcise en input.
}
\description{
\code{rc_load} est une methode permettant de charger la valeur de RC initiale dans un objet de type RC.
}
\author{
Prim'Act
}

