% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlmEngine_reallocate.R
\docType{methods}
\name{reallocate}
\alias{AlmEngine}
\alias{reallocate}
\title{Realise les operations d'achats ventes}
\usage{
reallocate(x, ptf_reference, alloc_cible)
}
\arguments{
\item{x}{objet de la classe \code{\link{PortFin}}.}

\item{ptf_reference}{est le portefeuille de reinvestissement. C'est un objet de la classe \code{\link{PortFin}}.}

\item{alloc_cible}{est un vecteur de type \code{numeric} constitue de 4 elements, il contient les proportions cibles d'allocations
action, immobilier, obligataire et de tresorerie.}
}
\value{
\code{portFin} l'objet initial de la classe \code{\link{PortFin}} realloue a l'allocation cible.

\code{pmvr} le montant total des pus ou moins values realisess.

\code{pmvr_oblig} le montant des plus ou moins values obligataires realisees lors de la reallocation.

\code{pmvr_action} le montant des plus ou moins values action realisees lors de l'etape de reallocation.

\code{pmvr_immo}  le montant des plus ou moins values immobilieres realisees lors de l'etape de reallocation.

\code{var_rc} la variation de la reserve de capitalisation induite par la reallocation.

\code{var_pre} la variation de la provision pour risque d'exigibilite induite par la reallocation.

\code{plac_moy_vm} la valeur de marche moyenne des placements de l'assureur au cours de l'operation de reallocation.

\code{plac_moy_vnc} la valeur nette comptable moyenne des placements de l'assureur au cours de l'operation de reallocation.
}
\description{
\code{reallocate} est une methode permettant d'ajuster l'allocation du \code{\link{PortFin}} de l'assureur.
}
\note{
Les operations d'achat/vente sont effectuees en termes de nombre d'unite d'achat/vente.
}
\author{
Prim'Act
}
\seealso{
La classe \code{\link{PortFin}}.
}

