% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action_flux.R
\docType{methods}
\name{revalo_action}
\alias{Action}
\alias{revalo_action}
\title{Calcul les valeurs de marches de chaque composante du portefeuille action.}
\usage{
revalo_action(x, S, S_prev)
}
\arguments{
\item{x}{objet de la classe Action (decrivant le portefeuille d'action).}

\item{S}{vecteur de valeur de chaque stock du ptf en milieu d'annee N (date de versement des dividendes)}

\item{S_prev}{vecteur de valeur de chaque stock du ptf en milieu d'annee N-1.}
}
\value{
Un data frame composé de deux colonnes et autant de lignes que le portefeuille action a de lignes.
La premiere colonne decrit de le rendement annuel de chacune des actions composants le portefeuille action.
La seconde colonne decrit les dividendes annuelles percues au titre de chacune des actions composants le portefeuille action.
}
\description{
\code{revalo_action} est une methode permettant de calculer les valeurs de marche.
}
\author{
Prim'Act
}

