% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Treso_flux.R
\docType{methods}
\name{revalo_treso}
\alias{revalo_treso}
\alias{treso}
\title{Calcul les valeurs de marches de chaque composante du portefeuille treso.}
\usage{
revalo_treso(Rt, Rt_prev)
}
\arguments{
\item{S}{vecteur de valeur de chaque ligne du ptf en milieu d'annee N (date de calcul des flux).}

\item{S_prev}{vecteur de valeur de chaque ligne du ptf en milieu d'annee N-1.}
}
\value{
Un vecteur ayant autant d elements que les vecteurs inputs. Chaque element correspondant au rendement annuel d'une lige de tresorerie.
}
\description{
\code{revalo_treso} est une methode permettant de calculer les valeurs de marche.
}
\author{
Prim'Act
}

