% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Be-run_be.R
\docType{methods}
\name{run_be}
\alias{Be}
\alias{run_be}
\title{Calcul d'un BE}
\usage{
run_be(x, pre_on)
}
\arguments{
\item{x}{est un objet de type \code{Be}.}

\item{pre_on}{est une valeur \code{logical} qui lorsqu'elle vaut \code{TRUE} prend en compte la variation
de PRE dans le resultat technique utilisee pour le calcul de la participation aux benefices reglementaires.}
}
\value{
L'objet BE avec les tables de resultats mise a jour et la liste des simulations ignorees en
cas d'erreur.
}
\description{
\code{run_be} est une methode permettant de calcul un best estimate pour un canton.
}
\author{
Prim'Act
}

