% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Be-run_be_simu.R
\docType{methods}
\name{run_be_simu}
\alias{Be}
\alias{run_be_simu}
\title{Calcul d'un BE par simulation}
\usage{
run_be_simu(x, i)
}
\arguments{
\item{x}{est un objet de type \code{Be}.}

\item{i}{est un entier (\code{integer}) correspondant au numero de la simulation.}
}
\value{
Une liste contenant les noms de produits calcules, les flux par produits et par annee pour une simuation sous
forme de matrice et les flux actualises par produits.
}
\description{
\code{run_be_simu} est une methode intermediare permettant de calcul un best estimate
pour une simulation donnee.
}
\author{
Prim'Act
}

