% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChocSolvabilite2_save_bscr.R
\docType{methods}
\name{save_bscr}
\alias{save_bscr}
\title{Enregistre les resultats d'une evaluation BSCR}
\usage{
save_bscr(nom, path, x)
}
\arguments{
\item{nom}{est un objet de type \code{character} utilise pour nommer le fichier de resultats.}

\item{path}{est un objet de type \code{character} utilise pour indiquer le chemin d'enregistrement des resultats.}

\item{x}{est un objet de type \code{list} contenant les r?sultats d'une evalution BSCR.}
}
\description{
\code{save_bscr} est une methode permettant d'enregistrer en \code{.cvs} les resultats
d'une evaluation BSCR.
}
\author{
Prim'Act
}
