% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immo_buy_sell.R
\docType{methods}
\name{sell_immo}
\alias{immo}
\alias{sell_immo}
\title{Mise a jour de chaque composante du portefeuille immo suite a la vente de tout ou partie de ce portefeuille.}
\usage{
sell_immo(x, num_sold, nb_sold)
}
\arguments{
\item{x}{objet de la classe Immo (decrivant le portefeuille immo en detention).}

\item{num_sold}{vecteur contenant le numero de model point des biens immobiliers du portefeuille que l'on souhaite vendre.}

\item{nb_sold}{vecteur contenant le nombre d'unite que l'on souhaite vendre (a autant de ligne que le vecteur num_sold).}
}
\value{
L'objet x mis a jour de l'operation de vente (suppression des lignes vendues).
}
\description{
\code{sell_immo} est une methode permettant de mettre a jour chaque composante d'un portefeuille immo suite a la vente
de tout ou partie de ce portefeuille.
}
\author{
Prim'Act
}

