% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_buy_sell.R
\docType{methods}
\name{sell_oblig}
\alias{Oblig}
\alias{sell_oblig}
\title{Mise a jour de chaque composante d'un portefeuille obligataire suite a la vente de tout ou partie du portefeuille obligataire.}
\usage{
sell_oblig(x, num_sold, nb_sold)
}
\arguments{
\item{x}{objet de la classe Oblig (decrivant le portefeuille obligataire en detention).}

\item{num_sold}{vecteur contenant le numero de model point obligataire du portefeuille que l'on souhaite vendre.}

\item{nb_sold}{vecteur contenant le nombre d'unite que l'on souhaite vendre (a autant de ligne que le vecteur num_sold).}
}
\value{
L'objet x mis a jour de l'operation de vente (suppression des lignes vendues).
}
\description{
\code{sell_oblig} est une methode permettant de mettre a jour chaque composante d'un portefeuille obligataire suite a la vente
de tout ou partie du portefeuille obligataire.
}
\author{
Prim'Act
}

