% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action_sell_pvl.R
\docType{methods}
\name{sell_pvl_action}
\alias{sell_pvl_action}
\title{Mise a jour du portefeuille action suite a une realisation de plus-values latentes action.}
\usage{
sell_pvl_action(x, montant)
}
\arguments{
\item{x}{un objet de la classe \code{\link{Action}} (decrivant le portefeuille action en detention).}

\item{montant}{un montant \code{numeric} de plus-values latentes que l'on souhaite realiser.}
}
\value{
\code{action} l'objet \code{x} mis a jour de l'operation de vente (suppression des lignes vendues).

\code{pmvr} le montant des plus ou moins-values realisees.
}
\description{
\code{sell_pvl_action} est une methode permettant de mettre a jour chaque composante d'un portefeuille action
 suite a la vente de tout ou partie de ce portefeuille afin de realiser un montant de plus-values latentes.
}
\note{
Les cessions sont realisees au prorata des plus-values latentes actions.
}
\author{
Prim'Act
}
