% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialisation_set_architecture.R
\docType{methods}
\name{set_architecture}
\alias{Initialisation}
\alias{set_architecture}
\title{Definition de l'architecture d'un workspace.}
\usage{
set_architecture(x)
}
\arguments{
\item{x}{un objet de la classe \code{Initialisation}.}
}
\value{
Objet mis a jour des chemins.
}
\description{
\code{set_architecture}.
}
\author{
Prim'Act
}

