% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabEpEuroInd-class.R
\docType{class}
\name{TabEpEuroInd}
\alias{TabEpEuroInd}
\title{La classe \code{TabEpEuroInd}.}
\description{
Une classe pour le stockage en memoire de variable de calcul au niveau du model point \code{\link{EpEuroInd}}.
}
\section{Slots}{

\describe{
\item{\code{tab}}{un objet \code{list} au format fige contenant l'ensemble des variables stockees.}
}}

\author{
Prim'Act
}
\keyword{classes}
