% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabProbaEpEuroInd-class.R
\docType{class}
\name{TabProbaEpEuroInd}
\alias{TabProbaEpEuroInd}
\title{La classe \code{TabProbaEpEuroInd}.}
\description{
Une classe pour le stockage en memoire des differentes probabilites invariantes au niveau du model point \code{\link{EpEuroInd}}.
}
\section{Slots}{

\describe{
\item{\code{rachat_tot}}{est un \code{data.frame} contenant les taux de rachat totaux pour chaque model point.}

\item{\code{rachat_part}}{est un \code{data.frame} contenant les taux de rachat partiels pour chaque model point.}

\item{\code{qx_mort}}{est un \code{data.frame} contenant les probabilite des deces pour chaque model point.}
}}

\author{
Prim'Act
}
\keyword{classes}
