% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabProbaRetEuroRest-class.R
\docType{class}
\name{TabProbaRetEuroRest}
\alias{TabProbaRetEuroRest}
\title{La classe \code{TabProbaRetEuroRest}.}
\description{
Une classe pour le stockage en memoire des differentes probabilites invariantes au niveau du model point \code{\link{RetraiteEuroRest}}.
}
\section{Slots}{

\describe{
\item{\code{ax}}{est un \code{data.frame} contenant les valeurs des coefficients actuariels pour chaque model point.}

\item{\code{sortie_retraite}}{est un \code{data.frame} contenant les la probabilite de sortie d'un contrat retraite pour chaque model point.}

\item{\code{survie_un_an}}{est un \code{data.frame} contenant les valeurs des coefficients actuariels pour chaque model point.}
}}

\author{
Prim'Act
}
\keyword{classes}
