% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabRetEuroRest_class.R
\docType{class}
\name{TabRetEuroRest}
\alias{TabRetEuroRest}
\title{La classe \code{TabRetEuroRest}.}
\description{
Une classe pour le stockage en memoire de variable de calcul au niveau du model point de retraite euro en phase de restitution \code{\link{TabRetEuroRest}}.
}
\section{Slots}{

\describe{
\item{\code{tab}}{un objet \code{list} au format fige contenant l'ensemble des variables stockees.}
}}

\author{
Prim'Act
}
\keyword{classes}
