% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TauxPB-class.R
\docType{class}
\name{TauxPB}
\alias{TauxPB}
\title{La classe \code{TauxPB}.}
\description{
Une classe pour le stockage des parametres de taux de participation contractuelle par produit.
}
\section{Slots}{

\describe{
\item{\code{mp}}{un \code{data frame} contenant les parametres des taux de participation contractuelle par produit.}
}}

\author{
Prim'Act
}
\keyword{classes}
