% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TauxPB_load.R
\docType{methods}
\name{tauxpb_load}
\alias{tauxpb_load}
\title{Methode permettant de charger les valeurs des hypotheses et des donnees de TauXPB}
\usage{
tauxpb_load(file_tauxpb_address)
}
\arguments{
\item{file_tauxpb_address}{est un \code{character} contenant l'adresse exacte
du fichier d'input utilisateur permettant de renseigner un objet \code{\link{TauxPB}}.}
}
\value{
L'objet de la classe \code{\link{TauxPB}} construit a partir des inputs renseignes par l'utilisateur.
}
\description{
\code{tauxpb_load} est une methode permettant de charger les parametres associees a un objet de classe \code{\link{TauxPB}}.
}
\seealso{
La classe \code{\link{Initialisation}} et sa methode \code{\link{set_architecture}}
 pour renseigner l'input.
}
\author{
Prim'Act
}
