% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_update_cc.R
\docType{methods}
\name{update_cc_oblig}
\alias{Oblig}
\alias{update_cc_oblig}
\title{Mise a jour des coupons courus d'un portefeuille obligataire.}
\usage{
update_cc_oblig(x, coupon)
}
\arguments{
\item{x}{objet de la classe \code{Oblig} (decrivant le portefeuille obligataire en detention).}

\item{coupon}{un vecteur de \code{numeric} a assigner a l'objet \code{Obligation}.}
}
\value{
L'objet \code{x} dont les coupons courus ont ete mis a jour
}
\description{
\code{update_cc_oblig} est une methode permettant de mettre a jour les coupons courus des composantes d'un portefeuille obligataire.
}
\author{
Prim'Act
}

