% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_update_dur.R
\docType{methods}
\name{update_dur_oblig}
\alias{update_dur_oblig}
\title{Mise a jour des durations d'un portefeuille obligataire.}
\usage{
update_dur_oblig(x, duration)
}
\arguments{
\item{x}{objet de la classe \code{\link{Oblig}} (decrivant le portefeuille obligataire en detention).}

\item{duration}{un vecteur de \code{numeric} a assigner a l'objet \code{\link{Oblig}}.}
}
\value{
L'objet \code{x} dont les durations ont ete mises a jour.
}
\description{
\code{update_dur_oblig} est une methode permettant de mettre a jour la duration des composantes d'un portefeuille obligataire.
}
\author{
Prim'Act
}
