% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_update.R
\docType{methods}
\name{update_mat_res}
\alias{Oblig}
\alias{update_mat_res}
\title{Mise a jour de la maturite residuelle et de la duree de detention de chaque composante d'un portefeuille obligataire.}
\usage{
update_mat_res(x)
}
\arguments{
\item{x}{objet de la classe Oblig (decrivant le portefeuille obligataire).}
}
\description{
\code{update_mat_res} est une methode permettant de mettre a jour la maturite residuelle et la duree de detention
de chaque composante d'un portefeuille obligataire.
}
\author{
Prim'Act
}

