% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_update_sd.R
\docType{methods}
\name{update_sd_oblig}
\alias{update_sd_oblig}
\title{Mise a jour des surcotes decotes d'un portefeuille obligataire.}
\usage{
update_sd_oblig(x, sd)
}
\arguments{
\item{x}{objet de la classe \code{\link{Oblig}} (decrivant le portefeuille obligataire en detention).}

\item{sd}{un vecteur de \code{numeric} a assigner a l'objet \code{\link{Oblig}}.}
}
\value{
L'objet \code{x} dont les surcotes decotes ont ete mises a jour.
}
\description{
\code{update_sd_oblig} est une methode permettant de mettre a jour la surcotes decotes des composantes d'un portefeuille obligataire.
}
\author{
Prim'Act
}

