% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBase_update_table_pb.R
\docType{methods}
\name{update_table_output_pb}
\alias{update_table_output_pb}
\title{Mise a jour de la table des PB. Ajoute une colonne "diff_pb".}
\usage{
update_table_output_pb(table)
}
\arguments{
\item{table}{est un objet de type \code{data.frame}.}
}
\value{
La table mise a jour.
}
\description{
Mise a jour de la table des PB. Ajoute une colonne "diff_pb".
}
\author{
Prim'Act
}
