% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBase_update_table_output_produit.R
\docType{methods}
\name{update_table_output_produit}
\alias{update_table_output_produit}
\title{Mise a jour de la table OUTPUT_PRODUIT. Ajoute plusieurs colonnes : chgt, delta_pm, frais, credit, debit, resulat.}
\usage{
update_table_output_produit(table)
}
\arguments{
\item{table}{est un objet de type \code{data.frame}.}
}
\value{
La table mise a jour.
}
\description{
Mise a jour de la table OUTPUT_PRODUIT. Ajoute plusieurs colonnes : chgt, delta_pm, frais, credit, debit, resulat.
}
\author{
Prim'Act
}
