% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Treso_update.R
\docType{methods}
\name{update_treso}
\alias{update_treso}
\title{Permet d'integrer un flux (entrant ou sortant) au compte de tresorerie d'un portefeuille financier.}
\usage{
update_treso(x, flux)
}
\arguments{
\item{x}{objet de la classe \code{\link{Treso}}, correspondant a l'actif Tresorerie d'un assureur anterieur a integration d'un flux.}

\item{flux}{est un \code{numeric} correspondant a un flux.
S'il est positif, le flux est entrant.
S'il est negatif, le flux est sortant.}
}
\value{
L'objet \code{x} mis a jour du flux precise en input.
}
\description{
\code{update_treso} est une methode permettant d'integrer un flux au compte de tresorerie.
}
\author{
Prim'Act
}
