% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immo_update_vm.R
\docType{methods}
\name{update_vm_immo}
\alias{Immo}
\alias{update_vm_immo}
\title{Mise a jour de chaque composante du portefeuille immobilier suite a la vente de tout ou partie de ce portefeuille.}
\usage{
update_vm_immo(x, vm)
}
\arguments{
\item{x}{objet de la classe \code{Immo} (decrivant le portefeuille immobilier en detention).}
}
\value{
L'objet \code{x} mis a jour du vieillissement de la duree de detention.
}
\description{
\code{update_vm_immo} est une methode permettant de mettre a jour la duree de detention des composantes d'un portefeuille immobilier.
}
\author{
Prim'Act
}

