% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_update_vnc.R
\docType{methods}
\name{update_vnc_oblig}
\alias{update_vnc_oblig}
\title{Mise a jour des valeurs nettes comptables d'un portefeuille obligataire.}
\usage{
update_vnc_oblig(x, vnc)
}
\arguments{
\item{x}{objet de la classe \code{\link{Oblig}} (decrivant le portefeuille obligataire en detention).}

\item{vnc}{un vecteur de \code{numeric} a assigner a l'objet \code{\link{Oblig}}.}
}
\value{
L'objet \code{x} dont les valeurs nettes comptables ont ete mis a jour
}
\description{
\code{update_vnc_oblig} est une methode permettant de mettre a jour les valeurs nettes comptables des composantes d'un portefeuille obligataire.
}
\author{
Prim'Act
}

