% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_update_zsp.R
\docType{methods}
\name{update_zsp_oblig}
\alias{update_zsp_oblig}
\title{Mise a jour des z-spreads d'un portefeuille obligataire.}
\usage{
update_zsp_oblig(x, zspread)
}
\arguments{
\item{x}{objet de la classe \code{\link{Oblig}} (decrivant le portefeuille obligataire en detention).}

\item{zspread}{un vecteur de \code{numeric} a assigner a l'objet \code{Obligation}.}
}
\value{
L'objet \code{x} dont les zspreads ont ete mis a jour
}
\description{
\code{update_zsp_oblig} est une methode permettant de mettre a jour les z-spreads des composantes d'un portefeuille obligataire.
}
\author{
Prim'Act
}

