% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpEuroInd-vieilli_mp.R
\docType{methods}
\name{vieilli_mp}
\alias{EpEuroInd}
\alias{vieilli_mp}
\title{Veilli un model point.}
\usage{
vieilli_mp(x, pm_fin_ap_pb, tx_revalo)
}
\arguments{
\item{x}{un objet de la classe \code{EpEuroInd} contenant les model points epargne euros}

\item{pm_fin_ap_pb}{un vecteur de type \code{numeric} contenant par model points
les montants de PM revalorises.}

\item{tx_revalo}{un vecteur de type \code{numeric} contenant par model points
les taux de revalorisation nets appliques.}
}
\value{
l'objet \code{x} vieilli d'une annee.
}
\description{
\code{vieilli_mp} est une methode permettant de vieillir
 un model point de 1 an.
}
\author{
Prim'Act
}

