% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortPassif-vieillissement_av_pb.R
\docType{methods}
\name{vieillissement_av_pb}
\alias{EpEuroInd}
\alias{vieillissement_av_pb}
\title{Vieillissement du portefeuille sur l'annee avant attribution
de participation aux benefices.}
\usage{
viellissement_av_pb(an, x, coef_inf, list_rd, tx_soc)
}
\arguments{
\item{an}{l'annee de realisation des calcul.}

\item{x}{un objet de la classe \code{PortPassif} contenant les produits du portefeuille.}

\item{coef_inf}{est le coefficient d'inflation considere pour le traitement des frais.}

\item{list_rd}{est une liste contenant les rendements des actifs de references.}

\item{tx_soc}{le taux de charges sociales.}
}
\value{
Une liste comprenant  :
\describe{
\item{\code{ptf} : }{Le portefeuille \code{x} mis a jour.}
\item{\code{result_av_pb} : }{Une liste dont le premier element designe les noms des produits,
 puis deux matrices de resultats aggreges : une pour les flux et une pour le stock.}
\item{\code{result_autres_passifs} : }{un vecteur contenant les resultats des passifs non modelises.}
\item{\code{var_psap} : }{La variation de PSAP sur l'annee.}
\item{\code{var_pgg} : }{La variation de PGG sur l'annee.}
\item{\code{flux_milieu} : }{Les flux de milieu d'annee entrant en tresorerie en milieu de periode.}
\item{\code{flux_fin} : }{Les flux de milieu d'annee entrant en tresorerie en fin de periode.}
}
}
\description{
\code{viellissement_av_pb} est de vieillir du portefeuille sur l'annee avant attribution
de participation aux benefices.
}
\author{
Prim'Act
}

