% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortFin_vieillissement_oblig.R
\docType{methods}
\name{vieillissement_oblig_PortFin}
\alias{PortFin}
\alias{vieillissement_oblig_PortFin}
\title{Effectue le vieillissement/la projection du portefeuille obligataire d'un portefeuille financier.}
\usage{
vieillissement_oblig_PortFin(x, new_mp_ESG)
}
\arguments{
\item{x}{objet de la classe \code{PortFin}, correspondant au portefeuille financier de l'assureur avant l'etape de vieillissement de son atribut \code{ptf_oblig} de la classe \code{Oblig}.}

\item{table_rdt}{est une \code{liste}, construite par la fonction \code{\link{calc_rdt}}.
Cette table contient les tables d'evolution des cours et rendements sur l'annee consideree de chacune des classes d'actif.
Les tables sont constuites a partir des extractions du Generateur de Scenario Economique de Prim'Act.}
}
\description{
\code{vieillissement_oblig_PortFin} est une methode permettant de projeter la composante obligataire d'un portefeuille financier.
}
\author{
Prim'Act
}
\seealso{
La fonction de calcul des rendements des actifs \code{\link{calc_rdt}}.
}

