% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppb_vieillissement_ppb.R
\docType{methods}
\name{vieillissement_ppb}
\alias{vieillissement_ppb}
\title{Vieillissement d'un an de la PPB.}
\usage{
vieillissement_ppb(x, pm_fin_ap_pb, tx_revalo)
}
\arguments{
\item{x}{objet de la classe \code{\link{Ppb}}.}
}
\value{
L'objet \code{x} vieilli d'une annee.
}
\description{
\code{vieillissement_ppb} est une methode permettant de vieillir d'un an la PPB. Cette methode permet de
reinitialiser les montants de dotation et de reprise cumules sur l'annee, de re-initialiser le montant de PPB de debut de periode,
et de mettre a jour le vecteur historique de la PPB.
}
\author{
Prim'Act
}
