% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortFin_vieillissement_treso.R
\docType{methods}
\name{vieillissement_treso_PortFin}
\alias{vieillissement_treso_PortFin}
\title{Effectue le vieillissement/la projection du portefeuille tresorerie d'un portefeuille financier.}
\usage{
vieillissement_treso_PortFin(x, flux_milieu, flux_fin, table_rdt)
}
\arguments{
\item{x}{objet de la classe \code{PortFin}, correspondant au portefeuille financier
de l'assureur avant l'etape de vieillissement.}

\item{flux_milieu}{est un \code{numeric} correspondant aux revenus percus en milieu d'annee (coupons obligataires, loyers, dividendes).}

\item{flux_fin}{est un \code{numeric} correspondant aux revenus percus en fin d'annee
(tombees d'echeances et revenus de tresorerie).}

\item{table_rdt}{est une \code{liste}, construite par la fonction \code{\link{calc_rdt}}.
Cette table contient les tables d'evolution des cours et rendements sur l'annee consideree de chacune
 des classes d'actif. Les tables sont constuites a partir des extractions du Generateur de Scenario Economique de Prim'Act.}
}
\value{
L'objet renvoye de la classe \code{PortFin} correspond au portefeuille financier initial dont l'attribut \code{ptf_treso} a ete vieilli d'une annee.
}
\description{
\code{vieillissement_treso_PortFin} est une methode permettant de projeter la composante obligataire d'un portefeuille financier.
}
\author{
Prim'Act
}
\seealso{
La fonction de calcul des rendements des actifs \code{\link{calc_rdt}}.
}

