% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortPassif-vieillissement_ap_pb.R
\docType{methods}
\name{vieillissment_ap_pb}
\alias{PortPassif}
\alias{vieillissment_ap_pb}
\title{Calcul les flux et les PM des produits modelises apres participation aux benefices}
\usage{
vieillissment_ap_pb(x, rev_nette_alloue, tx_soc)
}
\arguments{
\item{x}{un objet de la classe \code{PortPassif} contenant l'ensemble des produits de passifs.}

\item{rev_nette_alloue}{un vecteur \code{numeric} contenant par produit le supplement de revalorisation par rapport au taux minimum.}

\item{tx_soc}{le taux de charges sociales.}
}
\value{
Une liste dont le premier element designe le portefeuille mis a jour, puis
les noms des produits, puis deux matrices de resultats aggreges : une pour les flux
et une pour le stock.
}
\description{
\code{vieillissment_ap_pb} est une methode permettant de calculer les PM et les flux sur une annee apres PB.
Cette methode produit le portefeuille de passifs mis a jour.
}
\author{
Prim'Act
}

