% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Be-write_results.R
\docType{methods}
\name{write_be_results}
\alias{write_be_results}
\title{Enregistre les resultats d'une evaluation best estimate}
\usage{
write_be_results(nom_run, path, x)
}
\arguments{
\item{nom_run}{est un objet de type \code{character} utilise pour nommer le fichier de resultats.}

\item{path}{est un objet de type \code{character} utilise pour indiquer le chemin d'enregistrement des resultats.}

\item{x}{est un objet de type \code{\link{Be}}.}
}
\description{
\code{write_be_results} est une methode permettant d'enregistrer en \code{.cvs} les resultats
d'une evaluation best estimate.
}
\author{
Prim'Act
}
