% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action_class.R
\docType{class}
\name{Action}
\alias{Action}
\title{La classe \code{Action}.}
\description{
Classe pour les actifs de type Action.
}
\section{Slots}{

\describe{
\item{\code{ptf_action}}{est un \code{data.frame}, chaque ligne represente un actif action du portefeuille d'action.}
}}
\author{
Prim'Act
}
\seealso{
Les operations d'achat-vente action  \code{\link{buy_action}},
\code{\link{sell_action}} et \code{\link{sell_pvl_action}}.
}
\keyword{classes}

