% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutresPassifs-class.R
\docType{class}
\name{AutresPassifs}
\alias{AutresPassifs}
\title{La classe \code{AutresPassifs}.}
\description{
Une classe pour la gestion des passifs hors modele.
}
\section{Slots}{

\describe{
\item{\code{mp}}{un objet \code{data.frame} au format fige contenant les flux des passifs hors modele.}
}}
\author{
Prim'Act
}
\seealso{
La lecture des flux d'une annee \code{\link{proj_annee_autres_passifs}}.
}
\keyword{classes}

