% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutresReserves-class.R
\docType{class}
\name{AutresReserves}
\alias{AutresReserves}
\title{La classe AutreReserves}
\description{
Une classe de parametres permettant de gerer le stock de provision globale de gestion (PGG) et de
provision pour sinistres a payer (PSAP).
}
\section{Slots}{

\describe{
\item{\code{pgg_debut}}{la valeur de la PGG en debut de periode.}

\item{\code{psap_debut}}{la valeur de la PSAP en debut de periode.}

\item{\code{pgg_valeur}}{la valeur courant de la PGG.}

\item{\code{psap_valeur}}{la valeur courant de la PSAP.}

\item{\code{tx_pgg_ep}}{le taux de PGG applique sur l'epargne.}

\item{\code{tx_pgg_autres}}{le taux de PGG applique sur les autres passifs.}

\item{\code{tx_psap_ep}}{le taux de PGG applique sur l'epargne.}

\item{\code{tx_psap_autres}}{le taux de PGG applique sur les autres passifs.}
}}

\seealso{
Le calcul et la mise a jour des autres reserves \code{\link{update_reserves}} et
\code{\link{init_debut_pgg_psap}}.
}
\author{
Prim'Act
}
\keyword{classes}
