% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevaloEngine_base_prod_fin.R
\docType{methods}
\name{base_prod_fin}
\alias{base_prod_fin}
\title{Calcule la base de produits financiers attribuables.}
\usage{
base_prod_fin(tra, pm_moy, ppb)
}
\arguments{
\item{tra}{est une valeur \code{numeric} donnant le taux de rendement de l'actif.}

\item{pm_moy}{est un vecteur \code{numeric} comprenant le montant de PM moyenne par produit.}

\item{ppb}{est un objet de la classe \code{\link{Ppb}} qui renvoie l'etat courant de la PPB.}
}
\value{
La valeur de la base de produit financier par produit et au total pour le portefeuille.
}
\description{
\code{base_prod_fin} est une methode permettant de calculer la base de produits financiers attribuables
 pour la revalorisation des contrats.
}
\author{
Prim'Act
}
\seealso{
\code{\link{Ppb}}.
}

