% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_flux_annee.R
\docType{methods}
\name{calc_flux_annee}
\alias{calc_flux_annee}
\title{Calcule les flux percus dans l'annee du portefeuille obligataire.}
\usage{
calc_flux_annee(x)
}
\arguments{
\item{x}{un objet de la classe \code{\link{Oblig}}.}
}
\value{
\code{tombee_coupon} un vecteur correspondant aux tombees de coupon dans l'annee. Ce vecteur a autant d'elements
que le portefeuille obligataire d'inputs a de lignes.

\code{tombee_echeance} un vecteur correspondant aux tombees d echeances dans l'annee. Ce vecteur a autant d'elements
que le portefeuille obligataire d'inputs a de lignes.
}
\description{
\code{calc_flux_annee} est une methode permettant de calculer les tombees de coupons et les echeances
 l'ensemble des obligations composant un portefeuille obligataire.
}
\author{
Prim'Act
}
