% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevaloEngine_calc_marge_fin.R
\docType{methods}
\name{calc_marge_fin}
\alias{calc_marge_fin}
\title{Calcule la marge financiere de l'assureur.}
\usage{
calc_marge_fin(base_fin, rev_prest_nette, rev_stock_nette,
  contrib_tmg_prest, contrib_tmg_stock, contrib_ppb_tx_cible)
}
\arguments{
\item{base_fin}{est un vecteur de type \code{numeric} comprenant par produit la base de produits financiers.}

\item{rev_prest_nette}{est un vecteur de type \code{numeric} comprenant par produit
la revalorisation nette sur prestations.}

\item{rev_stock_nette}{est un vecteur de type \code{numeric} comprenant par produit
la revalorisation nette sur stock.}

\item{contrib_tmg_prest}{est une valeur \code{numeric} comprenant par produit
la contribution de la PPB au financement des TMG sur prestations.}

\item{contrib_tmg_stock}{est une valeur \code{numeric} comprenant par produit
la contribution de la PPB au financement des TMG sur stock.}

\item{contrib_ppb_tx_cible}{une valeur de type \code{numeric} comprenant par produit
la contribution de la PPB au financement du taux cible sur stock.}
}
\value{
Le montant de la marge de l'assureur.
}
\description{
\code{calc_marge_fin} est une methode permettant
 de calculer la marge financiere de l'assureur apres attribution d'un certain niveau de revalorisation.
}
\author{
Prim'Act
}
