% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_flux.R
\docType{methods}
\name{calc_nominal}
\alias{Oblig}
\alias{calc_nominal}
\title{Calcul le nominal des models points constituant le portefeuille obligataire.}
\usage{
calc_nominal(x)
}
\arguments{
\item{x}{un objet de la classe Oblig.}
}
\value{
Un vecteur dont chaque element correspond a la valeur du nominal de l'obligation consideree : parite * nominal * nb_unit.
Le vecteur renvoye a autant d'elements que le portefeuille obligataire en input a de lignes.
}
\description{
\code{calc_nominal} est une methode permettant de calculer les valeurs de nominal de l'ensemble des obligations
composant un portefeuille obligataire.
}
\author{
Prim'Act
}

