% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpEuroInd-calc_pm.R
\docType{methods}
\name{calc_pm}
\alias{calc_pm}
\title{Calcul les PM pour des contrats epargne en euros.}
\usage{
calc_pm(x, tab_prime, tab_prest, tx_cible, tx_min, an, method, tx_soc)
}
\arguments{
\item{x}{un objet de la classe \code{\link{EpEuroInd}} contenant les model points epargne euros.}

\item{tab_prime}{une liste contenant les flux de primes pour chaque ligne de model points.
Le format de cette liste correspond a la sortie \code{flux} de la methode \code{\link{calc_primes}}.}

\item{tab_prest}{est une liste contenant les flux de prestations pour chaque ligne de model points.
Le format de cette liste correspond a la sortie \code{flux} de la methode \code{\link{calc_prest}}.}

\item{tx_cible}{est une liste conteant les taux cible annuel et semestriel par model points.
Le format de cette liste correspond a la sortie de la methode \code{\link{calc_tx_cible}}.}

\item{tx_min}{une liste contenant le taux de revalorisation minimum associes a chaque ligne de model points.
Le format de cette liste correspond a la sortie de la methode \code{\link{calc_tx_min}}.}

\item{an}{une valeur \code{numeric} represantant l'annee de projection courante.}

\item{method}{un \code{character} prenant pour valeur \code{normal} pour le calcul
des flux avec application de la revalorisation au titre de la participation aux benefices,
 et la valeur \code{gar} pour le calcul avec uniquement les flux garanti (calcul de la FDB).}

\item{tx_soc}{est une valeur \code{numeric} correspondant au taux de prelevements sociaux.}
}
\value{
Une liste contenant :
\describe{
\item{\code{method} : }{la valeur de l'argument \code{method}}
\item{\code{flux} : }{une liste comprenant les flux de l'annee}
\item{\code{stock} : }{une liste comprenant les nombres de sorties}
}

Le format de la liste \code{flux} est :
\describe{
\item{\code{rev_stock_brut} : }{un vecteur contenant la revalorisation minimale
brute de l'annee appliquee au PM}
\item{\code{rev_stock_nette} : }{un vecteur contenant la revalorisation minimale
nette de l'annee appliquee au PM}
\item{\code{enc_charg_stock} : }{un vecteur contenant les chargement sur encours de l'annee,
calcules en prenant en compte la revalorisation minimale}
\item{\code{enc_charg_base_th} : }{un vecteur contenant les chargements sur encours theoriques
de l'annee, evalues sur la base de la PM non revalorisees}
\item{\code{enc_charg_rmin_th} : }{un vecteur contenant les chargements sur encours theoriques
 de l'annee, evalues sur la seule base de la revalorisation minimale des PM}
\item{\code{base_enc_th} : }{un vecteur contenant l'assiette de calcul des chargements sur encours de l'annee}
\item{\code{soc_stock} : }{un vecteur contenant le prelevements sociaux de l'annee}
\item{\code{it_tech_stock} : }{un vecteur contenant les interets techniques sur stock de l'annee}
\item{\code{it_tech} : }{un vecteur contenant les interets techniques sur stock et
sur prestations de l'annee}
\item{\code{bes_tx_cible} : }{un vecteur contenant le besoin de financement de l'annee pour
atteindre le taux cible de chaque assure.}
}

Le format de la liste \code{stock} est :
\describe{
\item{\code{pm_deb : }}{un vecteur contenant le montant de PM en debut d'annee}
\item{\code{pm_fin : }}{un vecteur contenant le montant de PM en fin d'annee, avec
revalorisation au taux minimum}
\item{\code{pm_moy : }}{un vecteur contenant le montant de PM moyenne sur l'annee.}
}
}
\description{
\code{calc_pm} est une methode permettant de calculer les provisions mathematiques (PM)
 de fin de periode avant application de la revalorisation au titre de la participation aux benefices.
}
\details{
Cette methode permet de calculer les montants de PM de fin d'annee avec une revalorisation
minimale. Les chargements sur encours sont egalement preleves. Cette methode permet de gerer les contrats a taux de
revalorisation net negatif. Cette methode permet egalement de calculer le besoin de financement necessaire
pour atteindre les exigences de revalorisation des assures.
}
\author{
Prim'Act
}
\seealso{
\code{\link{calc_primes}}, \code{\link{calc_prest}}, \code{\link{calc_tx_cible}},
\code{\link{calc_tx_min}}.
}

