% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortFin_calc_pmvl.R
\docType{methods}
\name{calc_pmvl}
\alias{calc_pmvl}
\title{Mets a jour les valeurs des plus ou moins-values actions et immobilier.}
\usage{
calc_pmvl(x)
}
\arguments{
\item{x}{objet de la classe \code{\link{PortFin}}.}
}
\value{
L'objet \code{x} dont les attributs en PVL et en MVL a ete mise a jour.
}
\description{
\code{calc_pmvl} est une methode permettant de calculer les valeurs des plus ou moins values
latentes sur actions et immobilier. Met a jour la valeur des attributs d'un objet \code{\link{PortFin}}.
}
\author{
Prim'Act
}
