% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action_calc_pmvl.R
\docType{methods}
\name{calc_pmvl_action}
\alias{Action}
\alias{calc_pmvl_action}
\title{Calcul les valeurs de marches de chaque composante du portefeuille action.}
\usage{
calc_pmvl_action(x)
}
\arguments{
\item{x}{objet de la classe \code{Action} (decrivant le portefeuille d'action).}
}
\value{
Une liste composee de deux elements \code{(pvl, mvl)} correspondant respectivement 
aux sommes des plus values latentes actions et aux somme des moins values latentes action.
}
\description{
\code{calc_pmvl_action} est une methode permettant de calculer les valeurs de marche.
}
\author{
Prim'Act
}

