% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassifBase-calc_prest.R
\docType{methods}
\name{calc_prest}
\alias{calc_prest}
\title{Calcul les flux de prestations pour des contrats epargne en euros ou retraite euros en phases de restitution.}
\usage{
calc_prest(x, method, an, y)
}
\arguments{
\item{x}{un objet de la classe \code{\link{EpEuroInd}} ou de la classe \code{\link{RetraiteEuroRest}} contenant les model points epargne euros,
ou les model points retraite euro en phase de restitution.}

\item{method}{un \code{character} prenant pour valeur \code{normal} pour le calcul des flux avec application de la revalorisation au titre de
la participation aux benefices, et la valeur \code{gar} pour le calcul avec uniquement les flux garanti (calcul de la FDB).}

\item{an}{une valeur \code{integer} correspondant a l'annee du calcul des prestations.}

\item{y}{une liste a remplir uniquement si \code{x} est de type \code{\link{EpEuroInd}} contenant les parametres :
\describe{
\item{\code{proba_dyn} : }{une liste contenant le taux de rachats dynamiques (totaux et partiels) par model points.}
\item{\code{tx_min} : }{une liste contenant le taux de revalorisation minimum associes a chaque ligne de model points.
Le format de cette liste correspond a la sortie de la methode \code{\link{calc_tx_min}}.}
\item{\code{tx_soc} : }{est une valeur \code{numeric} correspondant au taux de prelevements sociaux.}
\item{\code{choc_lapse_mass} : }{est une valeur \code{numeric} correspondant au choc de rachat massif.}
}}
}
\value{
Une liste contenant :
\describe{
\item{\code{method} : }{la valeur de l'argument \code{method}}
\item{\code{flux} : }{une liste comprenant les flux de l'annee}
\item{\code{stock} : }{une liste comprenant les nombres de sorties}
}

Le format de la liste \code{flux} est :
\describe{
\item{\code{ech} : }{un vecteur contenant les flux de sortie en echeance de l'annee : nul si l'objet est de type \code{\link{RetraiteEuroRest}}.}
\item{\code{rach_tot} : }{un vecteur contenant les flux de rachat totaux de l'annee : nul si l'objet est de type \code{\link{RetraiteEuroRest}}.}
\item{\code{dc} : }{un vecteur contenant les flux de deces de l'annee : nul si l'objet est de type \code{\link{RetraiteEuroRest}}.}
\item{\code{rach_part} : }{un vecteur contenant les flux de rachat partiel de l'annee : nul si l'objet est de type
\code{\link{RetraiteEuroRest}}.}
\item{\code{rente} : }{le flux annuel de rente par model point : nul si l'objet est de type \code{\link{EpEuroInd}}.}
\item{\code{prest} : }{un vecteur contenant les flux prestations de l'annee (renseigne que l'objet x soit de type \code{\link{RetraiteEuroRest}}
ou \code{\link{EpEuroInd}}).}
\item{\code{rev_ech} : }{un vecteur contenant la revalorisation des echeances de l'annee : nul si l'objet est de type
\code{\link{RetraiteEuroRest}}.}
\item{\code{rev_rach_tot} : }{un vecteur contenant la revalorisation des rachats totaux de l'annee : nul si l'objet est de type
\code{\link{RetraiteEuroRest}}.}
\item{\code{rev_dc} : }{un vecteur contenant la revalorisation des deces de l'annee : nul si l'objet est de type \code{\link{RetraiteEuroRest}}.}
\item{\code{rev_rach_part} : }{un vecteur contenant la revalorisation des rachats partiels de l'annee : nul si l'objet est de type
\code{\link{RetraiteEuroRest}}.}
\item{\code{rev_prest} : }{un vecteur contenant la revalorisation brute des prestations de l'annee : nul si l'objet est de type
\code{\link{RetraiteEuroRest}}.}
\item{\code{rev_prest_nette} : }{un vecteur contenant la revalorisation des prestations nette de l'annee : nul si l'objet est de type
\code{\link{RetraiteEuroRest}}.}
\item{\code{enc_charg} : }{un vecteur contenant les chargements sur l'encours de l'annee : nul si l'objet est de type
\code{\link{RetraiteEuroRest}}.}
\item{\code{rach_charg} : }{un vecteur contenant les chargements sur les rachats de l'annee : nul si l'objet est de type
\code{\link{RetraiteEuroRest}}.}
\item{\code{soc_prest} : }{un vecteur contenant les prelevements sociaux sur prestations de l'annee : nul si l'objet est de type
\code{\link{RetraiteEuroRest}}.}
\item{\code{it_tech_prest} : }{un vecteur contenant les interets techniques sur prestations de l'annee. : nul si l'objet est de type
\code{\link{RetraiteEuroRest}}.}
\item{\code{arr_charg} : }{un vecteur contenant les chargements sur arrerages. : nul si l'objet est de type \code{\link{EpEuroInd}}.}
}

Le format de la liste \code{stock} est :
\describe{
\item{\code{nb_ech : }}{un vecteur contenant le nombre de sorties en echeance de l'annee : nul si l'objet est de type
\code{\link{RetraiteEuroRest}}.}
\item{\code{nb_rach_tot : }}{un vecteur contenant le nombre de rachats totaux de l'annee : nul si l'objet est de type
\code{\link{RetraiteEuroRest}}.}
\item{\code{nb_dc : }}{un vecteur contenant le nombre de deces de l'annee}
\item{\code{nb_sortie : }}{un vecteur contenant le nombre de sorties de l'annee}
\item{\code{nb_contr_fin : }}{un vecteur contenant le nombre de contrats en cours en fin d'annee}
\item{\code{nb_contr_moy : }}{un vecteur contenant la moyenne du nombre de contrats sur l'annee.}
}
}
\description{
\code{calc_prest} est une methode permettant de calculer les flux de prestations,
les chargements sur encours relatifs a ces prestations et les nombres de sorties sur une periode.
}
\details{
En epargne, cette methode permet de calculer les flux de sortie en echeance, les flux de rachat totaux et partiels et
les flux de deces d'un contrat epargne en euros. Ces prestations font l'objet d'une relavorisation
au taux minimum contractuel. Les nombres de sortie sont egalement produits. Il est possible de realiser un choc de rachat
massif si \code{an}. Dans ce cas, les prestations de rachats massifs sortent en debut d'annee et ne sont pas revalorisees.
Des chargements sont appliques sur flux de rachats. Des prelevements sur encours sont appliques sur les
prestations revalorises au taux minimum contractuel. Cette methode permet de gerer les contrats a taux de
revalorisation net negatif.
Pour un contrat de retraite, elle permet de sortir les flux de rente et
 les nombres de sorties.
}
\seealso{
\code{\link{calc_proba_flux}}, \code{\link{calc_tx_min}}.
}
\author{
Prim'Act
}
