% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpEuroInd-calc_prest.R
\docType{methods}
\name{calc_prest}
\alias{EpEuroInd}
\alias{calc_prest}
\title{Calcul les flux de prestations pour des contrats epargne en euros.}
\usage{
calc_prest(x, tx_sortie, tx_min, an, method, tx_soc)
}
\arguments{
\item{x}{un objet de la classe \code{\link{EpEuroInd}} contenant les model points epargne euros.}

\item{tx_sortie}{une matrice contenant les taux de sortie associes a chaque ligne de model points.
Le format de cette matrice correspond a la sortie de la methode \code{\link{calc_tx_sortie}}.}

\item{tx_min}{une liste contenant le taux de revalorisation minimum associes a chaque ligne de model points.
Le format de cette liste correspond a la sortie de la methode \code{\link{calc_tx_min}}.}

\item{an}{une valeur \code{numeric} represantant l'annee de projection courante.}

\item{method}{un \code{character} prenant pour valeur \code{normal} pour le calcul
des flux avec application de la revalorisation au titre de la participation aux benefices,
 et la valeur \code{gar} pour le calcul avec uniquement les flux garanti (calcul de la FDB).}

\item{tx_soc}{est une valeur \code{numeric} correspondant au taux de prelevements sociaux.}
}
\value{
Une liste contenant :
\describe{
\item{\code{method} : }{la valeur de l'argument \code{method}}
\item{\code{flux} : }{une liste comprenant les flux de l'annee}
\item{\code{stock} : }{une liste comprenant les nombres de sorties}
}

Le format de la liste \code{flux} est :
\describe{
\item{\code{ech} : }{un vecteur contenant les flux de sortie en echeance de l'annee}
\item{\code{rach_tot} : }{un vecteur contenant les flux de rachat totaux de l'annee}
\item{\code{dc} : }{un vecteur contenant les flux de deces de l'annee}
\item{\code{rach_part} : }{un vecteur contenant les flux de rachat partiel de l'annee}
\item{\code{prest} : }{un vecteur contenant les flux prestations de l'annee}
\item{\code{rev_ech} : }{un vecteur contenant la revalorisation des echeances de l'annee}
\item{\code{rev_rach_tot} : }{un vecteur contenant la revalorisation des rachats totaux de l'annee}
\item{\code{rev_dc} : }{un vecteur contenant la revalorisation des deces de l'annee}
\item{\code{rev_rach_part} : }{un vecteur contenant la revalorisation des rachats partiels de l'annee}
\item{\code{rev_prest} : }{un vecteur contenant la revalorisation brute des prestations de l'annee}
\item{\code{rev_prest_nette} : }{un vecteur contenant la revalorisation des prestations nette de l'annee}
\item{\code{enc_charg} : }{un vecteur contenant les chargements sur l'encours de l'annee}
\item{\code{rach_charg} : }{un vecteur contenant les chargements sur les rachats de l'annee}
\item{\code{soc_prest} : }{un vecteur contenant les prelevements sociaux sur prestations de l'annee}
\item{\code{it_tech_prest} : }{un vecteur contenant les interets techniques sur prestations de l'annee.}
}

Le format de la liste \code{stock} est :
\describe{
\item{\code{nb_ech : }}{un vecteur contenant le nombre de sorties en echeance de l'annee}
\item{\code{nb_rach_tot : }}{un vecteur contenant le nombre de rachats totaux de l'annee}
\item{\code{nb_dc : }}{un vecteur contenant le nombre de deces de l'annee}
\item{\code{nb_sortie : }}{un vecteur contenant le nombre de sorties de l'annee}
\item{\code{nb_contr_fin : }}{un vecteur contenant le nombre de contrats en cours en fin d'annee}
\item{\code{nb_contr_moy : }}{un vecteur contenant la moyenne du nombre de contrats sur l'annee.}
}
}
\description{
\code{calc_prest} est une methode permettant de calculer les flux de prestations,
les chargements sur encours relatifs a ces prestations et les nombres de sorties sur une periode.
}
\details{
Cette methode permet de calculer les flux de sortie en echeance, les flux de rachat totaux et partiels et
les flux de deces d'un contrat epargne en euros. Ces prestations font l'objet d'une relavorisation
au taux minimum contractuel. Les nombres de sortie sont egalement produits.
Des chargements sont appliques sur flux de rachats. Des prelevements sur encours sont appliques sur les
prestations revalorises au taux minimum contractuel. Cette methode permet de gerer les contrats a taux de
revalorisation net negatif.
}
\author{
Prim'Act
}
\seealso{
\code{\link{calc_tx_sortie}}, \code{\link{calc_tx_min}}.
}

