% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassifBase-calc_primes.R
\docType{methods}
\name{calc_primes}
\alias{calc_primes}
\alias{EpEuroInd}
\title{Calcule les flux de primes pour des contrats epargne en euros et des retraites en phase de restitution.}
\usage{
calc_primes(x)
}
\arguments{
\item{x}{un objet de la classe \code{\link{EpEuroInd}} ou de la classe \code{\link{RetraiteEuroRest}}
contenant les model points epargne euros ou de retraite en phase de restitution.}
}
\value{
\code{stock} : une liste contenent le nombre de versements \code{nb_vers} associe a chaque model point.

\code{flux} : une liste contenant pour chaque model point les montants de primes brutes \code{pri_brut},
les montants de primes nettes \code{pri_net} et les chargements sur primes \code{pri_chgt}.
}
\description{
\code{calc_primes} est une methode permettant de calculer les flux de primes sur
 une periode.
}
\details{
Cette fonction permet de projeter uniquement des primes periodiques de contrats epargne en euros. Pour
la retraite en phase de restitution, il n'y a pas de prime et la methode renvoie des valeurs nulles.
}
\author{
Prim'Act
}
