% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamTableMort-calc_qx.R
\docType{methods}
\name{calc_qx}
\alias{calc_qx}
\title{Calcule le taux de deces.}
\usage{
calc_qx(table_mort, age, gen)
}
\arguments{
\item{table_mort}{un objet de la classe \code{\link{ParamTableMort}} contenant la table de mortalite.}

\item{age}{une valeur \code{numeric} correspondant a l'age.}

\item{gen}{une valeur \code{numeric} correspondant a la generation.}
}
\value{
La valeur du taux de deces calcule.
}
\description{
\code{calc_qx} est une methode permettant de calculer le taux de deces.
}
\author{
Prim'Act
}

