% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamTableRach-calc_rach.R
\docType{methods}
\name{calc_rach}
\alias{calc_rach}
\title{Calcule le taux de rachat.}
\usage{
calc_rach(table_rach, age, anc)
}
\arguments{
\item{table_rach}{un objet de la classe \code{\link{ParamTableRach}} contenant la table de rachat.}

\item{age}{une valeur \code{numeric} correspondant a l'age.}

\item{anc}{une valeur \code{numeric} correspondant a l'anciennete.}
}
\value{
La valeur du taux de rachat calcule.
}
\description{
\code{calc_rach} est une methode permettant de calculer le taux de rachat.
}
\author{
Prim'Act
}

