% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamRachDyn-calc_rach_dyn.R
\docType{methods}
\name{calc_rach_dyn}
\alias{calc_rach_dyn}
\title{Calcule la composante rachats dynamique.}
\usage{
calc_rach_dyn(p, tx_cible, tx_serv)
}
\arguments{
\item{p}{un objet de la classe \code{\link{ParamRachDyn}} contenant les parametres de rachats dynamiques.}

\item{tx_cible}{une valeur \code{numeric} correspondant au taux de revalorisation cible.}

\item{tx_serv}{une valeur \code{numeric} correspondant au taux de revalorisation servi.}
}
\value{
La valeur du taux rachat.
}
\description{
\code{calc_rach_dyn} est une methode permettant de calculer la composante rachat dynamique
selon la methodologie transmise dans le ONC de l'ACPR de 2013.
}
\author{
Prim'Act
}
