% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC_calc.R
\docType{methods}
\name{calc_RC}
\alias{calc_RC}
\title{Calcule de la RC.}
\usage{
calc_RC(x, pmvr_oblig)
}
\arguments{
\item{x}{objet de la classe \code{\link{RC}}, necessaire pour connaitre le stock de RC initial.}

\item{pmr_oblig}{est un \code{numeric} correspondant au montant global annuel de plus ou moins values realisees sur des actifs obligataires.}
}
\value{
Le format de la liste renvoyee est :
\describe{
\item{\code{RC_courante} : }{ valeur de la RC courante initiale augmentee des plus ou moins values annuelles realisees}
\item{\code{var_RC} : }{ variation de la RC courante.}
}
}
\description{
\code{calc_RC} est une methode permettant de calculer le montant de Reserve de Capitalisation.
}
\author{
Prim'Act
}

