% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevaloEngine_calc_revalo.R
\docType{methods}
\name{calc_revalo}
\alias{RevaloEngine}
\alias{calc_revalo}
\title{Applique la politique de revalorisation d'un canton.}
\usage{
calc_revalo(x, passif_av_pb, tra, plac_moy_vnc, result_tech)
}
\arguments{
\item{x}{un objet de la classe \code{\link{Canton}}.}

\item{passif_av_pb}{est une liste produit par la methode \code{viellissement_av_pb}
appliquee a un portefeuille de passif.}

\item{tra}{est la valeur \code{numeric} du taux de rendement de l'actif.}

\item{plac_moy_vnc}{est la valeur \code{numeric} moyenne des actifs.}

\item{result_tech}{est la valeur \code{numeric} du resultats techniques prise en compte de la PB.}
}
\value{
Une liste avec la valeur de la revalorisation additionelle nette servie par
 produit, la ppb mise a jour, la PPB mise a jour et le resultat technique avant PB.
}
\description{
\code{calc_revalo} est une methode permettant de
 d'appliquer l'ensemble de la politique de revalorisation d'un assureur.
}
\author{
Prim'Act
}

