% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpEuroInd-calc_revalo_pm.R
\docType{methods}
\name{calc_revalo_pm}
\alias{EpEuroInd}
\alias{calc_revalo_pm}
\title{Calcul la revalo pour des PM d'un model point epargne en euros.}
\usage{
calc_revalo_pm(x, rev_net_alloue, tx_soc)
}
\arguments{
\item{x}{un objet de la classe \code{EpEuroInd} contenant les model points epargne euros.}

\item{rev_net_alloue}{une valeur de type \code{numeric} correspondant au montant de revalorisation a allouer}

\item{tx_soc}{le taux de charges sociales}

\item{rev_net_alloue}{une valeur de type \code{numeric} correspondant au montant de revalorisation a allouer}
}
\description{
\code{calc_revalo_pm} est une methode permettant de calculer les PM sur une annee.
}
\author{
Prim'Act
}

