% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_calc_value.R
\docType{methods}
\name{calc_sur_dec}
\alias{Oblig}
\alias{calc_sur_dec}
\title{Calcul les surcote/decote de chaque composante d'un portefeuille obligataire.}
\usage{
calc_sur_dec_vnc(x)
}
\arguments{
\item{x}{objet de la classe Oblig (decrivant le portefeuile obligataire).}
}
\value{
Un data.frame composé de deux colonnes : 1 ere colonne :surcotes decotes ; 2de colonne : valeurs nettes comptables.
}
\description{
\code{calc_sur_dec} est une methode permettant de calculer les surcotes/decotes de chaque composante d'un portefeuille obligataire.
}
\author{
Prim'Act
}

