% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpEuroInd-calc_tx_cible.R
\docType{methods}
\name{calc_tx_cible}
\alias{EpEuroInd}
\alias{calc_tx_cible}
\title{Calcul du taux cible pour des contrats epargne en euros.}
\usage{
calc_tx_cible(x, ht, list_rd)
}
\arguments{
\item{x}{un objet de la classe \code{\link{EpEuroInd}} contenant les model points epargne euros.}

\item{ht}{un objet de la classe \code{\link{HypTech}} contenant differentes lois de comportement.}

\item{list_rd}{une liste contenant les rendements de reference. Le format de cette liste est :
\describe{
\item{le taux de rendement obligataire}{}
\item{le taux de rendement de l'indice action de reference}{}
\item{le taux de rendement de l'indice immobilier de reference}{}
\item{le taux de rendement de l'indice tresorerie de reference}{}
}}
}
\value{
\code{tx_cible_an} : un vecteur contenant les taux cible de l'annee

\code{tx_cible_se} : un vecteur contenant les taux cible de l'annee sur base semestrielle
}
\description{
\code{calc_tx_cible} est une methode permettant d'evaluer le taux de revalorisation cible
  de chaque model point.
}
\note{
Pour les besoins des calculs a mi-annee, des taux semestriels sont produits.
}
\author{
Prim'Act
}
\seealso{
La recuperation des taux cibles calcules : \code{\link{get_comport}}.
}

