% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassifBase-calc_tx_cible.R
\docType{methods}
\name{calc_tx_cible}
\alias{calc_tx_cible}
\title{Calcul du taux cible pour des contrats epargne en euros et de retraite en phase de restitution.}
\usage{
calc_tx_cible(x, y)
}
\arguments{
\item{x}{un objet de la classe \code{\link{EpEuroInd}} ou de la classe \code{\link{RetraiteEuroRest}}
contenant les model points epargne euros ou retraite euro en phases de restitution.}

\item{y}{une liste contenant les parametres.
\describe{
\item{\code{list_rd} : }{est un vecteur contenant les rendements de reference.}
\item{\code{ht} : }{est un objet de la classe \code{\link{HypTech}}.}
}
Le format de \code{list_rd} est :
\describe{
\item{le taux de rendement obligataire}{}
\item{le taux de rendement de l'indice action de reference}{}
\item{le taux de rendement de l'indice immobilier de reference}{}
\item{le taux de rendement de l'indice tresorerie de reference}{}
}}
}
\value{
\code{tx_cible_an} : un vecteur contenant les taux cible de l'annee

\code{tx_cible_se} : un vecteur contenant les taux cible de l'annee sur base semestrielle
}
\description{
\code{calc_tx_cible} est une methode permettant d'evaluer le taux de revalorisation cible
  de chaque model point.
}
\note{
Pour les besoins des calculs a mi-annee, des taux semestriels sont produits.
}
\seealso{
La recuperation des taux cibles calcules : \code{\link{get_comport}}.
}
\author{
Prim'Act
}
