% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassifBase-calc_tx_min.R
\docType{methods}
\name{calc_tx_min}
\alias{calc_tx_min}
\title{Calcul le taux de revalorisation contractuel minimum pour des contrats epargne en euros.}
\usage{
calc_tx_min(x, an)
}
\arguments{
\item{x}{un objet de la classe \code{\link{EpEuroInd}} contenant les model points epargne euros.}

\item{an}{un \code{numeric} representant l'annee de projection courante.}
}
\value{
\code{tx_tech_an} : un vecteur contenant les taux de technique de l'annee

\code{tx_tech_se} : un vecteur contenant les taux de technique de l'annee sur base semestrielle

\code{tx_an} : un vecteur contenant les taux de revalorisation minimum de l'annee

\code{x_se} : un vecteur contenant les taux de revalorisation minimum de l'annee exprimes en semestriel.
}
\description{
\code{calc_tx_min} est une methode permettant de calculer les taux de revalorisation minimum
  sur une periode. La revalorisation minimum est le maximum entre le taux technique et
  le taux minimim garanti (TMG) du contrat.
}
\note{
Pour les besoins des calculs a mi-annee, des taux semestriels sont produits.
}
\author{
Prim'Act
}
