% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immo_calc_vm.R
\docType{methods}
\name{calc_vm_immo}
\alias{Immo}
\alias{calc_vm_immo}
\title{Calcul les valeurs de marches de chaque composante du portefeuille immobilier.}
\usage{
calc_vm_immo(x, rdt)
}
\arguments{
\item{x}{objet de la classe \code{Immo} (decrivant le portefeuille d'immobilier).}

\item{rdt}{vecteur de type \code{numeric} decrivant le rendement de chacune des lignes d'immobilier du portefeuille immobilier de l'assureur.
Contient autant d'elements que le portefeuille a de lignes.}
}
\value{
L'objet \code{x} dont les valeurs de marche ont ete mises a jour.
}
\description{
\code{calc_vm_immo} est une methode permettant de calculer les valeurs de marche.
}
\author{
Prim'Act
}

