% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immo_calc_vm.R
\docType{methods}
\name{calc_vm_immo}
\alias{calc_vm_immo}
\title{Calcule les valeurs de marches du portefeuille immobilier.}
\usage{
calc_vm_immo(x, rdt)
}
\arguments{
\item{x}{objet de la classe \code{\link{Immo}} (decrivant le portefeuille d'immobilier).}

\item{rdt}{vecteur de type \code{numeric} decrivant le rendement de chacune des lignes du portefeuille immobilier
 de l'assureur.
Contient autant d'elements que le portefeuille immobilier a de lignes.}
}
\value{
L'objet \code{x} dont les valeurs de marche ont ete mises a jour.
}
\description{
\code{calc_vm_immo} est une methode permettant de calculer les valeurs de marche du portefeuille immobilier.
}
\author{
Prim'Act
}
