% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Canton_class.R, R/ParamAlmEngine_class.R
\docType{class}
\name{Canton}
\alias{Canton}
\title{La classe Canton}
\description{
Classe pour le canton d'un assureur

Classe pour le parametre ALM d'un canton.
}
\section{Slots}{

\describe{
\item{\code{annee}}{est une valeur \code{integer} correspondant a l'annee de projection.}

\item{\code{ptf_fin}}{est un objet de type \code{PortFin}, qui represente le portefeuille d'investissement d'un canton.}

\item{\code{ptf_passif}}{est un objet de type \code{PortPassif}, qui represente le portefeuille de passif d'un canton.}

\item{\code{mp_esg}}{est un objet de type \code{ModelPointESG}, qui represente la situation courante
en annee et simulations des valeurs de l'ESG.}

\item{\code{ppb}}{est un objet de type \code{Ppb}, qui represente la provision pour participation aux benefices.}

\item{\code{hyp_canton}}{est un objet de type \code{HypCanton}, qui regroupe les hypotheses generales applicables au canton.}

\item{\code{param_alm}}{est un objet de type \code{ParamAlmEngine}, qui contient les parametres
utilises dans les methodes de gestion de l'allocation d'actifs.}

\item{\code{param_revalo}}{est un objet de type \code{ParamRevaloEngine}, qui contient les parametres
utilises dans les methodes de gestion de la revalorisation.}

\item{\code{ptf_reference}}{est un objet de type \code{PortFin}, qui represente le portefeuille
d'investissement de reference d'un canton.}
}}
\section{Lien a creer}{



}
\author{
Prim'Act

Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe

Mettre le lien vers les methodes de la classe
}
\keyword{classes}

