% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChocSolvabilite2_load.R
\docType{methods}
\name{chargement_choc}
\alias{chargement_choc}
\title{Permet de charger les parametres de choc de la formule standard.}
\usage{
chargement_choc(x, folder_chocs_address)
}
\arguments{
\item{x}{objet de la classe \code{\link{ChocSolvabilite2}}.}

\item{folder_choc_address}{est un \code{character}. Cette chaine de caractere est construite par la methode
 \code{\link{set_architecture}} de la classe \code{\link{Initialisation}}.
Elle contient l'adresse du dossier contenant les fichiers de parametres des chocs de la formule standard a appliquer.
Ces derniers doivent etre renseignes par l'utilisateur.}
}
\value{
\code{x} l'objet  de la classe \\code{\link{ChocSolvabilite2}} dont les attributs \code{param_choc_mket}
 et \code{param_choc_sousc} ont ete mis a jour.
}
\description{
\code{chargement_choc} est une methode permettant de charger les parametres
l'ensemble des parametres necessaires a la bonne application des chocs de marche et
de souscription au sens de la formule standard de la directive Solvabilite 2,
 tels que renseignes par l'utilisateur.
}
\author{
Prim'Act
}
\seealso{
La creation de l'architecture de chargement des donnees et parametres renseignes par l'utilisateur
 \code{\link{set_architecture}},
ainsi que les classes \code{\link{ParamChocMket}} et \code{\link{ParamChocSousc}}.
}

