% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChocSolvabilite2_do_choc_spread.R
\docType{methods}
\name{do_choc_spread}
\alias{ChocSolvabilite2}
\alias{do_choc_spread}
\title{Permet a partir d'un canton initial de creer un canton choque action.}
\usage{
do_choc_spread(x, canton)
}
\arguments{
\item{x}{objet de la classe \code{ChocSolvabilite2}.}

\item{canton}{est un objet de la classe \code{canton}. Il correspond au canton non choque (i.e. central) de l'assureur.}
}
\value{
\code{canton} l'objet  de la classe \code{canton} correspondant au scenario choque spread au sens de la formule standard Solvabilite 2.
}
\description{
\code{do_choc_spread} est une methode permettant d'appliquer le choc spread de la formule standard Solvabilite 2 a un canton.
}
\note{
Il est possible d'appliquer des chocs de spreads distincts a chaque ligne du portefeuille obligataire selon l'index de la ligne obligataire.
Cette parametrisation est effectuee dans les fichiers d'inputs utilisateurs.
}
\author{
Prim'Act
}

