% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChocSolvabilite2_do_choc_spread_unitaire.R
\docType{methods}
\name{do_choc_spread_unitaire}
\alias{do_choc_spread_unitaire}
\title{Applique le choc spread de la formule standard Solvabilite 2 a une ligne obligataire.}
\usage{
do_choc_spread_unitaire(table_choc_spread, ligne_oblig)
}
\arguments{
\item{table_choc_spread}{un \code{data.frame} contenant la table de parametres avec les chocs de spreads.}

\item{ligne_oblig}{un \code{data.frame}. Il correspond a une ligne obligataire d'un portefeuille \code{\link{Oblig}}
d'un assureur.}
}
\value{
\code{vm_choquee} une valeur \code{numeric} correspondant a la valeur de marche de la ligne obligataire
 suite a l'application du choc de spread a cette ligne.
}
\description{
\code{do_choc_spread_unitaire} Permet a partir d'une table contenant les elements du choc de spread obligataire
 Solvabilite 2 et d'une ligne obligataire d'un element \code{\link{Oblig}} d'un portefeuille financier
 \code{\link{PortFin}}
d'appliquer le choc de spread a cette ligne obligataire.
}
\author{
Prim'Act
}
\seealso{
La classe \code{\link{PortFin}}.
}

