% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortFin_do_update_vm_vnc_precedent.R
\docType{methods}
\name{do_update_vm_vnc_precedent}
\alias{PortFin}
\alias{do_update_vm_vnc_precedent}
\title{Evalue et met a jour les objets constituants un PortFin.}
\usage{
do_update_vm_vnc_precedent(x)
}
\arguments{
\item{x}{objet de la classe \code{PortFin}, correspondant au portefeuille financier de l'assureur avant mise a jour de l'attribut vm_vnc_precedent.}
}
\value{
L'objet de la classe \code{PortFin} renvoye correspond au portefeuille financier de l'assureur dont l'attribut vm_vnc_precedent a ete mis a jour.
}
\description{
\code{do_update_vm_vnc_precedent} est une methode permettant de calculer et mettre a jour un portefeuille financier
suite a un vieillissement.
}
\author{
Prim'Act
}

