% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_calc_dur.R
\docType{methods}
\name{duration_sensi}
\alias{duration_sensi}
\title{Calcule la duration de chaque composante d'un portefeuille obligataire.}
\usage{
duration_sensi(x)
}
\arguments{
\item{x}{objet de la classe \code{\link{Oblig}} (decrivant le portefeuille obligataire).}
}
\value{
Un data frame compose de deux colonnes : la premiere est composee de la duration de chacune des obligations du portefeuille obligataire.
La seconde est compose de la sensibilite de chacune des obligations du portefeuille obligataire.
Le dataframe de sortie a autant d'elements que le portefeuille obligataire a de lignes.
}
\description{
\code{duration_sensi} est une methode permettant de calculer la duration de chaque composante d'un portefeuille obligataire.
}
\author{
Prim'Act
}
