% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_echeancier.R
\docType{methods}
\name{echeancier}
\alias{Oblig}
\alias{echeancier}
\title{Calcule les flux d'un model point ou d'un ensemble de models points obligataires.}
\usage{
echeancier(coupon, maturite, zspread, nominal, yield)
}
\arguments{
\item{coupon}{vecteur contenant les taux de coupons}

\item{maturite}{vecteur d'entiers contenant les maturites residuelles}

\item{zspread}{vecteur contenant les zero-spreads}

\item{nominal}{vecteur contenant les valeurs nominales de chaque obligation}

\item{yield}{vecteur contenant la courbe de taux consideree (peut-etre vide)}
}
\value{
Une matrice contenant :
\describe{
\item{\code{grid_flux} : }{la matrice d'ecoulement des flux. Cette matrice a autant de colonnes
que le max du vecteur de maturite residuelle, et autant de lignes que les vecteurs d'input \code{coupon,maturite,zspread,nominal}.
Chaque ligne decrit les flux annuels a venir pour l'actif obligataire de caracteristique renseigne en input.}
}
}
\description{
\code{echeancier} est une methode permettant de calculer les flux jusqu'a maturite residuelle.
}
\author{
Prim'Act
}

