% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESG_extract_mp.R
\docType{methods}
\name{extract_ESG}
\alias{extract_ESG}
\title{permet de construire et charger les trajectoires simulees par le Generateur de Scenarios Economiques de Prim'Act.}
\usage{
extract_ESG(x, num_trajectoire, annee)
}
\arguments{
\item{x}{un objet de la classe \code{\link{ESG}}.}

\item{num_trajectoire}{une valeur de type \code{integer} correspondant a la trajectoire de simulation
dont on souhaite obtenir les valeurs.}

\item{annee}{une valeur de type \code{integer} correspondant a l'annee d'interet pour le model point
(possibilite de selectionner les annees 0 a \code{nb_annee_proj}).}
}
\value{
\code{x} l'objet de la classe \code{\link{ModelPoint_ESG}} construit.
}
\description{
\code{extract_ESG} construit l'objet de classe \code{\link{ModelPoint_ESG}} a partir d'un objet
 de la classe \code{\link{ESG}}.
Le \code{ModelPointESG} ainsi construit correspond a l'extraction de donnees de l'ESG
 pour une annee specifique et pour une simulation specifique.
}
\author{
Prim'Act
}
\seealso{
La classe \code{\link{ModelPoint_ESG}}.
}

