% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevaloEngine_finance_cible_marge.R, R/RevaloEngine_finance_cible_ppb.R
\docType{methods}
\name{finance_cible_ppb}
\alias{RevaloEngine}
\alias{finance_cible_ppb}
\title{Evalue le financement d'un taux cible par la PPB.}
\usage{
finance_cible_marge(marge_fin, bes_cible, rev_stock_nette, marge_min)

finance_cible_ppb(bes_cible, rev_stock_nette, ppb)
}
\arguments{
\item{marge_fin}{est valeur \code{numeric} donnant le montant courant de la marge financiere de l'assureur.}

\item{bes_cible}{est un vecteur de type \code{numeric} comprenant par produit
le besoin de financement aux taux cible.}

\item{rev_stock_nette}{est un vecteur de type \code{numeric} comprenant par produit
le montant de revalorisation nette atteint.}

\item{marge_min}{est une valeur de type \code{numeric} correspondant
au montant minimum de marge financiere souhaite par l'assureur.}

\item{ppb}{est un objet de la classe \code{Ppb} qui renvoie l'etat courant de la PPB.}

\item{bes_cible}{est un vecteur de type \code{numeric} comprenant par produit le besoin de financement aux taux cible.}

\item{rev_stock_nette}{est un vecteur de type \code{numeric} comprenant par produit le montant de revalorisation nette atteint.}
}
\value{
Une liste avec la valeur de la revalorisation nette servie par produit
 et le montant de marge de l'assureur.

Une liste avec la valeur de la revalorisation nette servie par produit, les montants de dotation
et reprise sur PPB et la PPB mise à jour.
}
\description{
\code{finance_cible_ppb} est une methode permettant de
 determiner le financement de taux cibles par la provision pour participation aux benefices.

\code{finance_cible_ppb} est une methode permettant de
 determiner le financement de taux cibles par la provision pour participation aux benefices.
}
\details{
Lorsque la revalorisation nette est superieure au besoin de financement des taux cibles, on sert le taux cible
et on dote le reste a la PPB dans la limite du plafond de dotation annuel. A l'inverse, les taux cible sont finances par
les compensations entre produits lorsque certains prevoient une revalorisation superieure au taux cible, et
par une reprise sur PPB.
}
\author{
Prim'Act

Prim'Act
}

