% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevaloEngine_finance_tmg.R
\docType{methods}
\name{finance_tmg}
\alias{RevaloEngine}
\alias{finance_tmg}
\title{Calcule la contribution de la PPB au financement des taux minimums garantis.}
\usage{
finance_tmg(bes_tmg_prest, bes_tmg_stock, ppb)
}
\arguments{
\item{bes_tmg_prest}{un vecteur \code{numeric} comprenant
par produit le besoin de finance des TMG sur prestations.
@param bes_tmg_stock un vecteur \code{numeric} comprenant
par produit le besoin de finance des TMG sur le stock.}

\item{ppb}{est un objet de la classe \code{\link{Ppb}} qui renvoie l'etat courant de la PPB.}
}
\value{
\code{contrib_tmg_prest} la valeur de la contribution au financement des TMG sur prestations.

\code{contrib_tmg_stock} la valeur de la contribution au financement des TMG sur stock.

\code{ppb} l'objet \code{ppb} mis a jour.
}
\description{
\code{finance_tmg} est une methode permettant d'evaluer le contribution de la PPB
au financement des taux minimums garantis (TMG) sur prestations et sur stock.
}
\details{
Dans cette methode, il est considere que le PPB peut venir financer les TMG sur prestations et sur stock.
Par convention, la PPB finance d'abord les TMG sur prestations, puis sur stock.
}
\author{
Prim'Act
}

