% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FraisPassif-class.R
\docType{class}
\name{FraisPassif}
\alias{FraisPassif}
\title{La classe \code{FraisPassif}}
\description{
Une classe de parametres pour les frais des produits du portefeuille de passif.
}
\section{Slots}{

\describe{
\item{\code{mp}}{un objet \code{data.frame} contenant les parametres de frais au passif par produit.}
}}

\seealso{
Le calcul des frais de passif \code{\link{calc_frais}}.
}
\author{
Prim'Act
}
\keyword{classes}
