% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypTech-get_choc_rach.R
\docType{methods}
\name{get_choc_rach}
\alias{HypTech}
\alias{get_choc_rach}
\title{donne le résulat d'un choc de rachat sur toutes les tables de rachat d'un objet HypTech}
\usage{
get_choc_rach(x, type_choc_rach, choc, choc_lim)
}
\description{
\code{get_choc_rach}
}
\author{
Prim'Act
}

