% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypTech-get_choc_table.R
\docType{methods}
\name{get_choc_table}
\alias{get_choc_table}
\title{Applique les chocs de mortalite et de longevite de la formule standard.}
\usage{
get_choc_table(x, choc)
}
\arguments{
\item{x}{un objet de la classe \code{\link{HypTech}} contenant differentes tables de
mortalite.}

\item{choc}{une valeur \code{numeric} indiquant le taux de choc.}
}
\value{
L'objet \code{x} apres choc.
}
\description{
\code{get_choc_table} est une methode permettant d'appliquer a l'ensemble des table de mortalite
d'un objet \code{\link{HypTech}} les chocs de mortalite ou de longevite de la formule standard.
}
\author{
Prim'Act
}
