% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypTech-get_qx_mort.R
\docType{methods}
\name{get_qx_mort}
\alias{HypTech}
\alias{get_qx_mort}
\title{Calcule le taux de mortalite}
\usage{
get_qx_mort(x, nom_table, age, gen)
}
\arguments{
\item{x}{un objet de la classe \code{HypTech}.}

\item{nom_table}{: nom de la table de mortalite consideree}

\item{age}{: age en numerique}

\item{gen}{: generation en numerique}
}
\value{
le taux de mortalite
}
\description{
\code{get_qx_mort} est une methode permettant de calculer le taux de mortalite.
}
\author{
Prim'Act
}

