% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypTech-get_qx_rach.R
\docType{methods}
\name{get_qx_rach}
\alias{HypTech}
\alias{get_qx_rach}
\title{Calcule le taux de rachat total ou partiel}
\usage{
get_qx_rach(x, nom_table, age, anc)
}
\arguments{
\item{x}{un objet de la classe \code{HypTech}.}

\item{nom_table}{: nom de la table de rachat consideree}

\item{age}{: age en numerique}

\item{anc}{: anciennete en numerique}
}
\value{
le taux de rachat total ou partiel
}
\description{
\code{get_qx_rach} est une methode permettant de calculer le taux de rachat total ou partiel.
}
\author{
Prim'Act
}

