% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypTech-get_rach_dyn.R
\docType{methods}
\name{get_rach_dyn}
\alias{get_rach_dyn}
\title{Recuperer les taux de rachat dynamiques calcules.}
\usage{
get_rach_dyn(x, nom_table, tx_cible, tx_serv)
}
\arguments{
\item{x}{un objet de la classe \code{\link{HypTech}}.}

\item{nom_table}{un nom de jeu de paramatre de rachat dynamique.}

\item{tx_cible}{est une valeur \code{numeric} correspondant taux de revalorisation cible.}

\item{tx_serv}{est une valeur \code{numeric} correspondant taux de revalorisation servi.}
}
\value{
Le taux de rachat dynamique.
}
\description{
\code{get_rach_dyn} est une methode permettant d'executer le calcul des taux de rachat dynamique.
}
\seealso{
Le calcul du taux de rachat dynamique \code{\link{calc_rach_dyn}}.
}
\author{
Prim'Act
}
