% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypTech-class.R
\docType{class}
\name{HypTech}
\alias{HypTech}
\title{La classe \code{HypTech}.}
\description{
Une classe contenant les listes de tables de mortalite, de rachat, les parametres de rachat dynamique et
les parametres comportementaux qui permettent
de calculer les attentes en matiere de taux de revalorisation cible.
}
\details{
Chaque element de ces listes doit avoir ete prealablement nomme.
}
\section{Slots}{

\describe{
\item{\code{tables_mort}}{une liste contenant des tables de mortalite au format \code{\link{ParamTableMort}}.}

\item{\code{tables_rach}}{une liste contenant des tables de rachat (structurel) au format \code{\link{ParamTableRach}}.}

\item{\code{param_rach_dyn}}{une liste contenant des parametres de rachat dynamique
au format \code{\link{ParamRachDyn}}.}

\item{\code{param_comport}}{une liste contenant des parametres comportementaux au format \code{\link{ParamComport}}.}
}}

\seealso{
Les classes de parametres contenues : \code{\link{ParamTableMort}}, \code{\link{ParamTableRach}},
\code{\link{ParamRachDyn}}, \code{\link{ParamComport}}.
La methode pour l'application des chocs de mortalite et de longevite : \code{\link{get_choc_table}}.
La methode pour l'application des chocs de rachat haut et bas : \code{\link{get_choc_rach}}.
La methode pour la recuperation des parametres comportementaux : \code{\link{get_comport}}.
La methode pour la recuperation des taux de deces : \code{\link{get_qx_mort}}.
La methode pour la recuperation des taux de rachat structurel : \code{\link{get_qx_rach}}.
La methode pour la recuperation des taux de rachat dynamique : \code{\link{get_rach_dyn}}.
}
\author{
Prim'Act
}
\keyword{classes}
