% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypTech-class.R
\docType{class}
\name{HypTech}
\alias{HypTech}
\title{La classe HypTech}
\description{
Classe pour pour les tables de rachat
}
\section{Slots}{

\describe{
\item{\code{4}}{listes contenant les data frame relatifs aux tables de mortalite, tables de rachat et parametres de rachat conjoncturels}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

