% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immo_class.R
\docType{class}
\name{Immo}
\alias{Immo}
\title{Classe pour les actifs de type immobilier.}
\description{
Classe pour les actifs de type immobilier.
}
\section{Slots}{

\describe{
\item{\code{ptf_immo}}{est un dataframe, chaque ligne represente un actif immobilier du portefeuille d'immobilier.}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Les operations d'achat vente immo  \code{\link{buy_immo}} et \code{\link{sell_immo}}.
}

