% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppb_init_debut.R
\docType{methods}
\name{init_debut_ppb}
\alias{init_debut_ppb}
\title{Re-initialise la PPB en debut d'annee.}
\usage{
init_debut_ppb(x)
}
\arguments{
\item{x}{un objet de la classe \code{\link{Ppb}}.}
}
\value{
L'objet \code{x} reinitialise.
}
\description{
\code{init_debut_ppb} est une methode permettant de re-initialiser les montants de
 dotation ou de reprise cumules sur l'annee et
de re-initialiser le montant de PPB de debut de periode.
}
\author{
Prim'Act
}

