% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialisation_initSimBEL.R
\docType{methods}
\name{init_SimBEL}
\alias{Initialisation}
\alias{init_SimBEL}
\title{Initialisation d'un workspace.}
\usage{
init_SimBEL(x)
}
\arguments{
\item{x}{un objet de la classe \code{Initialisation}.}

\item{nb_simu}{nombre de simulation.}

\item{nb_annee_proj}{nombre d'annee de projection.}
}
\value{
Pas de sortie.
}
\description{
\code{init_SimBEL} est la methode d'initialisation d'un workspace.
}
\author{
Prim'Act
}

