% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypTech-load.R
\docType{methods}
\name{load_ht}
\alias{load_ht}
\title{Methode permettant de charger la valeur des parametres techniques.}
\usage{
load_ht(x)
}
\arguments{
\item{x}{est un objet de la classe \code{\link{Initialisation}} qui est utilise pour renseigner le chemin
d'acces de tous les parametres techniques.}
}
\value{
L'objet de la classe \code{\link{HypTech}} construit a partir des inputs renseignes par l'utilisateur.
}
\description{
\code{load_ht} est une methode permettant de charger les parametres associees a un
objet de classe \code{\link{HypTech}}.
}
\author{
Prim'Act
}
\seealso{
La classe \code{\link{Initialisation}} et sa methode \code{\link{set_architecture}}
 pour renseigner l'input.
}

