% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBase_merge_table_produit.R
\docType{methods}
\name{merge_table_produit}
\alias{merge_table_produit}
\title{Jointure et insertion de donnees dans la \code{DataBase}. Les donnees sont les ouputs \code{output_produit} de la fonction \code{\link{proj_an}}}
\usage{
merge_table_produit(
  flux,
  stock,
  hors_model,
  fin,
  nb_annee,
  sim,
  nom_produit,
  result_tech,
  result_fin,
  result_brut,
  result_net
)
}
\arguments{
\item{flux}{une \code{matrix} a inserer dans la base.}

\item{stock}{une \code{matrix} a inserer dans la base.}

\item{hors_model}{un \code{data.frame} a inserer dans la base.}

\item{fin}{un \code{data.frame} a inserer dans la base.}

\item{nb_annee}{un \code{integer} indiquant le nombre d'annees de projection.}

\item{sim}{un \code{integer} representant le numero de la simulation.}

\item{nom_produit}{une liste de \code{character} contenant les noms des differents produits.}

\item{result_tech}{une \code{matrix} contenant le resultat technique.}

\item{result_fin}{une \code{matrix} contenant le resultat financier.}

\item{result_brut}{une \code{matrix} contenant le resultat brut.}

\item{result_net}{une \code{matrix} contenant le resultat net.}
}
\description{
Jointure et insertion de donnees dans la \code{DataBase}. Les donnees sont les ouputs \code{output_produit} de la fonction \code{\link{proj_an}}
}
\author{
Prim'Act
}
