% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPointESG_class.r
\docType{class}
\name{ModelPointESG}
\alias{ModelPointESG}
\title{La classe ModelPointESG}
\description{
Classe pour une extraction de l'ESG.
}
\section{Slots}{

\describe{
\item{\code{annee}}{est une valeur \code{integer} correspondant a l'annee de projection.}

\item{\code{num_traj}}{est une valeur \code{integer} correspondant au numero de simulation de l'ESG.}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

