% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oblig_class.R
\docType{class}
\name{Oblig}
\alias{Oblig}
\title{Classe pour les actifs de type obligation.}
\description{
Classe pour les actifs de type obligation.
}
\section{Slots}{

\describe{
\item{\code{ptf_oblig}}{est un dataframe, chaque ligne represente un actif obligation du portefeuille d'obligation.}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe

Les operations d'achat vente obligations  \code{\link{buy_oblig}} et \code{\link{sell_oblig}}.
}

