% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamAlmEngine_class.R
\docType{class}
\name{ParamAlmEngine}
\alias{ParamAlmEngine}
\title{La classe \code{ParamAlmEngine}.}
\description{
Une classe pour le parametre ALM d'un canton.
}
\section{Slots}{

\describe{
\item{\code{ptf_reference}}{est un objet de type \code{\link{PortFin}}, qui represente le portefeuille
d'investissement de reference d'un canton.}

\item{\code{alloc_cible}}{un vecteur de 4 elements rendant compte du pourcentage
de l'actif composant respectivement les investissements: actions, immobiliers, obligataires et de tresorerie.}

\item{\code{seuil_realisation_PVL}}{une valeur \code{numeric} correspondant au pourcentage de plus-values actions
qui peut etre liquidee chaque annee pour atteindre l'objectif de revalorisation cible des passifs.}
}}

\author{
Prim'Act
}
\keyword{classes}
