% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamChocSousc-class.R
\docType{class}
\name{ParamChocSousc}
\alias{ParamChocSousc}
\title{La classe ParamChocSousc}
\description{
Classe pour pour les parametres de comportement
}
\section{Slots}{

\describe{
\item{\code{mp}}{: model point contenant l'ensemble des parametres pour faire les differents chocs}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

