% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamRachDyn-class.R
\docType{class}
\name{ParamRachDyn}
\alias{ParamRachDyn}
\title{La classe de parametres de rachat dynamique \code{ParamRachDyn}.}
\description{
Une classe pour les parametres de des lois de rachat dynamique.
}
\section{Slots}{

\describe{
\item{\code{vec_param}}{un \code{data frame} contenant les parametres pour les rachats dynamiques.}
}}

\seealso{
Le calcul du taux de rachat dynamique \code{\link{calc_rach_dyn}}.
}
\author{
Prim'Act
}
\keyword{classes}
