% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamTableMort-class.R
\docType{class}
\name{ParamTableMort}
\alias{ParamTableMort}
\title{La classe de parametres pour les tables de mortalite \code{ParamTableMort}.}
\description{
Une classe de parametres pour les tables de mortalite.
}
\section{Slots}{

\describe{
\item{\code{age_min}}{un entier correspondant a l'age minimal de la table.}

\item{\code{age_max}}{un entier correspondant a l'age maximal de la table.}

\item{\code{gen_min}}{un entier correspondant a la premiere generation de la table.}

\item{\code{gen_max}}{un entier correspondant a la derniere generation de la table.}

\item{\code{table}}{un \code{data frame} contenant la table de mortalite.}
}}
\note{
Les tables de mortalite doivent contenir des effectifs sous risque par age (Lx).
}
\author{
Prim'Act
}
\seealso{
Le calcul du taux de deces \code{\link{calc_qx}}.
}
\keyword{classes}

