% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamTableMort-class.R
\docType{class}
\name{ParamTableMort}
\alias{ParamTableMort}
\title{La classe ParamTableMort}
\description{
Classe pour pour les tables de mortalite
}
\section{Slots}{

\describe{
\item{\code{table}}{un data frame contenant les taux de mortalite}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

