% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamTableRach-class.R
\docType{class}
\name{ParamTableRach}
\alias{ParamTableRach}
\title{La classe ParamTableRach}
\description{
Classe pour pour les tables de rachat
}
\section{Slots}{

\describe{
\item{\code{table}}{un data frame contenant les taux de rachat}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

