% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevaloEngine_pb_contr.R
\docType{methods}
\name{pb_contr}
\alias{RevaloEngine}
\alias{pb_contr}
\title{Calcule la PB contractuelle.}
\usage{
pb_contr(base_fin, tx_pb, rev_stock_brut, ch_enc_th, tx_enc_moy)
}
\arguments{
\item{base_fin}{un vecteur \code{numeric} comprenant par produit la base de produits financiers.}

\item{tx_pb}{un vecteur \code{numeric} comprenant par produit les taux de participation aux benefices contractuels.}

\item{rev_stock_brut}{un vecteur de type \code{numeric} comprenant par produit la revalorisation
appliquee sur le stock au taux minimum.}

\item{ch_enc_th}{est un vecteur de type \code{numeric} comprenant par produit le montant total
des chargements sur encours appliques au stock et revalorises au taux minimum. Il s'agit ici des chargements
qui pourraient theoriquement etre preleves.}

\item{tx_enc_moy}{un vecteur \code{numeric} comprenant par produit
les taux de chargements sur encours moyens.}
}
\value{
\code{ch_enc_ap_pb_contr} un vecteur comprenant par produit les chargements sur encours appliques

\code{rev_stock_nette_contr} un vecteur comprenant par produit la
revalorisation contractuelle nette.
}
\description{
\code{pb_contr} est une methode permettant de calculer la participation aux benefices contractuelle par produit.
}
\details{
Le montant des chargements \code{ch_enc_th} est theorique et peut
conduire a l'application d'une revalorisation nette negative.
}
\author{
Prim'Act
}

