% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortFin_class.R
\docType{class}
\name{PortFin}
\alias{PortFin}
\title{La classe \code{PortFin}.}
\description{
Une Classe pour le portefeuille financier.
}
\section{Slots}{

\describe{
\item{\code{annee}}{est un objet de type \code{integer}correspondant a l'annee courante.}

\item{\code{ptf_action}}{est un objet de type \code{\link{Action}}, qui represente le portefeuille d'action d'un canton.}

\item{\code{ptf_immo}}{est un objet de type \code{\link{Immo}}, qui represente le portefeuille immobilier d'un canton.}

\item{\code{ptf_oblig}}{est un objet de type \code{\link{Oblig}}, qui represente le portefeuille obligataire d'un canton.}

\item{\code{ptf_treso}}{est un objet de type \code{\link{Treso}}, qui represente le portefeuille monetaire d'un canton.}

\item{\code{pre}}{est un objet de type \code{\link{PRE}}, qui represente la PRE d'un canton.}

\item{\code{rc}}{est un objet de type \code{\link{RC}}, qui represente la RC d'un canton.}

\item{\code{frais_fin}}{est un objet de type \code{\link{FraisFin}}, qui represente les frais financiers d'un canton.}

\item{\code{pvl_action}}{est un \code{numeric}, qui correspond a la somme des plus values latentes des actifs Actions qui sont en situation de plus values latentes.}

\item{\code{pvl_immo}}{est un \code{numeric}, qui correspond a la somme des plus values latentes des actifs Immo qui sont en situation de plus values latentes.}

\item{\code{pvl_oblig}}{est un \code{numeric}, qui correspond a la somme des plus values latentes des actifs Obligs qui sont en situation de plus values latentes.}

\item{\code{mvl_action}}{est un \code{numeric}, qui correspond a la somme des moins values latentes des actifs Actions qui sont en situation de moins values latentes.}

\item{\code{mvl_immo}}{est un \code{numeric}, qui correspond a la somme des moins values latentes des actifs Immos qui sont en situation de moins values latentes.}

\item{\code{mvl_oblig}}{est un \code{numeric}, qui correspond a la somme des moins values latentes des actifs Obligs qui sont en situation de moins values latentes.}

\item{\code{vm_vnc_precedent}}{est une liste composee de deux elements : la \code{vm_precedente} et la
 \code{vnc_precedente}, correspondant respectivement a la valeur de marche
et a la valeur nette comptable en debut d'annee de l'objet PortFin.}
}}

\seealso{
Le calcul des rendements : \code{\link{calc_rdt}}.
Le calcul des plus ou moins-values latentes : \code{\link{calc_pmvl}}.
Le chargement des donneees initiales : \code{\link{chargement_PortFin}}, \code{\link{chargement_PortFin_reference}}.
Les methodes de mise a jour : \code{\link{update_PortFin}}, \code{\link{update_PortFin_reference}}, \code{\link{do_update_pmvl}},
\code{\link{do_update_vm_vnc_precedent}}.
L'allocation d'actifs : \code{\link{print_alloc}}.
}
\author{
Prim'Act
}
\keyword{classes}
