% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortPassif-class.R
\docType{class}
\name{PortPassif}
\alias{PortPassif}
\title{La classe PortPassif}
\description{
La classe PortPassif
}
\section{Slots}{

\describe{
\item{\code{eei}}{une liste d'objets de classe \code{EpEuroInd} contenant l'ensemble portefeuilles
relatifs aux model points epargne en euros.}

\item{\code{names_class_prod}}{un vecteur \code{character} indique les noms de classes de produits.}

\item{\code{ht}}{un objet de classe \code{HypTech}.}

\item{\code{fp}}{un objet de classe \code{FraisPassif}.}

\item{\code{tx_pb}}{un objet de classe \code{Taux_PB}.}

\item{\code{autres_passifs}}{un objet de classe \code{AutresPassifs}.}

\item{\code{autres_reserves}}{un objet de classe \code{AutresReserves}.}

\item{\code{ere}}{une liste d'objets de classe EpRetrInd contenant l'ensemble portefeuilles relatifs aux model points epargne retraite en euros}

\item{\code{erp}}{une liste d'objets de classe EpRetrPoint contenant l'ensemble portefeuilles relatifs aux model points epargne retraite en points}

\item{\code{ren}}{une liste d'objets de classe Rente contenant l'ensemble portefeuilles relatifs aux model points rente en euros}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

