% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortPassif-class.R
\docType{class}
\name{PortPassif}
\alias{PortPassif}
\title{La classe \code{PortPassif}.}
\description{
Une classe regroupant l'ensemble des donnees de passifs et les hypotheses correspondantes.
}
\section{Slots}{

\describe{
\item{\code{annee}}{une valeur entiere correspondant a l'annee de projection.}

\item{\code{eei}}{une liste d'objets de la classe \code{\link{EpEuroInd}} contenant l'ensemble
des produits de type epargne en euros.}

\item{\code{rer}}{une liste d'objets de la classe \code{\link{RetraiteEuroRest}} contenant l'ensemble
des produits de retraite en phase de restitution.}

\item{\code{names_class_prod}}{un vecteur \code{character} indiquant les noms de classes de produits.}

\item{\code{ht}}{un objet de classe \code{\link{HypTech}} contenant les hypotheses techniques.}

\item{\code{fp}}{un objet de classe \code{\link{FraisPassif}} contenant les hypotheses de frais de passif
par produit.}

\item{\code{tx_pb}}{un objet de classe \code{\link{TauxPB}} contenant les taux contractuel de participation
aux benefices par produit.}

\item{\code{autres_passifs}}{un objet de classe \code{\link{AutresPassifs}}.}

\item{\code{autres_reserves}}{un objet de classe \code{\link{AutresReserves}}.}

\item{\code{calc_proba}}{un \code{logical} indiquant si les probabilites ont ete calculees.}

\item{\code{choc_lapse_mass}}{un \code{numeric} comprenant la valeur du choc de rachat massif. Cet attribut doit etre a 0 lorsque aucun
choc n'est souhaite. Sa valeur est comprise entre 0 et 1. Il ne s'applique que la premiere annee de projection.}

\item{\code{choc_mort_cat}}{un \code{numeric} comprenant la valeur du choc de mortalite catastrophique Cet attribut doit etre a 0 lorsque aucun
choc n'est souhaite. Sa valeur est comprise entre 0 et 1. Il ne s'applique que la premiere annee de projection.}
}}

\seealso{
La projection des produits sur l'annee avant attributiuon de participation
aux benefices : \code{\link{proj_annee_av_pb}}.
Le vieillissement des model points de passifs avant et apres attributiuon de participation
aux benefices : \code{\link{viellissement_av_pb}}, \code{\link{vieillissment_ap_pb}}.
}
\author{
Prim'Act
}
\keyword{classes}
