% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppb_class.R
\docType{class}
\name{Ppb}
\alias{Ppb}
\title{La classe Ppb}
\description{
Classe pour la provision pour participation aux benefices (PPB)
}
\section{Slots}{

\describe{
\item{\code{valeur_ppb}}{est la valeur courante prise par la PPB.}

\item{\code{ppb_debut}}{est la valeur prise par la PPB en debut d'annee.}

\item{\code{seuil_rep}}{est une valeur \code{numeric} correspond a la proportion de PPB de debut d'annee
que l'on peut reprendre sur une periode.
sur la PPB sur une periode.}

\item{\code{seuil_dot}}{est une valeur \code{numeric} correspond a la montant maximal de dotation possible sur la PPB
sur une periode, exprimee comme une fraction de la PPB de debut d'annee.}

\item{\code{compte_rep}}{est une valeur \code{numeric} qui totalise les montants de reprise effectuee sur une periode.}

\item{\code{compte_dot}}{est une valeur \code{numeric} qui totalise les montants de dotation effectuee sur une periode.}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

