% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_class.R
\docType{class}
\name{PRE}
\alias{PRE}
\title{La classe \code{PRE}.}
\description{
Une classe pour la gestion de la provision pour risque d'exigibilite (PRE).
}
\section{Slots}{

\describe{
\item{\code{val_debut}}{est une valeur \code{numeric} correspondant a la valeur de la PRE en debut d'annee.}

\item{\code{val_courante}}{est une valeur \code{numeric} correspondant a la valeur courante de la PRE.}
}}

\seealso{
Les methodes de calcul de la PRE \code{\link{calc_PRE}}, et de mises a jour des PRE initiales et courantes \code{\link{do_update_PRE_val_courante}}, \code{\link{do_update_PRE_val_debut}}.
}
\author{
Prim'Act
}
\keyword{classes}
