% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortFin_print_alloc.R
\docType{methods}
\name{print_alloc}
\alias{print_alloc}
\title{Calcul le poids de chaque composante du portefeuille.}
\usage{
print_alloc(x)
}
\arguments{
\item{x}{objet de la classe \code{\link{PortFin}}.}
}
\value{
Un \code{data.frame} compose de quatre colonnes et cinq lignes.
La colonne \describe{
\item{\code{alloc_valeur} : }{decrit le montant alloue en valeur de marche par poche d'actif.}
\item{\code{alloc_proportion} : }{decrit la proportion allouee en valeur de marche par poche d'actif.}
\item{\code{alloc_valeur_vnc} : }{decrit le montant alloue en valeur nette comptable par poche d'actif.}
\item{\code{alloc_proportion_vnc} : }{decrit la proportion allouee en valeur nette comptable par poche d'actif.}
}
Les lignes correspondent aux classes d'actifs : (Action / Immobilier / Obligation / Tresorerie / Actifs cumules)
}
\description{
\code{pint_alloc} est une methode permettant de calculer l'allocation absolue et relative du portefeuille.
}
\author{
Prim'Act
}

