% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Canton_proj_an.R
\docType{methods}
\name{proj_an}
\alias{Canton}
\alias{proj_an}
\title{Projette un canton sur une annee.}
\usage{
proj_an(x, annee_fin)
}
\arguments{
\item{x}{est un objet de type \code{Canton}.}

\item{annee_fin}{est une valeur \code{numeric} correpondant a l'annee de fin de projection.}
}
\value{
Une liste comprenant l'objet x mis a jour, les flux de best estimate, les resultats comptables.
}
\description{
\code{proj_an} est une methode permettant de projeter un canton sur une annee pour le calcul
des flux de best estimate.
}
\author{
Prim'Act
}

