% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action_revalo.R
\docType{methods}
\name{revalo_action}
\alias{revalo_action}
\title{Revalorise les valeurs de marche du portefeuille action.}
\usage{
revalo_action(x, S, S_prev)
}
\arguments{
\item{x}{un objet de la classe \code{\link{Action}} (decrivant le portefeuille d'action).}

\item{S}{un vecteur \code{numeric} correspondant a la valeur des indices actions}

\item{S_prev}{un vecteur \code{numeric}  correspondant a la valeur des indices actions a la periode precedente.}
}
\value{
Un data frame compose de deux colonnes et autant de lignes que le portefeuille action a de lignes.
La premiere colonne decrit de le rendement annuel de chacune des actions composants le portefeuille action.
La seconde colonne decrit les dividendes annuelles percues au titre de chacune des actions composants le portefeuille action.
}
\description{
\code{revalo_action} est une methode permettant de revaloriser et de calculer les dividendes
du portefeuille action sur une periode.
}
\author{
Prim'Act
}

