% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immo_flux.R
\docType{methods}
\name{revalo_immo}
\alias{Immo}
\alias{revalo_immo}
\title{Calcul les valeurs de marches de chaque composante du portefeuille Immobilier.}
\usage{
revalo_immo(x, S, S_prev)
}
\arguments{
\item{x}{objet de la classe Immo (decrivant le portefeuille d'immobilier).}

\item{S}{vecteur de valeur des biens immobiliers du ptf en milieu d'annee N (date de versement des dividendes)}

\item{S_prev}{vecteur de valeur des biens immobiliers du ptf en milieu d'annee N-1.}
}
\value{
Un data frame composé de deux colonnes et autant de lignes que le portefeuille immo a de lignes.
La premiere colonne decrit de le rendement annuel de chacune des actions composants le portefeuille immo.
La seconde colonne decrit les dividendes annuelles percues au titre
de chacun des biens immobilier composants le portefeuille action.
}
\description{
\code{revalo_immo} est une methode permettant de calculer les valeurs de marche.
}
\author{
Prim'Act
}

