% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Treso_revalo.R
\docType{methods}
\name{revalo_treso}
\alias{revalo_treso}
\title{Calcul les valeurs de marches de chaque composante du portefeuille treso.}
\usage{
revalo_treso(Rt, Rt_prev)
}
\arguments{
\item{Rt}{valeur du taux zero coupon 1 an de l'annee.}

\item{Rt_prev}{valeur du taux zero coupon 0 an de l'annee.}
}
\value{
Le taux de rendement.
}
\description{
\code{revalo_treso} est une methode permettant de calculer les valeurs de marche.
}
\author{
Prim'Act
}

