% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevaloEngine_class.R
\docType{class}
\name{RevaloEngine}
\alias{RevaloEngine}
\title{La classe \code{RevaloEngine}.}
\description{
Une classe comprenant les methodes pour l'application de la revalorisation des passifs.
}
\section{Slots}{

\describe{
\item{\code{param_revalo}}{est objet de type \code{\link{ParamRevaloEngine}} comprenant
les parametres utilises pour la revalorisation des contrats.}
}}

\author{
Prim'Act
}
\keyword{classes}
