% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Be-run_be.R
\docType{methods}
\name{run_be}
\alias{Be}
\alias{run_be}
\title{Calcul d'un BE}
\usage{
run_be(x)
}
\arguments{
\item{x}{est un objet de type \code{Be}.}
}
\value{
L'objet BE avec les tables de resultats mise a jour.
}
\description{
\code{run_be} est une methode permettant de calcul un best estimate pour un canton.
}
\author{
Prim'Act
}

