% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Be-run_be.R
\docType{methods}
\name{run_be}
\alias{run_be}
\title{Calcul d'un BE.}
\usage{
run_be(x, pre_on)
}
\arguments{
\item{x}{un objet de type \code{\link{Be}}.}

\item{pre_on}{une valeur \code{logical} qui lorsqu'elle vaut \code{TRUE} prend en compte la variation
de PRE dans le resultat technique utilisee pour le calcul de la participation aux benefices reglementaires.}
}
\value{
\code{be} l'objet \code{x} mis a jour : l'attribut \code{tab_be} contient le best estimate et sa
decomposition, l'attribut \code{tab_flux} contient les flux moyens du best estimate et ses
composantes.

\code{err_simu} un vecteur contenant la liste des simulations qui ont generes des erreurs et qui n'ont pu
etre utilisees pour le calcul du best estimate.
}
\description{
\code{run_be} est une methode permettant de calculer un best estimate pour un canton.
}
\details{
Il s'agit de la methode principale du package \code{SimBEL}. Cette methode requiert le chargement
d'un objet \code{\link{Be}} deja parametre et alimente en donnees. La methode \code{\link{init_scenario}}
permet d'alimenter un objet \code{\link{Be}} dans la situation "centrale" de la formule standard et en situation
de choc.
}
\seealso{
Le calcul du best estimate pour une simulation : \code{\link{run_be_simu}}.
L'initialisation d'un best estimate : \code{\link{init_scenario}}.
La classe \code{\link{Be}}.
La sortie des resultats au format ".csv" : \code{\link{write_be_results}}.
}
\author{
Prim'Act
}
