% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialisation_set_architecture.R
\docType{methods}
\name{set_architecture}
\alias{set_architecture}
\title{Definition de l'architecture d'un workspace.}
\usage{
set_architecture(x)
}
\arguments{
\item{x}{un objet de la classe \code{\link{Initialisation}}.}
}
\value{
Objet mis a jour de l'ensemble des chemins du workspace,
ceux ci sont stockes sous forme de liste dans l'attribut \code{address}
de l'objet \code{\link{Initialisation}} renseigne en input.
}
\description{
\code{set_architecture}.
}
\author{
Prim'Act
}
