% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabEpEuroInd-class.R
\docType{class}
\name{TabEpEuroInd}
\alias{TabEpEuroInd}
\title{La classe TabEpEuroInd}
\description{
Classe pour le stockage en memoire de variable de calcul au niveau du model points
}
\section{Slots}{

\describe{
\item{\code{tab}}{un objet \code{list} au format fige contenant l'ensemble des variables stockees}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

