% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TauxPB-class.R
\docType{class}
\name{TauxPB}
\alias{TauxPB}
\title{La classe TauxPB}
\description{
Classe pour pour Taux PB
}
\section{Slots}{

\describe{
\item{\code{mp}}{data frame contenant les paramètres des Taux PB}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

