% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Treso_class.R
\docType{class}
\name{Treso}
\alias{Treso}
\title{La classe \code{Treso}.}
\description{
Une classe pour les actifs de type Tresorerie.
}
\section{Slots}{

\describe{
\item{\code{ptf_treso}}{est un dataframe, chaque ligne represente un actif de tresorerie du portefeuille de monetaire.}
}}
\author{
Prim'Act
}
\seealso{
Les methodes de calcul des valeurs \code{\link{calc_vm_treso}},
de calcul des revenus de la tresorerie \code{\link{revenu_treso}},
de calcul de la revalorisation de la tresorerie \code{\link{revalo_treso}},
de mise a jour de la tresorerie \code{\link{update_treso}}.
}
\keyword{classes}

