% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Treso_class.R
\docType{class}
\name{Treso}
\alias{Treso}
\title{La classe Treso}
\description{
Classe pour les actifs de type Tresorerie
}
\section{Slots}{

\describe{
\item{\code{ptf_treso}}{est un dataframe, chaque ligne represente un actif de tresorerie du portefeuille de monetaire.}
}}
\section{Lien a creer}{

}
\author{
Prim'Act
}
\seealso{
Mettre le lien vers les methodes de la classe
}
\keyword{classes}

