% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immo_update_dur_det.R
\docType{methods}
\name{update_dur_det_immo}
\alias{update_dur_det_immo}
\title{Mise a jour des durees de detention d'un portefeuille immobilier.}
\usage{
update_dur_det_immo(x)
}
\arguments{
\item{x}{objet de la classe \code{\link{Immo}} (decrivant le portefeuille immo en detention).}
}
\value{
L'objet \code{x} mis a jour du vieillissement de la duree de detention.
}
\description{
\code{update_dur_det_immo} est une methode permettant de mettre a jour la duree de detention des composantes d'un portefeuille immobilier.
}
\author{
Prim'Act
}

