% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortFin_update_reference.R
\docType{methods}
\name{update_PortFin_reference}
\alias{update_PortFin_reference}
\title{Evalue et met a jour les objets constituants un PortFin_reference.}
\usage{
update_PortFin_reference(an, x, mp_ESG)
}
\arguments{
\item{an}{\code{numeric} correspond a l'annee de projection du portefeuille financier de reinvestissement.}

\item{x}{objet de la classe \code{\link{PortFin}}, correspondant au portefeuille financier de
reinvestissement avant l'etape de vieillissement.}

\item{mp_ESG}{est un objet de la classe \code{\link{ModelPointESG}}, decrivant les conditions
economiques permettant d'effectuer le vieillissement du portefeuille financier de reinvestissement.}
}
\value{
L'objet \code{x} renvoye correspond au portefeuille financier de reinvesitssement veilli d'une annee.
}
\description{
\code{update_PortFin_reference} est une methode permettant de calculer et mettre a jour un portefeuille financier
de reinvestissement pour le faire vieillir d'une annee.
}
\author{
Prim'Act
}
\seealso{
La fonction de mise a jour specifique au portefeuille \code{\link{update_PortFin}}.
}

