% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutresReserves-update_reserves.R
\docType{methods}
\name{update_reserves}
\alias{update_reserves}
\title{Evalue et met a jour la valeur des autres reserves.}
\usage{
update_reserves(x, prest_ep, prest_autres, pm_ep, pm_autres)
}
\arguments{
\item{x}{objet de la classe \code{\link{AutresReserves}}.}

\item{prest_ep}{est une valeur \code{numeric} correspondant a la somme des prestations nettes de chargement et
de charges sociales sur epargne.}

\item{prest_autres}{est une valeur \code{numeric} correspondant a la somme des prestations nettes
de chargements et de charges sociales sur autres passifs.}

\item{pm_ep}{est une valeur \code{numeric}  correspondant a la somme des PM nettes de chargements et
de charges sociales sur epargne.}

\item{pm_autres}{est une valeur \code{numeric} correspondant a la somme des PM nettes de chargement et
de charges sociales sur autres passifs.}
}
\value{
\code{x} l'objet\code{\link{AutresReserves}} mis a jour.

\code{var_psap} une valeur \code{numeric} correspondant a la variation de PSAP.

\code{var_gg} une valeur \code{numeric} correspondant a la variation de PGG.
}
\description{
\code{update_reserves} est une methode permettant de calculer la valeur de la nouvelle PGG et de la nouvelle
PSAP et les met a jour.
}
\note{
Il s'agit d'une methode simplifiee.
}
\author{
Prim'Act
}

