% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action_update_vm.R
\docType{methods}
\name{update_vm_action}
\alias{update_vm_action}
\title{Mise a jour de valeurs de marche d'un portefeuille action.}
\usage{
update_vm_action(x, vm)
}
\arguments{
\item{x}{objet de la classe \code{\link{Action}} (decrivant le portefeuille action en detention).}

\item{vm}{un vecteur de \code{numeric} ayant la meme longueur que le portefeuille action a de lignes
et correspondant aux nouvelles valeurs de marche du portefeuille action.}
}
\value{
L'objet \code{x} mis a jour.
}
\description{
\code{update_vm_action} est une methode permettant de mettre a jour la valeur de marche des composantes d'un portefeuille Action.
}
\author{
Prim'Act
}

