% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action_update_vm.R
\docType{methods}
\name{update_vm_action}
\alias{Action}
\alias{update_vm_action}
\title{Mise a jour de chaque composante du portefeuille action suite a la vente de tout ou partie de ce portefeuille.}
\usage{
update_vm_action(x, vm)
}
\arguments{
\item{x}{objet de la classe \code{Action} (decrivant le portefeuille action en detention).}
}
\value{
L'objet \code{x} mis a jour du vieillissement de la duree de detention.
}
\description{
\code{update_vm_action} est une methode permettant de mettre a jour la duree de detention des composantes d'un portefeuille Action.
}
\author{
Prim'Act
}

