% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassifBase-vieilli_mp.R
\docType{methods}
\name{vieilli_mp}
\alias{vieilli_mp}
\title{Veillissement d'un an des contrats epargne en euros.}
\usage{
vieilli_mp(x, pm_fin, pm_gar, tx_revalo, an)
}
\arguments{
\item{x}{un objet de la classe \code{\link{EpEuroInd}} ou de la classe \code{\link{RetraiteEuroRest}} contenant les model points epargne euros.}

\item{pm_fin}{un vecteur de type \code{numeric} contenant par model point les montants de PM revalorises apres participation aux benefices.}

\item{pm_gar}{un vecteur de type \code{numeric} contenant par model point les montants de PM garanties revalorises apres participation aux
benefices au titre de la PPB initiale.}

\item{tx_revalo}{un vecteur de type \code{numeric} contenant par model point les taux de revalorisation nets appliques.}

\item{an}{une valeur \code{integer} correspondant a l'annee du calcul des prestations.}
}
\value{
l'objet \code{x} vieilli d'une periode.
}
\description{
\code{vieilli_mp} est une methode permettant de vieillir
 les model points epargne en euros d'une peridoe.
}
\seealso{
Calcul de la revalorisation des PM \code{\link{calc_revalo_pm}}.
}
\author{
Prim'Act
}
