% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortFin_vieillissement_oblig.R
\docType{methods}
\name{vieillissement_oblig_PortFin}
\alias{vieillissement_oblig_PortFin}
\title{Effectue le vieillissement du portefeuille obligataire d'un portefeuille financier.}
\usage{
vieillissement_oblig_PortFin(x, new_mp_ESG)
}
\arguments{
\item{x}{objet de la classe \code{\link{PortFin}}, correspondant au portefeuille financier
de l'assureur avant l'etape de vieillissement.}

\item{new_mp_ESG}{est un objet de type \code{\link{ModelPointESG}}, correspondant aux conditions
economiques de l'annee du vieillissement.}
}
\value{
\code{portFin} le portefeuille financier dont l'attribut \code{ptf_oblig} a ete vieilli d'une annee.

\code{coupon} le montant des coupons percus en milieu d'annee suite au vieillissement du portefeuille obligataire.

\code{echeance} le montant des echeance percus en fin d'annee suite au vieillissement du portefeuille obligataire.

\code{var_vnc_oblig} la variation de VNC constatee sur les titres non arrives en echeance.
}
\description{
\code{vieillissement_oblig_PortFin} est une methode permettant de projeter la composante obligataire d'un portefeuille financier.
}
\seealso{
La fonction de calcul des rendements des actifs \code{\link{calc_rdt}}.
}
\author{
Prim'Act
}
