% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortPassif-vieillissement_ap_pb.R
\docType{methods}
\name{vieillissment_ap_pb}
\alias{vieillissment_ap_pb}
\title{Vieillissement du portefeuille sur l'annee apres attribution
de participation aux benefices.}
\usage{
vieillissment_ap_pb(x, rev_nette_alloue, tx_soc)
}
\arguments{
\item{x}{un objet de la classe \code{\link{PortPassif}} contenant l'ensemble des produits de passifs.}

\item{rev_nette_alloue}{un vecteur \code{numeric} contenant par produit
le supplement de revalorisation par rapport au taux minimum.}

\item{tx_soc}{une valeur \code{numeric} correspondant au taux de charges sociales.}
}
\value{
\code{x} l'objet \code{x} mis a jour.

\code{nom_produit} un vecteur de \code{character} contenant les noms des produits.

\code{flux_agg} une matrice contenant les flux aggreges par produits.

\code{stock_agg} une matrice contenant les stocks aggreges par produits.
}
\description{
\code{vieillissment_ap_pb} est une methode permettant de calculer les PM et les flux sur une annee apres PB.
Cette methode vieilli le portefeuille de passifs apres attribution de PB.
}
\seealso{
L'attribution de la revalorisation par model point : \code{\link{calc_revalo_pm}}
Le viellissement des model points : \code{\link{vieilli_mp}}.
}
\author{
Prim'Act
}
